% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{add_baseline}
\alias{add_baseline}
\title{Generate a baseline matrix}
\usage{
add_baseline(max_a, nrow, ncol)
}
\arguments{
\item{max_a}{Numeric. Maximum allowed baseline level.}

\item{nrow}{Integer. Number of row of the output matrix.}

\item{ncol}{Integer. Number of columns of the output matrix.}
}
\value{
A matrix with a constant value (the baseline)
}
\description{
Generate a baseline matrix
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
