% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{add_tail}
\alias{add_tail}
\title{Simulate peak tailing}
\usage{
add_tail(max_a, max_b, max_c, nrow)
}
\arguments{
\item{max_a}{Numeric. Maximum allowed offset.}

\item{max_b}{Numeric. Maximum allowed multiplicative constant.}

\item{max_c}{Numeric. Maximum allowed exponent.}

\item{nrow}{Integer. The output vector length.}
}
\value{
A vector that simulates peak tailing as an offset plus an exponential decay
}
\description{
Simulate peak tailing
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
