% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSDataset.R
\name{align,GCIMSDataset-method}
\alias{align,GCIMSDataset-method}
\title{Align a GCIMS dataset}
\usage{
\S4method{align}{GCIMSDataset}(object, method = "ptw", align_ip = TRUE)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
The alignment uses a multiplicative correction in drift time and a
Parametric Time Warping correction in retention time
}
