% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-align.R
\name{align,GCIMSSample-method}
\alias{align,GCIMSSample-method}
\title{Align a GCIMSSample object, in drift and retention time}
\usage{
\S4method{align}{GCIMSSample}(object, rip_ref_ms, ric_ref, ric_ref_rt)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rip_ref_ms}{The reference position of the Reactant Ion Peak in the dataset (in ms)}

\item{ric_ref}{The reference Reverse Ion Chromatogram}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample object, in drift and retention time
}
