% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{align}
\alias{align}
\title{Align an object}
\usage{
align(object, ...)
}
\arguments{
\item{object}{An object to align}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The object, modified
}
\description{
Align an object
}
