% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ip_GCIMSDataset.R
\name{align_ip,GCIMSDataset-method}
\alias{align_ip,GCIMSDataset-method}
\title{Align a GCIMS dataset to the injection point}
\usage{
\S4method{align_ip}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
Aligns all the samples to their injection point in retention time
}
