% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ip_GCIMSSample.R
\name{align_ip,GCIMSSample-method}
\alias{align_ip,GCIMSSample-method}
\title{Align a GCIMS Sample to the injection point}
\usage{
\S4method{align_ip}{GCIMSSample}(object, min_start, rt_ref)
}
\arguments{
\item{object}{A \link{GCIMSSample} object, modified in-place}

\item{min_start}{The points that will be used before the injection point}

\item{rt_ref}{Retention time vector to be used}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Aligns a sample to the reference injection point
}
