% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_pow-GCIMSDataset.R
\name{align_pow,GCIMSDataset-method}
\alias{align_pow,GCIMSDataset-method}
\title{Align a \link{GCIMSDataset} using the POW algorithm}
\usage{
\S4method{align_pow}{GCIMSDataset}(object, lambdas = pracma::logspace(-2, 4, 31), p = 10, max_it = 1000)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{lambdas}{vector of lambdas to be tested for the POW algorithm}

\item{p}{points to be taken for validation for the POW algorithm}

\item{max_it}{maximum iterations for applying the POW algorithm}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
Uses the POW algorithm to generate waprs and align the samples
}
