% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_pow-GCIMSSample.R
\name{align_pow,GCIMSSample-method}
\alias{align_pow,GCIMSSample-method}
\title{Warp a GCIMSSample}
\usage{
\S4method{align_pow}{GCIMSSample}(object, w)
}
\arguments{
\item{object}{A \link{GCIMSSample} object, modified in-place}

\item{w}{warp to be applied to the GCIMSSample}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Warps a GCIMSSample with the provided warp
}
