% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{align_rip}
\alias{align_rip}
\title{Multiplicative correction to the drift time axis}
\usage{
align_rip(spectra, drift_time, drift_time_rip_ref = stats::median)
}
\arguments{
\item{spectra}{Matrix with one spectra per row to be aligned}

\item{drift_time}{Numerical vector as with length equal to ncol(spectra)}

\item{drift_time_rip_ref}{A function that will take the drift times of the RIP for each spectra
and returns the expected RIP drift time (default: median).
It can also be the theoretical drift time position.}
}
\description{
This spectra alignment function computes the largest peak from
each spectrum, assuming it corresponds to the Reactant Ion Peak.
Then it computes the median RIP position of all the samples and
corrects the drift time for each sample.
The correction is multiplicative, meaning:
drift_time_corr = Kcorr * drift_time
}
\details{
Changes in pressure and temperature during measurements may lead to
correction factors up to \link{0.9, 1.1}.

This alignment technique can be used when the pressure and temperature
measurements are not available.
}
