% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW.R
\name{align_td}
\alias{align_td}
\title{Multiplicative correction to the drift time axis}
\usage{
align_td(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where aligned data files are stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
sample to be visualized from the dataset.}
}
\description{
This spectra alignment function computes the largest peak from
each spectrum, assuming it corresponds to the Reactant Ion Peak.
Then it computes the RIP position of all the samples and
corrects the all the spectra, for each sample.
The correction is multiplicative, meaning:
drift_time_corr = Kcorr * drift_time
}
\details{
Changes in pressure and temperature during measurements may lead to
correction factors between 0.9 and 1.1.

This alignment technique can be used when the pressure and temperature
measurements are not available.
}
\examples{
input_samples <- system.file("extdata", package = "GCIMS")
to_align <- tempfile("dir")
dir.create(to_align, recursive = TRUE, showWarnings = FALSE)
file.copy(
  file.path(input_samples, "M3.rds"),
  file.path(to_align, "M0.rds")
)
file.copy(
  file.path(input_samples, "M7.rds"),
  file.path(to_align, "M7.rds")
)
aligned <- tempfile("dir")
align_td(
  dir_in = to_align,
  dir_out = aligned,
  samples = 7
)
}
