% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW.R
\name{align_tr}
\alias{align_tr}
\title{Correction of retention time axis using Parametric Time Warping (PTW)}
\usage{
align_tr(dir_in, dir_out, samples, correction_type)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where aligned data files are stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
sample to be visualized from the dataset.}

\item{correction_type}{Numeric. Integer between 0 and 5. If zero, no correction
in retention time is needed. Between 1 and 5 it indicates
the polynomial degree of the warping function.}
}
\description{
This chromatrogram computes the RIC of the reference samples and
uses it to correct each of the EIC of a sample, for the all the selected
samples on the dataset. The degree ot the polynomial that relates the
retention time axes of the reference and the sample to be corrected
is optimized using the average correlation between samples and reference as a figure
of merit.
}
\examples{
input_samples <- system.file("extdata", package = "GCIMS")
to_align <- tempfile("dir")
dir.create(to_align, recursive = TRUE, showWarnings = FALSE)
file.copy(
  file.path(input_samples, "M3.rds"),
  file.path(to_align, "M0.rds")
)
file.copy(
  file.path(input_samples, "M7.rds"),
  file.path(to_align, "M7.rds")
)
aligned <- tempfile("dir")
align_tr(
  dir_in = to_align,
  dir_out = aligned,
  samples = 7,
  correction_type = 1L
)
}
