% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-align.R
\name{alignDt,GCIMSSample-method}
\alias{alignDt,GCIMSSample-method}
\title{Align a GCIMSSample in drift time with a multiplicative correction}
\usage{
\S4method{alignDt}{GCIMSSample}(x, y, rip_ref_ms)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{y}{Not used}

\item{rip_ref_ms}{The position of the RIP in ms}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in drift time with a multiplicative correction
}
