% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{alignDt}
\alias{alignDt}
\title{Align a GCIMSSample in drift time with a multiplicative correction}
\usage{
alignDt(object, rip_ref_ms)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rip_ref_ms}{The position of the RIP in ms}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in drift time with a multiplicative correction
}
