% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-align.R
\name{alignPlots}
\alias{alignPlots}
\title{Plots to interpret alignment results}
\usage{
alignPlots(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}
}
\value{
A list with plots created with ggplot2.
}
\description{
Plots to interpret alignment results
}
