% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-align.R
\name{alignRt,GCIMSSample-method}
\alias{alignRt,GCIMSSample-method}
\title{Align a GCIMSSample in retention time using parametric time warping}
\usage{
\S4method{alignRt}{GCIMSSample}(x, y, ric_ref, ric_ref_rt)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{y}{Not used}

\item{ric_ref}{The reference Reverse Ion Chromatogram}

\item{ric_ref_rt}{The retention times corresponding to \code{ric_ref}}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in retention time using parametric time warping
}
