% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{alignRt_ip,GCIMSSample-method}
\alias{alignRt_ip,GCIMSSample-method}
\title{Align a GCIMSSample in retention time with a multiplicative correction}
\usage{
\S4method{alignRt_ip}{GCIMSSample}(x, ip_ref_s)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{ip_ref_s}{The position of the injection point in s}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in retention time with a multiplicative correction
}
