% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{alignRt_pow,GCIMSSample-method}
\alias{alignRt_pow,GCIMSSample-method}
\title{Align a GCIMSSample in retention time with parametric optimized warping}
\usage{
\S4method{alignRt_pow}{GCIMSSample}(
  x,
  ric_ref,
  ric_ref_rt,
  lambdas = pracma::logspace(-2, 4, 31),
  p = 10,
  max_it = 5000,
  lambda1 = 10^6
)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{ric_ref}{The reference Reverse Ion Chromatogram}

\item{ric_ref_rt}{The retention times corresponding to \code{ric_ref}}

\item{lambdas}{a vector with the penalties to test the POW}

\item{p}{By default \code{10}, meaning to use one every \code{10} points to validate.}

\item{max_it}{Maximum number of iterations}

\item{lambda1}{Regularization parameter for second derivative of warp}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in retention time with parametric optimized warping
}
