% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{alignRt_ptw}
\alias{alignRt_ptw}
\title{Align a GCIMSSample in retention time using parametric time warping}
\usage{
alignRt_ptw(object, ric_ref, ric_ref_rt, ploynomial_order = 5)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{ric_ref}{The reference Reverse Ion Chromatogram}

\item{ric_ref_rt}{The retention times corresponding to \code{ric_ref}}

\item{ploynomial_order}{maximum order of the polynomial to be used by default 5}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample in retention time using parametric time warping
}
