% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{apply_cow}
\alias{apply_cow}
\title{Apply Correlation Optimized Warping path on a data matrix}
\usage{
apply_cow(X, Wrp)
}
\arguments{
\item{X}{X (mP x nP) matrix with data for mP row vectors of
length nP to be warped / corrected.}

\item{Warping}{Warping (1 x N x 2) or (mP x nP) interpolation
segment starting points (in "nP" units) after warping
(first slab) and before warping (second slab).}
}
\value{
Xw (mP x nP) warped/corrected data matrix.
}
\description{
Apply Correlation Optimized Warping path on a data matrix
}
\references{
{
 Niels-Peter Vest Nielsen, Jens Micheal Carstensen and Jørn Smedegaard 'Aligning of singel and multiple
          wavelength chromatographic profiles for chemometric data analysis using correlation optimised warping'
          J. Chrom. A 805(1998)17-35

 Correlation optimized warping and dynamic time warping as preprocessing methods for chromatographic Data
           Giorgio Tomasi, Frans van den Berg and Claus Andersson, Journal of Chemometrics 18(2004)231-241
 }
}
\seealso{
Other Alignment functions: 
\code{\link{cow}()},
\code{\link{cut_samples}()},
\code{\link{gcims_alignment}()},
\code{\link{gcims_cut_samples}()},
\code{\link{optimize_cow}()}
}
\author{
Luis Fernandez,  \email{lfernandez@.ub.edu}

Inspired on the work by:

Giorgio Tomasi, \email{gt@kvl.dk}

Frans van den Berg 070821 (GT) \email{fb@kvl.dk}

Royal Agricultural and Veterinary University - Department of Food Science
Quality and Technology - Spectroscopy and Chemometrics group - Denmark
www.models.kvl.dk
}
\concept{Alignment functions}
