% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaw-GCIMSSample.R
\name{as.data.frame.GCIMSSample}
\alias{as.data.frame.GCIMSSample}
\title{Turn the intensity matrix into a data frame}
\usage{
\method{as.data.frame}{GCIMSSample}(
  x,
  row.names = NULL,
  optional = FALSE,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL,
  ...
)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}

\item{...}{unused}
}
\value{
A data frame with \code{dt_ms}, \code{rt_s} and \code{Intensity} columns
}
\description{
Turn the intensity matrix into a data frame
}
