% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{build_peak_table}
\alias{build_peak_table}
\title{Build a peak table}
\usage{
build_peak_table(peak_list_clustered, aggregate_conflicting_peaks = NULL)
}
\arguments{
\item{peak_list_clustered}{The output of \link{group_peak_list}}

\item{aggregate_conflicting_peaks}{\code{NULL} or a function. When we build the peak table, with peaks in rows, samples in
columns, \code{peak_table[i,j]} is the volume of the peak from sample \code{j} in cluster \code{i}. If the clustering process
clusters together two peaks form the same sample, those peaks will conflict in the peak table. \code{NULL} will error
in that case, another function will be applied on the conflicting volumes (e.g \code{mean} or \code{max} would be reasonable options)}
}
\value{
A list with the peak table and the ROI duplicity information
}
\description{
Build a peak table
}
\examples{
pl <- data.frame(
  SampleID = c("S1", "S1", "S2", "S2"),
  cluster = c(1, 2, 1, 2),
  volume = c(10, 20, 8, 18)
)
build_peak_table(pl)
}
