% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPeaks.R
\name{clusterPeaks}
\alias{clusterPeaks}
\title{Group peaks in clusters}
\usage{
clusterPeaks(
  peaks,
  ...,
  distance_method = "euclidean",
  dt_cluster_spread_ms = 0.1,
  rt_cluster_spread_s = 20,
  distance_between_peaks_from_same_sample = 100,
  clustering = list(method = "hclust"),
  verbose = FALSE
)
}
\arguments{
\item{peaks}{A data frame with at least the following columns:
\itemize{
\item "UniqueID" A unique ID for each peak
\item "SampleID" The sample ID the peak belongs to
\item "dt_apex_ms", "rt_apex_s" The peak positions
\item "dt_max_ms", "dt_min_ms", "rt_max_s", "rt_min_s" (for filtering outlier peaks based on their size)
}}

\item{...}{Ignored. All other parameters beyond \code{peaks} should be named}

\item{distance_method}{A string. One of the distance methods from \link[stats:dist]{stats::dist}, "sd_scaled_euclidean" or "mahalanobis"}

\item{dt_cluster_spread_ms, rt_cluster_spread_s}{The typical spread of the clusters. Used for scaling.
dimensions when computing distances. When \code{clustering$method} is \code{"hclust"}, these spreads are used to cut cluster sizes.}

\item{distance_between_peaks_from_same_sample}{The distance between two peaks from the same sample will be set to \code{distance_between_peaks_from_same_sample*max(distance_matrix)}}

\item{clustering}{A named list with "method" and the supported method, as well as further options.
For \code{method = "kmedoids"}, you must provide \code{Nclusters}, with either the number of clusters
to use in the kmedoids algorithm (\link[cluster:pam]{cluster::pam}) or the string \code{"max_peaks_sample"} to use the maximum number of
detected peaks per sample.

For \code{method = "hclust"}, you can provide \code{hclust_method}, with the \code{method} passed to \code{\link[mdendro:linkage]{mdendro::linkage()}}.
For \code{method = "hclust_old"}, you can provide \code{hclust_method}, with the \code{method} passed to \code{\link[stats:hclust]{stats::hclust()}}.}

\item{verbose}{logical, to control printing in the function}
}
\value{
A list with :
\itemize{
\item \code{peak_list_clustered}: The peak list with a "cluster" column
\item \code{cluster_stats}: Cluster statistics (cluster size...)
\item \code{dist}: peak to peak distance object
\item \code{extra_clustering_info}: Arbitrary clustering extra information, that depends on the clustering method
}
}
\description{
Peak grouping function, exposing several options useful for benchmarking.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
peak_list <- readRDS(file.path(dir_in, "peak_list.rds"))

peak_clustering  <- clusterPeaks(
  peaks = peak_list,
  distance_method = "mahalanobis",
  distance_between_peaks_from_same_sample = Inf,
  clustering = list(method = "hclust"),
  verbose = FALSE
)
}
}
