% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{cow}
\alias{cow}
\title{Correlation Optimized Warping with linear interpolation}
\usage{
cow(T, X, Seg, Slack, plot = FALSE, corrpow = 1,
  equal_lengths = FALSE, max_correction = NA, debug = FALSE)
}
\arguments{
\item{T}{array with the target spectra to use as reference}

\item{X}{matrix with data to be warped, one spectra per row}

\item{Seg}{the segment length. The number of segments is N = floor(ncol(X)/Seg)
if Seg is a two row matrix, the first row contains indexes of the T
array, ranging from 1 to length(T). The second row contains the indexes
for the X matrix, ranging from 1 to ncol(X). These indexes define the
boundaries for each segment}

\item{Slack}{maximum range or degree of warping in each segment.}

\item{plot}{logical value (default=FALSE, no plots are made)}

\item{corrpow}{numerical value in [1,4] to determine the correlation power (default=1)}

\item{equal_lengths}{force equal segment lengths for the data and the reference, instead of
filling up the reference with boundary points.
(note that different number of boundaries in the reference and the data
will generate an error) (default=FALSE)}

\item{max_correction}{the maximum correction will be of +- max_corrections points from
the diagonal (default=NA, No maximum)}

\item{debug}{logical to determine if a table with the optimal values of loss function
and predecessor should be returned too (memory consuming in large
problems) (default=FALSE)}
}
\value{
a list with:
    - Warping: list with two elements containing the interpolation segment starting points
               (in X units)
                 - after:  Starting points after warping
                 - before: Starting points before warping
               The difference after-before is the alignment by repositioning segment boundaries;
               useful for comparing correction in different/new objects/samples
    - XWarped: corrected data
    - Diagnos  warping diagnostics: options, segment, slack,
               index in target ("xt", "warping" is shift compared to this) and sample ("xP"),
               search range in "xP", computation time (note: diagnostics are only saved for
               one - the last - signal in "xP")
}
\description{
Correlation Optimized Warping with linear interpolation
}
\references{
{
 Niels-Peter Vest Nielsen, Jens Micheal Carstensen and Jørn Smedegaard 'Aligning of singel and multiple
          wavelength chromatographic profiles for chemometric data analysis using correlation optimised warping'
          J. Chrom. A 805(1998)17-35

 Correlation optimized warping and dynamic time warping as preprocessing methods for chromatographic Data
           Giorgio Tomasi, Frans van den Berg and Claus Andersson, Journal of Chemometrics 18(2004)231-241
 }
}
\seealso{
Other Alignment functions: \code{\link{gcims_alignment}}
}
\author{
Sergio Oller-Moreno,  \email{soller@ibecbarcelona.eu}

Inspired on the work by:

Giorgio Tomasi, \email{gt@kvl.dk}

Frans van den Berg 070821 (GT) \email{fb@kvl.dk}

Royal Agricultural and Veterinary University - Department of Food Science
Quality and Technology - Spectroscopy and Chemometrics group - Denmark
www.models.kvl.dk
}
\concept{Alignment functions}
