% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_samples_functions.R
\name{cut_samples}
\alias{cut_samples}
\title{Cuts retention time and drift time}
\usage{
cut_samples(loaded_sample, ret_cut = NULL, drift_cut = NULL)
}
\arguments{
\item{loaded_sample}{Current sample to be cut.}

\item{ret_cut}{A  vector of two components. Beginning and end
of the retention time cut.}

\item{drift_cut}{A  vector of two components. Beginning and end
of the drift time cut.}
}
\value{
An cut matrix.
}
\description{
Cuts retention time and drift time
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}


}
\seealso{
Other Alignment functions: 
\code{\link{apply_cow}()},
\code{\link{cow}()},
\code{\link{gcims_alignment}()},
\code{\link{gcims_cut_samples}()},
\code{\link{optimize_cow}()}
}
\concept{Alignment functions}
