% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimate-GCIMSDataset.R
\name{decimate,GCIMSDataset-method}
\alias{decimate,GCIMSDataset-method}
\title{Decimate a GCIMS dataset keeping 1 out of n points}
\usage{
\S4method{decimate}{GCIMSDataset}(object, dt_factor = 1L, rt_factor = 1L)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{dt_factor}{Keep one every \code{dt_factor} measurement points in drift time}

\item{rt_factor}{Keep one every \code{rt_factor} measurement points in retention time}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
Decimate a GCIMS dataset keeping 1 out of n points
}
