% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-decimate.R
\name{decimate,GCIMSSample-method}
\alias{decimate,GCIMSSample-method}
\title{Decimates a GCIMS sample}
\usage{
\S4method{decimate}{GCIMSSample}(object, dt_factor = 1L, rt_factor = 1L)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{dt_factor}{Keep one every \code{dt_factor} measurement points in drift time}

\item{rt_factor}{Keep one every \code{rt_factor} measurement points in retention time}
}
\value{
The modified \link{GCIMSSample}
}
\description{
This method assumes that the sample has been low-pass filtered to avoid aliasing issues
}
