% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{decimate}
\alias{decimate}
\title{Decimate an object}
\usage{
decimate(object, ...)
}
\arguments{
\item{object}{An object to decimate}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The object, modified
}
\description{
Decimate an object
}
