% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksImpl1D.R
\name{defaultPeakDetectionCWTParams}
\alias{defaultPeakDetectionCWTParams}
\title{Get the default peakDetectionCWT parameters for GCIMS peak detection}
\usage{
defaultPeakDetectionCWTParams(x, peakwidth_range_xunits)
}
\arguments{
\item{x}{The x axis, to determine units}

\item{peakwidth_range_xunits}{A vector of length 2 with the minimum and maximum peak width.}
}
\value{
A list with the default parameters we will use with \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}.
}
\description{
Our selection of defaults for \code{\link[=peakDetectionCWT]{peakDetectionCWT()}} consist of the following:

\itemize{
\item{\code{exclude0scaleAmpThresh=TRUE}: Does not include the zeroth scale in the default amplitude threshold.}
\item{\code{nearbyPeak=TRUE}: This is already a default, but since we customize nearbyWinSize we prefer to be explicit.}
\item{\code{nearbyWinSize}: This is set to the maximum desired peak width in points (instead of the arbitrary 150 from \code{\link[MassSpecWavelet:identifyMajorPeaks]{MassSpecWavelet::identifyMajorPeaks()}})}
\item{\code{ridgeLength}: We set this to the second calculated scale, that would correspond to a peak with shorter width in our scale preparation.}
\item{\code{excludeBoundariesSize}: Instead of an arbitrary number, we exclude half of the peak minimum width}
\item{\code{SNR.Th}: We keep the default (3) but we explicitly specify it because we consider this a sensitive parameter}
\item{\code{minNoiseLevel}: \code{\link[MassSpecWavelet:identifyMajorPeaks]{MassSpecWavelet::identifyMajorPeaks()}} uses a default of 0.001 and we keep that default}
}
}
\examples{
chromat1 <- GCIMSChromatogram(
  retention_time = seq(from = 0, to = 10, length.out = 200),
  intensity = 1:200
)

peakDetectionCWTParams <- defaultPeakDetectionCWTParams(
  x = rtime(chromat1),
  peakwidth_range_xunits = c(5, 25)
)
peakDetectionCWTParams
}
\keyword{internal}
