% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-DelayedDatasetBase.R
\name{DelayedDatasetBase}
\alias{DelayedDatasetBase}
\title{DelayedDatasetBase}
\description{
Base R6 class to handle a dataset that can queue execute operations on it

This is an abstract class. See \link{DelayedDatasetRAM} and \link{DelayedDatasetDisk}
for specific implementations.

This class is not exported, but if you want to use it reach us at
\url{https://github.com/sipss/GCIMS/issues/} and we will export it.
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sampleNames}}{The character vector with unique sample names (virtual method)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DelayedDatasetBase-realize}{\code{DelayedDatasetBase$realize()}}
\item \href{#method-DelayedDatasetBase-new}{\code{DelayedDatasetBase$new()}}
\item \href{#method-DelayedDatasetBase-registerOptimization}{\code{DelayedDatasetBase$registerOptimization()}}
\item \href{#method-DelayedDatasetBase-appendDelayedOp}{\code{DelayedDatasetBase$appendDelayedOp()}}
\item \href{#method-DelayedDatasetBase-hasDelayedOps}{\code{DelayedDatasetBase$hasDelayedOps()}}
\item \href{#method-DelayedDatasetBase-getSample}{\code{DelayedDatasetBase$getSample()}}
\item \href{#method-DelayedDatasetBase-history_as_list}{\code{DelayedDatasetBase$history_as_list()}}
\item \href{#method-DelayedDatasetBase-pending_as_list}{\code{DelayedDatasetBase$pending_as_list()}}
\item \href{#method-DelayedDatasetBase-clone}{\code{DelayedDatasetBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-realize"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-realize}{}}}
\subsection{Method \code{realize()}}{
Execute all the queued operations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$realize(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments used by children classes, passed on to realize_impl(...)}

\item{\code{dataset}}{The dataset where extracted results will be aggregated on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-new}{}}}
\subsection{Method \code{new()}}{
Create a new object, initialize the slots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$new(dataset, dataset_class = NULL, sample_class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{The dataset R6 object where aggregated results from queued operations are stored}

\item{\code{dataset_class}}{The class of the given dataset, used just to validate the contract between
the delayed actions and the dataset. If \code{NULL} action return values are not checked}

\item{\code{sample_class}}{The class of the samples in the dataset, used just to validate the contract between
the delayed actions and the samples. If \code{NULL} action return values are not checked}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-registerOptimization"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-registerOptimization}{}}}
\subsection{Method \code{registerOptimization()}}{
The list of queued operations are usually processed sequentially over each sample when
the delayed dataset is realized.

In some cases we want to operate on the list of queued operations before processing them,
for instance to remove redundant operations or change their order, to improve performance.

Write a function that takes and returns a list of queued operations when used
as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{queued_ops <- your_function(queued_ops)
}\if{html}{\out{</div>}}

And call \code{obj$registerOptimization(your_function)} to register it, so every time
\code{obj$realize()} is called, the list of queued operations will be passed to your function
and optimized accordingly.

You can register as many optimizations as you like.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$registerOptimization(optimization)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimization}}{A function that takes a list of queued operations and
returns a more optimized version of it.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-appendDelayedOp"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-appendDelayedOp}{}}}
\subsection{Method \code{appendDelayedOp()}}{
Queues an operation to the dataset so it will run afterwards
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$appendDelayedOp(operation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{A \link{GCIMSDelayedOp} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified GCIMSDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-hasDelayedOps"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-hasDelayedOps}{}}}
\subsection{Method \code{hasDelayedOps()}}{
Find out if the dataset has pending operations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$hasDelayedOps()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns \code{TRUE} if the dataset has pending operations, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-getSample"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-getSample}{}}}
\subsection{Method \code{getSample()}}{
Get a sample from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$getSample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{Either an integer (sample index) or a string (sample name)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sample object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-history_as_list"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-history_as_list}{}}}
\subsection{Method \code{history_as_list()}}{
Get a list that describes the already executed operations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$history_as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list, with the operations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-pending_as_list"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-pending_as_list}{}}}
\subsection{Method \code{pending_as_list()}}{
Get a list that describes the queued operations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$pending_as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list, with the operations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
