% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-DelayedDatasetDisk.R
\name{DelayedDatasetDisk}
\alias{DelayedDatasetDisk}
\title{DelayedDatasetDisk}
\description{
A class that contains a dataset where its samples are backed on disk.
Each sample is stored on a file, and when queued operations are executed files
are loaded and saved on a new folder.
}
\details{
This class is not exported, but if you want to use it reach us at
\url{https://github.com/sipss/GCIMS/issues/} and we will export it.
}
\keyword{internal}
\section{Super class}{
\code{\link[GCIMS:DelayedDatasetBase]{GCIMS::DelayedDatasetBase}} -> \code{DelayedDatasetDisk}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sampleNames}}{The character vector with unique sample names. Renaming samples renames files on \code{obj$getCurrentDir()} as well}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DelayedDatasetDisk-new}{\code{DelayedDatasetDisk$new()}}
\item \href{#method-DelayedDatasetDisk-getSample}{\code{DelayedDatasetDisk$getSample()}}
\item \href{#method-DelayedDatasetDisk-getCurrentDir}{\code{DelayedDatasetDisk$getCurrentDir()}}
\item \href{#method-DelayedDatasetDisk-clone}{\code{DelayedDatasetDisk$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="appendDelayedOp"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-appendDelayedOp'><code>GCIMS::DelayedDatasetBase$appendDelayedOp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="hasDelayedOps"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-hasDelayedOps'><code>GCIMS::DelayedDatasetBase$hasDelayedOps()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="history_as_list"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-history_as_list'><code>GCIMS::DelayedDatasetBase$history_as_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="pending_as_list"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-pending_as_list'><code>GCIMS::DelayedDatasetBase$pending_as_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="realize"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-realize'><code>GCIMS::DelayedDatasetBase$realize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="registerOptimization"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-registerOptimization'><code>GCIMS::DelayedDatasetBase$registerOptimization()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetDisk-new"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetDisk-new}{}}}
\subsection{Method \code{new()}}{
Create a delayed dataset on disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetDisk$new(
  samples,
  scratch_dir,
  keep_intermediate = FALSE,
  dataset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A named vector. The names are sample ids, the values are
either filenames or sample objects. If they are not filenames then the objects are dumped to disk.
If they are filenames, the filenames are relative to \code{base_dir}.}

\item{\code{scratch_dir}}{The directory where samples being processed will be saved}

\item{\code{keep_intermediate}}{A logical value, whether intermediate realization steps should be saved.}

\item{\code{dataset}}{The dataset that stores aggregated results from the processing}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The DelayedDatasetDisk object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetDisk-getSample"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetDisk-getSample}{}}}
\subsection{Method \code{getSample()}}{
Get a sample from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetDisk$getSample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{Either an integer (sample index) or a string (sample name)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sample object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetDisk-getCurrentDir"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetDisk-getCurrentDir}{}}}
\subsection{Method \code{getCurrentDir()}}{
Get the path to the location of the processed samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetDisk$getCurrentDir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The full path to the directory where samples are saved
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetDisk-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetDisk-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetDisk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
