% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-DelayedDatasetRAM.R
\name{DelayedDatasetRAM}
\alias{DelayedDatasetRAM}
\title{DelayedDatasetRAM}
\description{
A class that contains a dataset where its samples are stored in memory (RAM).
This is usually faster than storing them on files, but the size of the
dataset becomes limited by the amount of RAM available.
}
\details{
This class is not exported, but if you want to use it reach us at
\url{https://github.com/sipss/GCIMS/issues/} and we will export it.
}
\keyword{internal}
\section{Super class}{
\code{\link[GCIMS:DelayedDatasetBase]{GCIMS::DelayedDatasetBase}} -> \code{DelayedDatasetRAM}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sampleNames}}{The character vector with unique sample names.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DelayedDatasetRAM-new}{\code{DelayedDatasetRAM$new()}}
\item \href{#method-DelayedDatasetRAM-getSample}{\code{DelayedDatasetRAM$getSample()}}
\item \href{#method-DelayedDatasetRAM-clone}{\code{DelayedDatasetRAM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="appendDelayedOp"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-appendDelayedOp'><code>GCIMS::DelayedDatasetBase$appendDelayedOp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="hasDelayedOps"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-hasDelayedOps'><code>GCIMS::DelayedDatasetBase$hasDelayedOps()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="history_as_list"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-history_as_list'><code>GCIMS::DelayedDatasetBase$history_as_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="pending_as_list"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-pending_as_list'><code>GCIMS::DelayedDatasetBase$pending_as_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="realize"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-realize'><code>GCIMS::DelayedDatasetBase$realize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GCIMS" data-topic="DelayedDatasetBase" data-id="registerOptimization"><a href='../../GCIMS/html/DelayedDatasetBase.html#method-DelayedDatasetBase-registerOptimization'><code>GCIMS::DelayedDatasetBase$registerOptimization()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetRAM-new"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetRAM-new}{}}}
\subsection{Method \code{new()}}{
Create a delayed dataset on RAM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetRAM$new(samples, sample_class = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A named list of samples.}

\item{\code{sample_class}}{The class of the samples in the dataset, used just to validate the contract between
the delayed actions and the samples. If \code{NULL} action return values are not checked}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetRAM-getSample"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetRAM-getSample}{}}}
\subsection{Method \code{getSample()}}{
Get a sample from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetRAM$getSample(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{Either an integer (sample index) or a string (sample name)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sample object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DelayedDatasetRAM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DelayedDatasetRAM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedDatasetRAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
