% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-DelayedOperation.R
\docType{class}
\name{DelayedOperation-class}
\alias{DelayedOperation-class}
\title{Delayed Operation class}
\description{
DelayedOperation is an S4 class to store a delayed operation

Delayed operations are not applied to the dataset immediately, but rather
when some data from the dataset is required. When working on large datasets,
keeping all samples in RAM may be impossible, and the \link{DelayedDatasetDisk}
architecture becomes convenient, where samples are stored in a directory, loaded
processed and saved individually.

Under such arquitecture, it is more efficient to load a sample, run as many operations
as possible on it and save the sample, instead of loading a sample, running one
operation, saving the sample.

See how to create such delayed operations and more details at
\code{vignette("creating-a-workflow-step", package = "GCIMS")}.
}
\section{Slots}{

\describe{
\item{\code{name}}{A named for de delayed operation, only used for printing.}

\item{\code{fun}}{A function that takes a sample object and returns a sample object, usually with some change (filtered,...)}

\item{\code{params}}{A named list with additional arguments to be passed to \code{fun}}

\item{\code{params_iter}}{A named list with additional arguments to be passed to
\code{fun}. Compared to \code{params}, each argument must be a named list of length the number of samples, so
each sample will receive its corresponding parameter according to its name}

\item{\code{fun_extract}}{A function that takes the modified sample object returned
by \code{fun} and extracts some component out of it. This component will be stored in the dataset for faster access.}

\item{\code{fun_aggregate}}{A function that takes a dataset object and a list of extracted results (the output of all \code{fun_extract} calls)
and modifies the dataset.}
}}

