% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{describeAsList}
\alias{describeAsList}
\title{This function returns a list of relevant object-related information to be
used when printing.}
\usage{
describeAsList(object, ...)
}
\description{
This function returns a list of relevant object-related information to be
used when printing.
}
\keyword{internal}
