% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_three_ketones_dataset.R
\name{download_three_ketones_dataset}
\alias{download_three_ketones_dataset}
\title{Download three samples (6-ketone mixture)}
\usage{
download_three_ketones_dataset(outdir = "2021-mixture-six-ketones-demo")
}
\arguments{
\item{outdir}{Name of the directory where the samples will be saved}
}
\value{
Nothing (Files are created in the given folder)
}
\description{
This function downloads three samples in .mea.gz format. It is useful
to run the introductory vignette.
}
\examples{
\dontrun{
download_three_ketones_dataset(outdir = "sample_dataset")
list.files("sample_dataset")
}

}
