% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{driftTime}
\alias{driftTime}
\title{Get drift time vector}
\usage{
driftTime(object)
}
\arguments{
\item{object}{An object with drift time}
}
\value{
A numeric vector with the drift time
}
\description{
Get drift time vector
}
