% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dt_rt_range_normalization.R
\name{dt_rt_range_normalization}
\alias{dt_rt_range_normalization}
\title{Index normalization}
\usage{
dt_rt_range_normalization(
  dt = numeric(0L),
  rt = numeric(0L),
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL
)
}
\arguments{
\item{dt}{A numeric vector with the drift time}

\item{rt}{A numeric vector with the retention time}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}
}
\value{
A named list with class \code{dt_rt_range_normalization}. The elements of the list are:
\itemize{
\item \code{dt_ms_min}, \code{dt_ms_max}: Drift time in milliseconds (min and max). This is the same value if the user did want a single value
\item \code{dt_idx_min}, \code{dt_idx_max}: The indices corresponding to the requested range limits in drift time
\item \code{dt_logical}: A logical vector of length equal to the given drift time vector, with \code{TRUE} where the range is included and \code{FALSE} elsewhere. Useful for indexing.
\item \code{rt_s_min}, \code{rt_s_max}: Retention time in seconds (min and max). This is the same value if the user did want a single value
\item \code{rt_idx_min}, \code{rt_idx_max}: The indices corresponding to the requested range limits in retention time.
\item \code{rt_logical}: A logical vector of length equal to the given retention time vector, with \code{TRUE} where the range is included and \code{FALSE} elsewhere. Useful for indexing.
}
}
\description{
Users may want to specify:
}
\details{
\itemize{
\item A single retention/drift time to extract, in physical units or in indices
\item A range of retention/drift time to extract, in physical units or in indices
}

This function is internal to the \link{GCIMS-package} and it is meant to be used
to normalize those specifications, returning an object that computes the rest
so the caller function can easily get the indices or physical units it needs.

The behaviour is intuitive in corner cases, where the conversion from physical
units to indices may have rounding issues:
\itemize{
\item If a single value is requested in physical units the closest one will be returned. If the
requested single value is outside the valid range, then an error is given. A tolerance of
one time step is accepted to prevent rounding issues.
\item If a range value is requested in physical units, all the points within the range
are included. The interval is considered closed at both ends, for consistency with how
indexes behave in R.
}
}
\keyword{internal}
