% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-intensity-GCIMSChromatogram.R
\name{dtime,GCIMSChromatogram-method}
\alias{dtime,GCIMSChromatogram-method}
\title{Get the drift time of the chromatogram}
\usage{
\S4method{dtime}{GCIMSChromatogram}(object)
}
\arguments{
\item{object}{A GCIMSChromatogram}
}
\value{
The drift time where this chromatogram was extracted from (in ms)
}
\description{
Get the drift time of the chromatogram
}
\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram}},
\code{\link{GCIMSChromatogram-class}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{findPeaks,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{rtime,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
