% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-get_dtime_rtime.R
\name{dtime,GCIMSDataset-method}
\alias{dtime,GCIMSDataset-method}
\title{Get A reference drift time vector for the dataset}
\usage{
\S4method{dtime}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A GCIMSDataset}
}
\value{
a drift time vector
}
\description{
Get A reference drift time vector for the dataset
}
