% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dtime}
\alias{dtime}
\title{Get drift time vector}
\usage{
dtime(object, ...)
}
\arguments{
\item{object}{An object with drift time}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A numeric vector with the drift time
}
\description{
Get drift time vector
}
\examples{
x <- dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
dtime(x) # c(1,2)

}
