% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-GCIMSDataset.R
\name{estimateBaseline,GCIMSDataset-method}
\alias{estimateBaseline,GCIMSDataset-method}
\title{Estimate the baseline of a GCIMS Sample using a connect local minima algorithm}
\usage{
\S4method{estimateBaseline}{GCIMSDataset}(
  object,
  dt_peak_fwhm_ms,
  dt_region_multiplier,
  rt_length_s,
  remove = TRUE
)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{dt_peak_fwhm_ms}{Full Width at Half Maximum in milliseconds. Used to
determine the length of the regions where local minima are searched.}

\item{dt_region_multiplier}{A multiplier to calculate the region}

\item{rt_length_s}{The length of the baseline region. It should be
comparable or longer than the peak width}

\item{remove}{A boolean, if TRUE it removes the baseline from the intensity}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
Estimate the baseline of a GCIMS Sample using a connect local minima algorithm
}
