% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-getTIS_getRIC.R
\name{extract_RIC_and_TIS}
\alias{extract_RIC_and_TIS}
\title{Extract the Reverse Ion Chromatogram and Total Ion Spectrum from the samples}
\usage{
extract_RIC_and_TIS(object)
}
\arguments{
\item{object}{A GCIMSDataset object}
}
\value{
The \link{GCIMSDataset} object, with the delayed operation to compute the
RIC and TIS matrices.
}
\description{
Extract the Reverse Ion Chromatogram and Total Ion Spectrum from the samples
}
