% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{fill_concentration}
\alias{fill_concentration}
\title{Generate profiles of concentration including non-idealities}
\usage{
fill_concentration(
  C,
  column,
  max_pos,
  max_value,
  rise_time,
  fall_time,
  rt_shift,
  nscans
)
}
\arguments{
\item{C}{Matrix. A matrix of concentrations to be filled}

\item{column}{Column matrix to be filled with its concentration profile.}

\item{max_pos}{Integer. The postion for the maximum of concentration
in the profile (in indexes)}

\item{max_value}{Numeric. Maximum concentration of the profile.}

\item{rise_time}{Integer. Number of indexes needed to reach the
maximum value of concentration from the baseline.}

\item{fall_time}{Integer. Number of indexes needed to reach the baseline
from the maximum value of concentration.}

\item{rt_shift}{Vector of integers. The shift in retention time indexes for
each of the peaks.}

\item{nscans}{Integer. Length of the concentration profile (in indexes). It
its equal to the number of spectra or scans.}
}
\value{
A matrix of Concentration profiles.
}
\description{
Generate profiles of concentration including non-idealities
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
