% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-filter.R
\name{filterDt,GCIMSDataset-method}
\alias{filterDt,GCIMSDataset-method}
\title{Filter GCIMSDataset samples by drift time}
\usage{
\S4method{filterDt}{GCIMSDataset}(object, dt_range)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}

\item{dt_range}{A numeric vector of length 2 with the drift time range to keep, in milliseconds}
}
\description{
Filter GCIMSDataset samples by drift time
}
\examples{
base_dir <- system.file("extdata", "sample_formats", package = "GCIMS")
annot <- data.frame(SampleID = "Sample1", FileName = "small.mea.gz")
dataset <- GCIMSDataset(annot, base_dir)
filterDt(dataset, dt_range = c(5, 10))
}
