% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{filterDt}
\alias{filterDt}
\title{Filter in Drift time}
\usage{
filterDt(object, ...)
}
\arguments{
\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The object, modified
}
\description{
Filter in Drift time
}
