% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-GCIMSDataset.R
\name{filterRt,GCIMSDataset-method}
\alias{filterRt,GCIMSDataset-method}
\title{Filter GCIMSDataset samples by retention time}
\usage{
\S4method{filterRt}{GCIMSDataset}(object, rt_range)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}
}
\value{
The given object, with a delayed operation to filter retention times
}
\description{
Filter GCIMSDataset samples by retention time
}
\examples{
base_dir <- system.file("extdata", "sample_formats", package = "GCIMS")
annot <- data.frame(SampleID = "Sample1", FileName = "small.mea.gz")
dataset <- GCIMSDataset$new(annot, base_dir)
filterRt(dataset, rt_range = c(5, 50))
}
