% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-filter.R
\name{filterRt,GCIMSSample-method}
\alias{filterRt,GCIMSSample-method}
\title{Filter GCIMSSample samples by retention time}
\usage{
\S4method{filterRt}{GCIMSSample}(object, rt_range)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rt_range}{A numeric vector of length 2 with the retention time range to keep, in seconds}
}
\description{
Filter GCIMSSample samples by retention time
}
\examples{
sample_file <- system.file("extdata", "sample_formats", "small.mea.gz", package = "GCIMS")
s <- read_mea(sample_file)
s <- filterRt(s, rt_range = c(5, 50))
}
