% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_alignment.R
\name{find_reference_ric}
\alias{find_reference_ric}
\title{This function provides the index corresponding to the reference Reactant Ion Chromatogram (RIC) to correct
misalignments in retention time.}
\usage{
find_reference_ric(rics)
}
\arguments{
\item{rics}{A matrix. Each row correspond to a different RIC. There are as many RICs as samples.}
}
\value{
An Integer number that indicates the reference sample.
}
\description{
This function provides the index corresponding to the reference Reactant Ion Chromatogram (RIC) to correct
misalignments in retention time.
}
\examples{
rics <- rbind(
  dnorm(1:100, mean=50, sd =1),
  dnorm(1:100, mean=51, sd =1),
  dnorm(1:100, mean=52, sd =1)
)
find_reference_ric(rics) == 2L

}
