% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks-GCIMSChromatogram.R
\name{findPeaks,GCIMSChromatogram-method}
\alias{findPeaks,GCIMSChromatogram-method}
\title{Peak detection for a GCIMSChromatogram}
\usage{
\S4method{findPeaks}{GCIMSChromatogram}(object, ...)
}
\arguments{
\item{object}{A \link{GCIMSChromatogram} object}

\item{...}{
  Arguments passed on to \code{\link[=findPeaksImpl1D]{findPeaksImpl1D}}
  \describe{
    \item{\code{verbose}}{If \code{TRUE} information will be printed on screen}
    \item{\code{length_in_xunits}}{Length of the filter used to compute the second derivative. See details.}
    \item{\code{peakwidth_range_xunits}}{A vector of length 2 with the minimum and maximum peak width. See details.}
    \item{\code{peakDetectionCWTParams}}{Additional parameters to \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}. See details.}
    \item{\code{extension_factor}}{A number to extend the ROIs beyond their default size}
    \item{\code{iou_overlap_threshold}}{A number, between 0 and 1. Pairs of ROIs with an intersection over union larger than this threshold are merged.}
    \item{\code{debug}}{If TRUE, return as well the debug information}
  }}
}
\value{
The modified \link{GCIMSChromatogram}, with a peak list
}
\description{
Peak detection for a GCIMSChromatogram
}
\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram}},
\code{\link{GCIMSChromatogram-class}},
\code{\link{dtime,GCIMSChromatogram-method}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{rtime,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
