% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-peaks.R
\name{findPeaks,GCIMSDataset-method}
\alias{findPeaks,GCIMSDataset-method}
\title{Peak detection on the GCIMS dataset}
\usage{
\S4method{findPeaks}{GCIMSDataset}(object, noise_level = 3, verbose = FALSE)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{noise_level}{Scalar number. The number of times the standard deviation
above the noise level needed to detect a peak.}

\item{verbose}{If \code{TRUE}, debug information will be printed}
}
\value{
The modified \link{GCIMSDataset}, with a peak list
}
\description{
Peak detection on the GCIMS dataset
}
