% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-AllGenerics.R
\name{findPeaks}
\alias{findPeaks}
\title{Find Peaks in an object}
\usage{
findPeaks(object, ...)
}
\arguments{
\item{object}{An object to find peaks on}

\item{...}{Additional arguments for downstream methods}
}
\value{
The object, with found peaks
}
\description{
Find Peaks in an object
}
