% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaksImpl}
\alias{findPeaksImpl}
\title{Peak and ROI detection}
\usage{
findPeaksImpl(
  drift_time,
  retention_time,
  int_mat,
  verbose = FALSE,
  dt_length_ms = 0.07,
  rt_length_s = 8,
  dt_peakwidth_range_ms = c(0.15, 0.4),
  rt_peakwidth_range_s = c(10, 25),
  dt_peakDetectionCWTParams = list(exclude0scaleAmpThresh = TRUE),
  rt_peakDetectionCWTParams = list(exclude0scaleAmpThresh = TRUE),
  dt_extension_factor = 0,
  rt_extension_factor = 0,
  exclude_rip = FALSE,
  iou_overlap_threshold = 0.2,
  debug_idx = list(dt = NULL, rt = NULL)
)
}
\arguments{
\item{drift_time}{The drift time vector}

\item{retention_time}{The retention time vector}

\item{int_mat}{The intensity matrix, of dimensions \code{length(drift_time)} rows and \code{length(retention_time)} columns}

\item{verbose}{If \code{TRUE} information will be printed on screen}

\item{dt_length_ms, rt_length_s}{Length of the filters used to compute the
second derivative. See details.}

\item{dt_peakwidth_range_ms, rt_peakwidth_range_s}{A vector of length 2 with the minimum and maximum peak width. See details}

\item{dt_peakDetectionCWTParams, rt_peakDetectinoCWTParams}{Additional parameters to \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}. See details}

\item{exclude_rip}{Whether to exclude ROIs with a drift time apex smaller than the RIP drift time end.}

\item{iou_overlap_threshold}{A number, between 0 and 1. Pairs of ROIs with an intersection over union larger than this threshold are merged.}

\item{debug_idx}{A list with two numeric vectors named \code{dt} and \code{rt} each of them having a the indices to where debug info is kept}
}
\value{
A list with the \code{peak_list} and \code{debug_info} elements.
}
\description{
Detects regions of interest with peaks in a sample.
}
\details{
Peaks are detected on the intensity matrix.

In detail, the approach is as follows:
\enumerate{
\item We compute the second derivative with respect to the drift and retention times.
\item Based on the given peak width ranges, mexican hat wavelets are scaled
\item Peaks are detected using \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}.
\item We intersect the maxima to discard saddle points
\item We merge similar ROIs using a threshold on the intersection over union
\item Get some ROI metrics and return.
}

For the MassSpecWavelet-based peak detection, the \code{scales} are computed based on
the requested peak widths. Besides, the scales, further tuning beyond the
MassSpecWavelet defaults is possible through the \verb{*_peakDetectionCWTParams} argument.
By default, the only change we introduce is the \code{exclude0scaleAmpThresh = TRUE}
which is a reasonable peak detection setting not enabled in MassSpecWavelet
for backwards compatibility reasons.
}
\keyword{internal}
