% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksImpl1D.R
\name{findPeaksImpl1D}
\alias{findPeaksImpl1D}
\title{Peak and ROI detection (1D)}
\usage{
findPeaksImpl1D(
  x,
  y,
  verbose = FALSE,
  length_in_xunits = 0.07,
  peakwidth_range_xunits = c(0.15, 0.4),
  peakDetectionCWTParams = list(exclude0scaleAmpThresh = TRUE),
  extension_factor = 0,
  iou_overlap_threshold = 0.2,
  debug = FALSE
)
}
\arguments{
\item{x}{The x axis, to determine units}

\item{y}{The vector where to find peaks, of dimensions \code{length(x)}}

\item{verbose}{If \code{TRUE} information will be printed on screen}

\item{length_in_xunits}{Length of the filter used to compute the second derivative. See details.}

\item{peakwidth_range_xunits}{A vector of length 2 with the minimum and maximum peak width. See details.}

\item{peakDetectionCWTParams}{Additional parameters to \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}. See details.}

\item{extension_factor}{A number to extend the ROIs beyond their default size}

\item{iou_overlap_threshold}{A number, between 0 and 1. Pairs of ROIs with an intersection over union larger than this threshold are merged.}

\item{debug}{If TRUE, return as well the debug information}
}
\value{
A list with the \code{peak_list} and \code{debug_info} elements.
}
\description{
Detects regions of interest with peaks in a signal
}
\details{
Peaks are detected on the intensity vector.

In detail, the approach is as follows:
\enumerate{
\item We compute the second derivative with respect to the drift and retention times.
\item Based on the given peak width ranges, mexican hat wavelets are scaled
\item Peaks on are detected using \code{\link[MassSpecWavelet:peakDetectionCWT]{MassSpecWavelet::peakDetectionCWT()}}.
\item We merge similar ROIs using a threshold on the 1D-intersection over union
\item Get some ROI metrics and return.
}

For the MassSpecWavelet-based peak detection, the \code{scales} are computed based on
the requested peak widths. Besides, the scales, further tuning beyond the
MassSpecWavelet defaults is possible through the \code{peakDetectionCWTParams} argument.
By default, the only change we introduce is the \code{exclude0scaleAmpThresh = TRUE}
which is a reasonable peak detection setting not enabled in MassSpecWavelet
for backwards compatibility reasons.
}
\keyword{internal}
