% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{force_interpolation}
\alias{force_interpolation}
\title{Forces that sampling frequencies in retention or drift times are not constant}
\usage{
force_interpolation(sampling_period, constant_sampling, axis_length)
}
\arguments{
\item{sampling_period}{Numeric. The actual sampling period.}

\item{constant_sampling}{Logical. TRUE: no modification is needed. FALSE: an
error in sampling is computed.
in the profile (in indexes)}

\item{axis_length}{Numeric. Maximum concentration of the profile.}
}
\value{
A vector of errors in sampling to be added to the actual retention or
drift time vectors. If constant sampling is TRUE, the error vector is filled with zeros.
}
\description{
Forces that sampling frequencies in retention or drift times are not constant
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
