% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{GCIMS-class}
\alias{GCIMS-class}
\title{An object to store GC-IMS datasets}
\description{
GCIMS is an S4 class to store one or more samples
with the same drift and retention time ranges.
}
\details{
The actual spectra is stored in the `data` slot, in a 3-D array,
where the first index corresponds to drift time, the second to retention time
and the third one to the sample.
}
