% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GCIMS-methods}
\alias{GCIMS-methods}
\alias{.DollarNames.GCIMS}
\alias{$.GCIMS}
\alias{$<-.GCIMS}
\alias{[.GCIMS}
\alias{subset.GCIMS}
\alias{subset}
\title{Methods for the GC-IMS class}
\usage{
\method{.DollarNames}{GCIMS}(x, pattern = "")

\method{$}{GCIMS}(x, i, ...)

\method{$}{GCIMS}(x, i, ...) <- value

\method{[}{GCIMS}(x, i, j, k, ...)

\method{subset}{GCIMS}(x, subset, dt_idx = NULL, rt_idx = NULL, samples_idx = NULL, ...)
}
\arguments{
\item{x}{ An R object for which valid names after \code{"$"}
    are computed and returned.
  }

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }

\item{subset}{Logical expression metadata filters to keep}

\item{dt_idx}{A vector of drift time indices to keep}

\item{rt_idx}{A vector of retention time indices to keep}

\item{samples_idx}{A vector of sample indices to keep}
}
\value{
\code{$}: metadata column \code{i} for object \code{x};
\strong{note}: unlike \code{[[}, \code{$} drops the shape of the metadata
to return a vector instead of a data frame

\code{$<-}: object \code{x} with metadata \code{value} saved as
\code{i}

\code{[}: object \code{x} with features \code{i} and cells \code{j}

\code{subset}: A subsetted \code{GCIMS} object
}
\description{
Methods for the GC-IMS class
}
\section{Functions}{
\itemize{
\item \code{.DollarNames.GCIMS}: Autocompletion for \code{$} access on a
\code{GCIMS} object

\item \code{$.GCIMS}: Metadata access for \code{GCIMS} objects

\item \code{$<-.GCIMS}: Metadata setter for \code{GCIMS} objects

\item \code{[.GCIMS}: Simple subsetter for \code{GCIMS} objects

\item \code{subset.GCIMS}: Subset a \code{\link{GCIMS}} object
}}

\examples{
\dontrun{
# Get metadata using `$'
head(mydataset$groups)
}

# Add metadata using the `$' operator
set.seed(42)
mydataset$value <- sample(1:3, size = length(mydataset), replace = TRUE)
head(mydataset[["value"]])

# `[' examples

# `subset' examples
subset(mydataset, subset = Sex > 'Male')
subset(mydataset, subset = Sex > 'Male', rt = 1:100, dt = 10:20)

}
\seealso{
\code{\link[base]{subset}} \code{\link{WhichSamples}}
}
