% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMS-package.R
\docType{package}
\name{GCIMS-package}
\alias{GCIMS}
\alias{GCIMS-package}
\title{GCIMS: Pre-processing for GC-IMS data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package for the analysis of ion mobility spectrometry (IMS) measurements, as well as samples from multicapillary columns coupled with IMS (MCC-IMS) and gas chromatography coupled to ion mobility spectrometry (GC-IMS). The package provides a complete workflow for the analysis, importing the data, preprocessing the spectra as well as classification and regression techniques for the modelling of the spectra. The package also includes visualization helpers, to represent topographic plots, extracted and total ion chromatograms and IMS spectra.
}
\author{
\strong{Maintainer}: Sergio Oller Moreno \email{soller@ibecbarcelona.eu} (\href{https://orcid.org/0000-0002-8994-1549}{ORCID})

Authors:
\itemize{
  \item Celia Mallafré Muro \email{cmallafre@ibecbarcelona.eu} (\href{https://orcid.org/0000-0001-8247-1719}{ORCID})
  \item Luis Fernandez \email{lfernandez@ibecbarcelona.eu} (\href{https://orcid.org/0000-0001-9790-6287}{ORCID})
  \item Eduardo Caballero Saldivar \email{ecaballero@ibecbarcelona.eu} (\href{https://orcid.org/0000-0002-9698-0749}{ORCID})
  \item Arnau Blanco Borrego \email{blancoarnau@gmail.com} (\href{https://orcid.org/0000-0002-5287-7491}{ORCID})
  \item Antonio Pardo Matínez \email{a.pardo@ub.edu} (\href{https://orcid.org/0000-0003-4369-544X}{ORCID})
  \item Santiago Marco Colás \email{smarco@ibecbarcelona.eu} (\href{https://orcid.org/0000-0003-2663-2965}{ORCID})
}

Other contributors:
\itemize{
  \item Institute for Bioengineering of Catalonia [copyright holder]
}

}
\keyword{internal}
