% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW.R,
%   R/alignment_functions_NEW_2.R
\name{gcims_align_rt}
\alias{gcims_align_rt}
\title{Correction of retention time axis using Parametric Time Warping (PTW)}
\usage{
gcims_align_rt(aux_list, ric_ref, correction_type)

gcims_align_rt(aux_list, ric_ref, correction_type)
}
\arguments{
\item{aux_list}{Object containing GCIMS data and metadata (one sample).}

\item{ric_ref}{A vector. The reference Reactant Ion Chromatogram.                                        as reference for aligning in retention time.}

\item{correction_type}{Numeric. Integer between 0 and 5. If zero, no correction
in retention time is needed. Between 1 and 5 it indicates
the polynomial degree of the warping function.}
}
\value{
And object containing GCIMS data and metadata (one sample).

And object containing GCIMS data and metadata (one sample).
}
\description{
This chromatrogram computes the RIC of the reference samples and
uses it to correct each of the EIC of a sample, for the all the selected
samples on the dataset. The degree ot the polynomial that relates the
retention time axes of the reference and the sample to be corrected
is optimized using the average correlation between samples and reference as a figure
of merit.

This chromatrogram computes the RIC of the reference samples and
uses it to correct each of the EIC of a sample, for the all the selected
samples on the dataset. The degree ot the polynomial that relates the
retention time axes of the reference and the sample to be corrected
is optimized using the average correlation between samples and reference as a figure
of merit.
}
