% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{gcims_alignment}
\alias{gcims_alignment}
\title{Peak alignment using Correlation Optimized Warping (COW)}
\usage{
gcims_alignment(dir_in, dir_out, samples, time, seg_vector, slack_vector)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}

\item{seg_vector}{Vector of segment lengths.}

\item{slack_vector}{Vector of slacks.}
}
\value{
An aligned gcims dataset.
}
\description{
Peak alignment using Correlation Optimized Warping (COW)
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3, 7)

# Example of Drift time Alignment :
# Before:
gcims_plot_spec(dir_in, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")

# After:
time <- "Drift"
seg_vector <- seq(from = 10, to = 30, by = 5)
slack_vector <- seq(from = 1, to = (seg_vector[length(seg_vector)] - 4), by = 3)
gcims_alignment(dir_in, dir_out, samples, time, seg_vector, slack_vector)
gcims_plot_spec(dir_out, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\references{
{
 Niels-Peter Vest Nielsen, Jens Micheal Carstensen and Jørn Smedegaard 'Aligning of singel and multiple
          wavelength chromatographic profiles for chemometric data analysis using correlation optimised warping'
          J. Chrom. A 805(1998)17-35

 Correlation optimized warping and dynamic time warping as preprocessing methods for chromatographic Data
           Giorgio Tomasi, Frans van den Berg and Claus Andersson, Journal of Chemometrics 18(2004)231-241
 }
}
\author{
Sergio Oller-Moreno,  \email{soller@ibecbarcelona.eu}

Inspired on the work by:

Giorgio Tomasi, \email{gt@kvl.dk}

Frans van den Berg 070821 (GT) \email{fb@kvl.dk}

Royal Agricultural and Veterinary University - Department of Food Science
Quality and Technology - Spectroscopy and Chemometrics group - Denmark
www.models.kvl.dk
}
\concept{Alignment functions}
