% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions.R
\name{gcims_baseline_removal}
\alias{gcims_baseline_removal}
\title{Remove baseline of a dataset using psalsa}
\usage{
gcims_baseline_removal(
  dir_in,
  dir_out,
  samples,
  time,
  lambda = 1e+07,
  p = 0.001,
  k = -1
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}

\item{lambda}{Smoothing parameter (generally 1e5 - 1e8)}

\item{p}{Asymmetry parameter}

\item{k}{Peak height parameter (usually 5\% of maximum intensity)}
}
\value{
A baseline removed gcims dataset.
}
\description{
Remove baseline of a dataset using psalsa
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}

}
\seealso{
Other Baseline removal functions: 
\code{\link{psalsa_one}()}
}
\concept{Baseline removal functions}
