% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_figures_of_merit.R
\name{gcims_compute_fom}
\alias{gcims_compute_fom}
\title{Compute figures of merit (fom)}
\usage{
gcims_compute_fom(
  dir_in,
  dir_out,
  peak_list,
  cluster_stats,
  integration_size = c("cluster_roi", "individual_roi"),
  rip_saturation_threshold = 0.1
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where files containing the Figures of Meir of each ROI are
stored.}

\item{peak_list}{A data frame. The peak list where we will add the figures of merit.}

\item{cluster_stats}{A data frame with cluster statistics (ROI limits...)}

\item{integration_size}{Either "cluster_roi" or "individual_roi". When computing the volume,
the integration size can be set individually for each sample or use the reference cluster size given by \code{cluster_stats}.}

\item{rip_saturation_threshold}{A number. The fraction of the maximum RIP. If the RIP at the ROI is below that
fraction, we will consider the peak to be saturated.}
}
\value{
The given peak list, with the added columns
}
\description{
Calculates the area, volume asymmetry and saturation of each peak roi in \code{peak_list}.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
roi_merging_out <- readRDS(file.path(dir_in, "peak_clustering.rds"))

peak_list_fom <- gcims_compute_fom(
  dir_in = dir_in,
  dir_out = dir_out,
  peak_list = roi_merging_out$peak_list_clustered,
  cluster_stats = roi_merging_out$cluster_stats
)
head(peak_list_fom)

files_fom <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files_fom))
}
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_shifting}()}
}
\concept{Utility functions}
