% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_of_merit_functions.R
\name{gcims_compute_numpeaks}
\alias{gcims_compute_numpeaks}
\title{Compute the number of peaks in samples}
\usage{
gcims_compute_numpeaks(dir_in, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{Samples to which the SNR is computed
sample characteristics (one per sample class)}
}
\value{
A dummy dataset to be stored in dir_out
}
\description{
Compute the number of peaks in samples
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Figure of merit functions: 
\code{\link{gcims_compute_snr}()}
}
\concept{Figure of merit functions}
