% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_of_merit_functions.R
\name{gcims_compute_snr}
\alias{gcims_compute_snr}
\title{Compute the signal to noise ratio}
\usage{
gcims_compute_snr(dir_in, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{Samples to which the SNR is computed
sample characteristics (one per sample class)}
}
\value{
A dummy dataset to be stored in dir_out
}
\description{
Compute the signal to noise ratio
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Figure of merit functions}
