% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{gcims_create_dummy_params}
\alias{gcims_create_dummy_params}
\title{Create sample params and labels for two different sample classes}
\usage{
gcims_create_dummy_params()
}
\value{
A list of with parameters for class1 (list), parameters for class (list)
and their corresponding labels.
}
\description{
Create sample params and labels for two different sample classes
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
