% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{gcims_create_dummy_sample}
\alias{gcims_create_dummy_sample}
\title{Create a dummy sample taking into account non-idealities in sampling process}
\usage{
gcims_create_dummy_sample(params)
}
\arguments{
\item{params}{A list that include sample characteristics.}
}
\value{
A dummy sample formally equal to the samples readed from an gc-ims
}
\description{
Create a dummy sample taking into account non-idealities in sampling process
}
\examples{
params <- gcims_create_dummy_params()
M1 <- gcims_create_dummy_sample(params$params_class1)
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
