% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{gcims_create_dummy_set}
\alias{gcims_create_dummy_set}
\title{Create a dummy dataset}
\usage{
gcims_create_dummy_set(dir_in, dir_out, samples_per_class, params)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples_per_class}{Integer. Number of sample repetitions per class}

\item{params}{A list that include two list of
sample characteristics (one per sample class)}
}
\value{
A dummy dataset to be stored in dir_out
}
\description{
Create a dummy dataset
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{modgauss}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
