% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_decimate.R
\name{gcims_decimate}
\alias{gcims_decimate}
\title{Data Smoothing using Savitzky-Golay filters}
\usage{
gcims_decimate(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{filter_length}{Numerical. Length of the filter.}

\item{polynomial_order}{Numerical. Order of the polynomial.}
}
\value{
A filtered  gcims dataset.
}
\description{
Data Smoothing using Savitzky-Golay filters
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Smoothing functions: 
\code{\link{gcims_smoothing}()}
}
\concept{Smoothing functions}
