% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_figures_of_merit.R
\name{gcims_figures_of_merit}
\alias{gcims_figures_of_merit}
\title{Figures Of Merit Calculation}
\usage{
gcims_figures_of_merit(dir_in, dir_out, samples, peak_list)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where files containing the Figures of Meir of each ROI are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which their Figures of Merit (FOM) have to be caulated.}

\item{peak_list}{A data frame. The peak list where we will add the figures of merit.}
}
\value{
A Set of S3 objects and a ".csv" file with a table that contains the figures
of merit for each ROI and the ROI information.
}
\description{
Figures Of Merit Calculation
}
\details{
\code{gcims_figures_of_merit} calculates a set of figures of merit for each
ROI of the sample. The FOMs calculated are: the area, the volume, the assymetry
and the saturation.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
roi_selection <- tempfile("dir")
fom <- tempfile("dir")

# Example of Calculating the Figures of Merit'
pl <- gcims_rois_selection(dir_in, roi_selection, samples = c(3, 7), noise_level = 3)
peak_list_fom <- gcims_figures_of_merit(
  dir_in = roi_selection,
  dir_out = fom,
  samples = 3,
  peak_list = pl
)
head(peak_list_fom)
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_rois_fusion}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
