% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx-deprecated-imputation_functions.R
\name{gcims_impute_missing}
\alias{gcims_impute_missing}
\title{Impute missing values}
\usage{
gcims_impute_missing(peak_table, dir_in, cluster_stats)
}
\arguments{
\item{peak_table}{The matrix to be imputed, with samples in rows (see example)}

\item{dir_in}{Directory with baseline removed samples}

\item{cluster_stats}{Cluster statistics (reference ROI limits)}
}
\value{
A matrix with samples in rows, clusters in columns and volumes
as values, without missing values
}
\description{
Compute the volume of ROIS with missing values.
}
\details{
\code{gcims_impute_missing()} uses the coordinates of the reference ROI for each
cluster and substitute the missing value by the volume in this region without
baseline.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
peak_table_obj <- readRDS(file.path(dir_in, "peak_table.rds"))
roi_merging_out <- readRDS(file.path(dir_in, "roi_fusion_out.rds"))
peak_table_to_impute <- peak_table_obj$peak_table_mat

# Peak table before imputation
peak_table_to_impute


peak_table_imputed <- gcims_impute_missing(peak_table = peak_table_to_impute,
                                               dir_in = dir_in, # use bsln
                                               cluster_stats = roi_merging_out$cluster_stats
                                               )
# Peak table after imputation
peak_table_imputed
}

}
\concept{Imputation functions}
