% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_interpolate}
\alias{gcims_interpolate}
\title{Interpolates data according to retention and drift time sampling frequencies}
\usage{
gcims_interpolate(dir_in, dir_out, samples, time)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the drift time or
in the retention time. It should be introduce "Retention" for correcting
the retention time; or "Drift" for the drift time}

\value{
An interpolated dataset
}
\description{
Interpolates data according to retention and drift time sampling frequencies
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Utility functions:
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_unfold}()},
\code{\link{reshape_samples}()}
}
\concept{Utility functions}
