% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_rois_fusion.R
\name{gcims_merge_rois}
\alias{gcims_merge_rois}
\title{Merge regions of interest}
\usage{
gcims_merge_rois(peak_list_clustered, cluster_stats)
}
\arguments{
\item{peak_list_clustered}{The peak list clustered data frame from \code{\link[=group_peak_list]{group_peak_list()}}}

\item{cluster_stats}{The cluster statistics data frame from \code{\link[=group_peak_list]{group_peak_list()}}}
}
\value{
The peak list clustered table with added \verb{ref_roi_[dt|rt]_[min|max]_idx} columns and the cluster statistics
}
\description{
For each ROI in \code{peak_list_clustered}, we define additional columns with ROI boundaries.
The center of the ROI is not changed, but the limits are taken so the size of the ROI is
the same for all ROIs in the same cluster, using the median size of the ROIs in the cluster.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
clustering <- readRDS(file.path(dir_in, "peak_clustering.rds"))

merge_rois_out <- gcims_merge_rois(clustering$peak_list_clustered, clustering$cluster_stats)
samplestocheck <- merge_rois_out$peak_list_clustered[
  which(merge_rois_out$peak_list_clustered$cluster == "Cluster01"),
]
library(ggplot2)
ggplot() +
 geom_rect(
   data = samplestocheck,
   mapping = aes(
     xmin = dt_min_idx,
     xmax = dt_max_idx,
     ymin = rt_min_idx,
     ymax = rt_max_idx,
     color = SampleID,
     fill = SampleID
   ),
   alpha=0.5
) +
labs(x = "Drift time (index)", "Retention time (index)", title = "Cluster01")
}

}
