% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{gcims_missing_imputation}
\alias{gcims_missing_imputation}
\title{Missing Values Imputation}
\usage{
gcims_missing_imputation(dir_in, peak_table_with_na, peak_list_foms)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from. It should be the one with the baseline removed.}

\item{peak_table_with_na}{Peak table that contains missing values}

\item{peak_list_foms}{Peak list containing all the figures of merit.}
}
\value{
A Set of S3 objects.
}
\description{
Missing Values Imputation
}
\details{
\code{gcims_missing_imputation} calculates the volume of the ROIs that
have missing values. It uses the coordinates of the reference ROI for each
cluster and subistitue the missing value by the volume in this region without
baseline.
}
\examples{
peak_table_imputed <- gcims_missing_imputation(dir_in = bslnr,
peak_table_with_na, peak_list_foms)
head(peak_table_imputed)
}
\seealso{
Other Imputation functions: 
\code{\link{gcims_peak_imputation}()}
}
\concept{Imputation functions}
