% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW_2.R
\name{gcims_optimize_polynomial}
\alias{gcims_optimize_polynomial}
\title{This function finds the optimal polynomial degree to apply ptw to GMIMS data
when correcting retention time axis. The figure of merit is the mean of the
correlation between the reference RIC and samples RIC.}
\usage{
gcims_optimize_polynomial(rics, ref_ric_sample_idx)
}
\arguments{
\item{rics}{Matrix. One Reactant Ion Chromatogram per row.}

\item{ref_ric_sample_idx}{A scalar. It indicates the sample that is chosen
as reference for aligning in retention time.}
}
\value{
A vector of integers with as many components as samples to be corrected, with values
between 0 and 5. If zero, no correction
in retention time is needed. Between 1 and 5, it indicates
the polynomial degree of the warping function.
}
\description{
This function finds the optimal polynomial degree to apply ptw to GMIMS data
when correcting retention time axis. The figure of merit is the mean of the
correlation between the reference RIC and samples RIC.
}
