% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_functions.R
\name{gcims_peak_grouping}
\alias{gcims_peak_grouping}
\alias{"mahalanobis"}
\alias{or}
\alias{"overlapping"}
\title{Two dimensional peak picking}
\usage{
gcims_peak_grouping(dir_in, dir_out, samples, distance = "overlapping")
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{Samples to which perform the peak picking.}

\item{distance}{A measure of dissimilarity of Samples. Either}
}
\value{
A peak table per sample.
}
\description{
Two dimensional peak picking
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Peak Grouping functions}
