% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_picking_functions.R
\name{gcims_peak_picking}
\alias{gcims_peak_picking}
\title{Two dimensional peak picking}
\usage{
gcims_peak_picking(
  dir_in,
  dir_out,
  samples,
  rem_baseline = TRUE,
  min_length_tr = 50,
  min_length_td = 10,
  preprocess = TRUE
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{Samples to which perform the peak picking.}

\item{rem_baseline}{Perform a baseline removal in retention / drift time
axes. TRUE by default.}

\item{min_length_tr}{Minimum peak length in retention time (now indexes).}

\item{min_length_td}{Minimum peak length in drift time (now indexes).}

\item{preprocess}{Preprocess the samples?. TRUE or FALSE}
}
\value{
A peak table per sample.
}
\description{
Two dimensional peak picking
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Peak Peaking functions}
