% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{gcims_peaks_clustering}
\alias{gcims_peaks_clustering}
\title{ROIs Clustering}
\usage{
gcims_peaks_clustering(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where peak table data file is
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples that are going to be included in the peak table.}
}
\value{
A Set of S3 objects.
}
\description{
ROIs Clustering
}
\details{
\code{gcims_peaks_clustering}  does a clustering along the ROIs
for a peak table creation. The Figures of merits of each ROI are also
reported. In this table are included all samples in \code{samples}. Use this
function if you are interested in obtaining a final peak table for future
classification techniques.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 1:3

# Example of ROIs Clustering for Peak Table Creation
gcims_peaks_clustering <- function(dir_in, dir_out, samples)
}
\concept{Peaks CLustering function}
