% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_chrom}
\alias{gcims_plot_chrom}
\title{Plots TICs or EICs.}
\usage{
gcims_plot_chrom(
  dir_in,
  samples,
  dt_value = NULL,
  rt_range = NULL,
  colorby = "Name"
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{dt_value}{Selected drift time value. NULL by default.}

\item{rt_range}{Min a Max retention time values. NULL by
default.}

\item{colorby}{Metadata parameter used for color the samples.}
}
\value{
A set of Chromatograms.
}
\description{
Plots TICs or EICs.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
samples <- c(3, 7 ,8, 14, 20, 21)
gcims_plot_chrom(dir_in, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")
setwd(current_dir)

}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_spec}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
