% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_eics}
\alias{gcims_plot_eics}
\title{Plot Sample Extracted Ion Chromatrogram}
\usage{
gcims_plot_eics(dir_in, samples, td_ind)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The number corresponding to the sample
to be visualized.}

\item{td_ind}{Drift time Index (to be modified in future)}
}
\value{
An extracted ion chromatogram

An Extracted Ion Chromatogram,
}
\description{
Plot Sample Extracted Ion Chromatrogram
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_spec}()},
\code{\link{gcims_plot_tics}()},
\code{\link{gcims_plot_tisa}()},
\code{\link{gcims_visualization}()}
}
\concept{Visualization functions}
