% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_spec}
\alias{gcims_plot_spec}
\title{Plot Sample Spectra}
\usage{
gcims_plot_spec(dir_in, samples, tr_ind)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{tr_ind}{Retention time Index (to be modified in future)}
}
\value{
A plot of the Sample Spectra.
}
\description{
Plot Sample Spectra
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_eics}()},
\code{\link{gcims_plot_tics}()},
\code{\link{gcims_plot_tisa}()},
\code{\link{gcims_visualization}()}
}
\concept{Visualization functions}
