% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_tics}
\alias{gcims_plot_tics}
\title{Total Ion Chromatogram visualization}
\usage{
gcims_plot_tics(dir_in, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}
}
\value{
A plot of the Total Ion Chromatograms.
}
\description{
Total Ion Chromatogram visualization
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_tisa}()},
\code{\link{gcims_visualization}()}
}
\concept{Visualization functions}
