% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{gcims_prepare_data}
\alias{gcims_prepare_data}
\title{Data preparation for individual samples.}
\usage{
gcims_prepare_data(dir_in, dir_out, samples, params)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where smoothed data files are
stored.}

\item{samples}{A vector. Set of samples to which remove the baseline
(e.g.: c(1, 2, 3)).}

\item{params}{A list of lists with parameter values to perform
filtering and decimation in retention and drift time axes.}
}
\value{
A set of S3 objects.Additionally, it returns a list with two variables needed to perform sample alignment in drift
and retention time axes: \code{tis} and \code{rics}. First variable (a matrix) corresponds to the Total Ion Spectra of
samples, while second one (a matrix) to their Reactive Ion Chromatograms, respectively.
}
\description{
Data preparation for individual samples.
}
\details{
Description of the parameters found in the variable \code{params}.
\itemize{
\item{\code{params$filter$do_filter}}{ is a Boolean variable. If TRUE a Savitzky-Golay filters are applied.}
\item{\code{params$filter$order_rt}}{ is a positive integer. Polynomial order in retention time.}
\item{\code{params$filter$length_rt}}{ is a positive integer. Filter length in retention time.}
\item{\code{params$filter$order_dt}}{ is a positive integer. Polynomial order in drift time.}
\item{\code{params$filter$length_dt}}{ is a positive integer. Filter length in drift time.}
\item{\code{params$decimate$do_decimate}}{ is a Boolean variable. If TRUE decimation is performed on data.}
\item{\code{params$decimate$dfactor_rt}}{ is a positive integer. Decimation factor in retention time.}
\item{\code{params$decimate$factor_dt}}{ is a positive integer. Decimation factor in drift time.}
}

\code{gcims_prepare_data} prepares samples for the subsequent data
pre-processing stage.
Each of the samples is interpolated, filtered (optional),
and decimated (optional).
Note that filter length must be an odd number bigger than the
polynomial order of the filter.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of digital smoothing and decimation, in both axes:
# Before:
gcims_plot_chrom(dir_in, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

# Set parameters:
params <-list()

# Filtering
params$filter$do <- TRUE
params$filter$order_rt <- 2
params$filter$length_rt <- 19
params$filter$order_dt <- 2
params$filter$length_dt <- 19

# Decimation
params$decimate$do <- TRUE
params$decimate$factor_rt <- 2
params$decimate$factor_dt <- 2

# After:
gcims_prepare_data(dir_in, dir_out, samples, params)
gcims_plot_chrom(dir_out, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\references{
{ García, S., Luengo, J. and Herrera, F., 2015. Data preprocessing
in data mining (Vol. 72, pp. 59-139). Cham, Switzerland: Springer International
Publishing.
\doi{10.1007/978-3-319-10247-4}
}
}
\concept{Data preparation functions}
