% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_mat}
\alias{gcims_read_mat}
\title{Read GC-IMS Samples from MATLAB files}
\usage{
gcims_read_mat(dir_in, dir_out)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}
}
\value{
An invisible NULL. This function is called to convert mat files to rds files.
}
\description{
Read GC-IMS Samples from MATLAB files
}
\examples{
\dontrun{
gcims_read_mat("/path/to/dir/with/mat/files", "where_to_save_samples_in_rds/")
}
}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_metadata}()}
}
\concept{Reading functions}
