% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_mea}
\alias{gcims_read_mea}
\title{Read GC-IMS samples from .mea or .mea.gz files and save them as RDS list objects}
\usage{
gcims_read_mea(dir_in, dir_out)
}
\arguments{
\item{dir_in}{Input directory. Where the raw data files are stored.}

\item{dir_out}{Output directory. Where data files are stored as R
objects.}
}
\value{
The filenames that have been created in \code{dir_out}
}
\description{
Read GC-IMS samples from .mea or .mea.gz files and save them as RDS list objects
}
\details{
\code{gcims_read_samples} stores a set of S3 objects in the
directory \code{dir_out} (one per sample sample stored in \code{dir_in}).
Each object is a list containing the following variables: \code{metadata}
and \code{data}.In \code{metadata} you can find information related to the
sample that can be added by the user. On the other hand in \code{data} you
can find information of the measurement.By default, when the
\code{metadata} field is created only includes one variable, called
\code{Name}, although the user can include more fields using the function
\code{\link[=gcims_read_metadata]{gcims_read_metadata()}}. Regarding \code{data}, it is consist in
three fields:  \describe{
\item{\code{retention_time}}{ The vector of retention times associated to a sample measurement.}
\item{\code{drift_time}}{ The vector of drift times associated to a sample
measurement.}
\item{\code{data_df}}{ A dataframe containing the intensities
of a GCIMS measurement of a sample. Each column in the dataset corresponds
to a particular drift time, while each row to a different retention time.}
}
}
\examples{
dir_in <- system.file("extdata", "sample_formats", package = "GCIMS")
dir_out <- tempfile("dir_out")
dir.create(dir_out, recursive = TRUE, showWarnings = FALSE)
gcims_read_mea(dir_in, dir_out)
list.files(dir_out)

}
\seealso{
Other Reading functions.: 
\code{\link{gcims_read_samples}()}
}
\concept{Reading functions.}
