% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_metadata}
\alias{gcims_read_metadata}
\title{Read metadata form a csv}
\usage{
gcims_read_metadata(dir_in, samples, file)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be processed.}

\item{file}{Name of the file that contains the metadata. It must be an excel file}
}
\value{
An R object that contains the metadata information
}
\description{
Read metadata form a csv
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_mat}()},
\code{\link{gcims_read_samples}()}
}
\concept{Reading functions}
