% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_samples}
\alias{gcims_read_samples}
\title{Read GC-IMS Samples from .csv files}
\usage{
gcims_read_samples(dir_in, dir_out, sftwr)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{sftwr}{Use 1 if the software that has converted the original files into
csv format is the new one. Use 2 if the software is the old one.}
}
\value{
An R object that contains the matrix and the retention times
}
\description{
Read GC-IMS Samples from .csv files
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
gcims_read_samples(dir_in, dir_out, sftwr = 2)
files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
print(files)
file.remove(files)
setwd(current_dir)

}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_mat}()},
\code{\link{gcims_read_metadata}()}
}
\concept{Reading functions}
