% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions_NEW.R
\name{gcims_remove_baseline}
\alias{gcims_remove_baseline}
\title{Remove baseline of a dataset using psalsa}
\usage{
gcims_remove_baseline(dir_in, dir_out, samples, peak_list_clustered)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are
loaded from.}

\item{dir_out}{Output directory. Where data files after baseline
correction are stored.}

\item{samples}{A vector. Set of samples to which remove the baseline
(e.g.: c(1, 2, 3)).}

\item{peak_list_clustered}{A peak list after grouping by cluster.}
}
\value{
A set of S3 objects.
}
\description{
Remove baseline of a dataset using psalsa
}
\details{
\code{gcims_remove_baseline} performs baseline correction on sample
spectra (in drift time) and chromatograms (in retention time) .This function corrects
the baseline in drift time and afterwards, in retention time. To do it, it cuts the
axis in regions of equal length and looks for the local minima of each region. A simple
linear interpolation is performed to create a baseline of the same length as the number
of data points per axis.The region width is estimated differently in each of the the axes:
In drift time, it is computed as 12 times the standard deviation of the RIP width, while
in retention time as the third quantile of peak width distributions in chromatograms (obtained
from the clustered peak table). Take into account that when removing the baseline in retention
time, the shape of the Reactant Ion Peak is severely affected.
}
\seealso{
Other Baseline removal functions: 
\code{\link{gcims_baseline_removal}()}
}
\concept{Baseline removal functions}
