% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_remove_rip}
\alias{gcims_remove_rip}
\title{Removes RIP from all samples}
\usage{
gcims_remove_rip(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}
}
\value{
A RIP removed dataset
}
\description{
Removes RIP from all samples
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Drift time Interpolation
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
gcims_remove_rip(dir_in, dir_out, samples)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
