% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_rois_fusion.R
\name{gcims_rois_fusion}
\alias{gcims_rois_fusion}
\title{ROIs Fusion}
\usage{
gcims_rois_fusion(clustering_object)
}
\arguments{
\item{clustering_object}{Objects obtained from the clustering function
containing a list of all the ROIs with tehir corresponding cluster.}
}
\value{
A Set of S3 objects.
}
\description{
ROIs Fusion
}
\details{
\code{clustering_object} calculates de median coordinates of all the ROIs
of each cluster. All the ROis inside a cluster are resized according to their
geometrical center and the size of the median ROI.
}
\examples{
# Example of ROIs FUsion
tt <- gcims_rois_fusion(peak_clustering)
samplestocheck <- tt$peak_list_clustered[which(tt$peak_list_clustered$cluster == 35), ]
ggplot() +
 geom_rect(data = samplestocheck, mapping = aes(xmin=dt_min_idx, xmax=dt_max_idx,
 ymin=rt_min_idx, ymax=rt_max_idx), color= as.factor(samplestocheck$SampleID), alpha=0.5)
invisible(file.remove(files))
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_figures_of_merit}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
