% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakDetection.R
\name{gcims_rois_selection}
\alias{gcims_rois_selection}
\title{ROIs Selection}
\usage{
gcims_rois_selection(dir_in, dir_out, samples, noise_level)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where ouput data files are stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which peaks must be detected.}

\item{noise_level}{Scalar number. The number of times the standard deviation
above the noise level needed to detect a peak. IUPAC recommends \code{noise_level = 3} for detection.}
}
\value{
A set of S3 objects. Additionally, a peak/roi list.
}
\description{
ROIs Selection
}
\examples{
\donttest{
# Use BiocParallel library for parallelization
library(BiocParallel)
register(SnowParam(workers = 3, progressbar = TRUE, exportglobals = FALSE), default = TRUE)

dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3,7)

# Example of Peak Picking
peak_list <- gcims_rois_selection(dir_in, dir_out, samples , noise_level = 3)
head(peak_list)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
}

}
\concept{Peak detection functions}
