% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_shift_rt}
\alias{gcims_shift_rt}
\title{Shift samples in retention time.}
\usage{
gcims_shift_rt(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where decimated data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to be shifted in retention time.}
}
\value{
A set of S3 objets.
}
\description{
Perform a shift in retention time so to the injection point of all samples
coincide.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()

# Before:
samples <- c(3, 7)
gcims_plot_chrom(dir_in, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")

# After:
gcims_shift_rt(dir_in, dir_out, samples)
gcims_plot_chrom(dir_out, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\seealso{
Other Utility functions: 
\code{\link{gcims_compute_fom}()},
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()}
}
\concept{Utility functions}
