% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_shifting}
\alias{gcims_shifting}
\title{Shifting}
\usage{
gcims_shifting(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where decimated data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to be decimated.}
}
\value{
A set of S3 objets.
}
\description{
Shifting
}
\details{
\code{gcims_shifting} performs peak shifting in retention time axes
of GCIMS data. In particular, it shifts all the peaks along the retention
time axis in order to have the RIP peaks at the exact same time.
}
\note{
\code{gcims_shifting} reduce the retention time axis of some of
the samples.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()

# Before:
samples <- c(3, 7)
gcims_plot_chrom(dir_in, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")
gcims_plot_spec(dir_in, samples, rt_value = NULL,  dt_range = NULL, colorby = "Class")

# After:
gcims_shifting(dir_in, dir_out, samples)
# c(50, 226)
gcims_plot_chrom(dir_out, samples, dt_value = NULL,  rt_range = c(60, 98) , colorby = "Class")
gcims_plot_spec(dir_out, samples, rt_value = NULL,  dt_range = c(7.8, 9.0), colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\references{
{ Oppenheim, Alan V.; Schafer, Ronald W.; Buck, John R. (1999).
"4". Discrete-Time Signal Processing (2nd ed.). Upper Saddle River, N.J.:
Prentice Hall. p. 168. ISBN 0-13-754920-2. }
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_figures_of_merit}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
