% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{gcims_smoothing}
\alias{gcims_smoothing}
\title{Data Smoothing using Savitzky-Golay filters}
\usage{
gcims_smoothing(
  dir_in,
  dir_out,
  samples,
  time,
  filter_length = 19,
  polynomial_order = 2
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}

\item{filter_length}{Numerical. Length of the filter.}

\item{polynomial_order}{Numerical. Order of the polynomial.}
}
\value{
A filtered  gcims dataset.
}
\description{
Data Smoothing using Savitzky-Golay filters
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Retention time smoothing:
# Before:
gcims_plot_chrom(dir_in, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

# After:
gcims_smoothing(dir_in, dir_out, samples, time = "Retention")
gcims_plot_chrom(dir_out, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\concept{Smoothing functions}
