% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_unfold}
\alias{gcims_unfold}
\title{Converts Sample Matrices to Feature Vectors}
\usage{
gcims_unfold(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}
}
\value{
A set of feature vectors
}
\description{
Converts Sample Matrices to Feature Vectors
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", "to_interpolate", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Sample Data unfolding
# Before:
setwd(dir_in)
M3 <- readRDS("M3.rds")
data <- M3$data$data_df
data_dimensions <- dim(data)

message("Before unfolding data has ", data_dimensions[1],
" rows and ", data_dimensions[2], " columns.")

# After:
gcims_unfold(dir_in, dir_out, samples)
setwd(dir_out)
M3 <- readRDS("M3.rds")
data <- M3$data$data_df
data_length <- length(data)

message("After unfolding data is a vector of length ", data_length,".
This value corresponds to product of the number
of columns by the number of rows of the original data.")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)


}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()}
}
\concept{Utility functions}
