% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_view_sample}
\alias{gcims_view_sample}
\title{Sample matrix visualization}
\usage{
gcims_view_sample(dir_in, sample_num, rt_range = NULL, dt_range = NULL)
}
\arguments{
\item{dir_in}{The input directory.}

\item{sample_num}{The number corresponding to the sample
to be visualized.}

\item{rt_range}{Min a Max retention time values. NULL by
default.}

\item{dt_range}{Min a Max drift time values. NULL by
default.}
}
\value{
An image of the sample matrix.
}
\description{
Sample matrix visualization
}
\examples{
wd <- getwd()
dir_in <- tempdir()
dir_out <-  file.path(dir_in,"dummy")
dir.create (dir_out, FALSE)
samples_per_class <- 1
params <- gcims_create_dummy_params()
gcims_create_dummy_set(dir_in, dir_out, samples_per_class, params)
gcims_view_sample(dir_out, 1)
unlink(dir_out, recursive = TRUE)
setwd(wd)
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_plot_spec}()}
}
\concept{Visualization functions}
