% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_visualization}
\alias{gcims_visualization}
\title{SAmples Visualization}
\usage{
gcims_visualization(dir_in, sample_num)
}
\arguments{
\item{dir_in}{The input directory.}

\item{sample_num}{The number corresponding to the sample
to be visualized.}
}
\value{
A set of plots of the samples: full espectra, Intensity VS Retention Time
and Intensity VS Drift Time.
}
\description{
SAmples Visualization
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Visualization functions}
