% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSChromatogram-class.R,
%   R/GCIMSChromatogram-dtime-rtime-intensity.R
\docType{class}
\name{GCIMSChromatogram-class}
\alias{GCIMSChromatogram-class}
\alias{GCIMSChromatogram}
\alias{dtime,GCIMSChromatogram-method}
\alias{rtime,GCIMSChromatogram-method}
\alias{intensity,GCIMSChromatogram-method}
\title{GCIMSChromatogram class}
\usage{
GCIMSChromatogram(...)

\S4method{dtime}{GCIMSChromatogram}(object)

\S4method{rtime}{GCIMSChromatogram}(object)

\S4method{intensity}{GCIMSChromatogram}(object, rt_range = NULL, rt_idx = NULL)
}
\arguments{
\item{...}{See the slots section in this page}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}
}
\value{
A GCIMSChromatogram object
}
\description{
GCIMSChromatogram is an S4 class to store a GCIMS Chromatogram It can be a single chromatogram
or the aggregation of several chromatograms.
}
\section{Functions}{
\itemize{
\item \code{GCIMSChromatogram()}: class

\item \code{dtime(GCIMSChromatogram)}: Get the drift time vector

\item \code{rtime(GCIMSChromatogram)}: Get the retention time where this spectrum was extracted

\item \code{intensity(GCIMSChromatogram)}: Get the intensity matrix

}}
\section{Slots}{

\describe{
\item{\code{retention_time}}{A numeric vector with retention times}

\item{\code{intensity}}{A numeric vector with the corresponding intensities}

\item{\code{drift_time_idx}}{The index or indices used to get the intensity}

\item{\code{drift_time_s}}{The drift times corresponding to the retention time indices.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}
}}

