% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSChromatogram.R
\name{GCIMSChromatogram}
\alias{GCIMSChromatogram}
\title{Create a \linkS4class{GCIMSChromatogram} object}
\usage{
GCIMSChromatogram(
  retention_time,
  retention_index = NULL,
  intensity,
  drift_time_idx = NA_integer_,
  drift_time_ms = NA_real_,
  description = "",
  baseline = NULL,
  peaks = NULL,
  peaks_debug_info = NULL
)
}
\arguments{
\item{retention_time}{A numeric vector with retention times}

\item{retention_index}{A numeric vector with retention indexes}

\item{intensity}{A numeric vector with the corresponding intensities}

\item{drift_time_idx}{The index or indices used to get the intensity}

\item{drift_time_ms}{The drift times corresponding to \code{drift_time_idx}.}

\item{description}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}

\item{baseline}{A numeric vector of the same length as \code{intensity} with the corresponding baseline,
or \code{NULL} if not set. Use \code{\link[=estimateBaseline]{estimateBaseline()}} to estimate it, \code{\link[=baseline]{baseline()}} to directly access it.}

\item{peaks}{A data frame with peaks, use \code{\link[=findPeaks]{findPeaks()}} to fill it, or \code{\link[=peaks]{peaks()}} to set/get it}

\item{peaks_debug_info}{A list with arbitrary debug information from \code{\link[=findPeaks]{findPeaks()}}}
}
\value{
A \linkS4class{GCIMSChromatogram} object
}
\description{
Create a \linkS4class{GCIMSChromatogram} object
}
\examples{
GCIMSChromatogram(
  retention_time = seq(from = 0, to = 10, length.out = 200),
  intensity = 1:200
)
}
\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram-class}},
\code{\link{dtime,GCIMSChromatogram-method}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{findPeaks,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{rtime,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
