% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSChromatogramSet.R,
%   R/plot-GCIMSChromatogramSet.R
\docType{class}
\name{GCIMSChromatogramSet-class}
\alias{GCIMSChromatogramSet-class}
\alias{sampleNames,GCIMSChromatogramSet-method}
\alias{sampleNames<-,GCIMSChromatogramSet,ANY-method}
\alias{pData,GCIMSChromatogramSet-method}
\alias{length,GCIMSChromatogramSet-method}
\alias{[[,GCIMSChromatogramSet-method}
\alias{plot,GCIMSChromatogramSet,ANY-method}
\title{GCIMSChromatogramSet class}
\usage{
\S4method{sampleNames}{GCIMSChromatogramSet}(object)

\S4method{sampleNames}{GCIMSChromatogramSet,ANY}(object) <- value

\S4method{pData}{GCIMSChromatogramSet}(object)

\S4method{length}{GCIMSChromatogramSet}(x)

\S4method{[[}{GCIMSChromatogramSet}(x, i)

\S4method{plot}{GCIMSChromatogramSet,ANY}(x, color_by = "SampleID", ...)
}
\arguments{
\item{object}{A \link{GCIMSChromatogramSet} object}

\item{value}{A character vector of length the number of chromatograms with the new sample names}

\item{x}{A \link{GCIMSChromatogramSet} object to plot}

\item{i}{A number or a string with the sample index or name}

\item{color_by}{The name of a \code{pData(x)} column (or \code{"SampleID"}) used to
color the chromatograms}

\item{y}{From the generic \code{plot} function, ignored for \code{GCIMSChromatogramSet} class objects}
}
\value{
A character vector with the sample names

The \link{GCIMSChromatogramSet} object, with samples renamed in both
\code{chromatograms} and \code{pData()}

A tibble with the phenotype data, or \code{NULL} if not set

An integer with the number of chromatograms

The \link{GCIMSChromatogram} of the requested sample

A ggplot2 plot object
}
\description{
GCIMSChromatogramSet is an S4 class to store one \link{GCIMSChromatogram} per
sample of a \link{GCIMSDataset}, together with a copy of \code{pData()} so plots can
use the dataset's annotations.

Samples are not required to share a common retention time axis: each
chromatogram keeps its own, exactly as extracted from its sample. No
interpolation is performed.
}
\section{Functions}{
\itemize{
\item \code{sampleNames(GCIMSChromatogramSet)}: Get the sample names

\item \code{sampleNames(object = GCIMSChromatogramSet) <- value}: Set the sample names

\item \code{pData(GCIMSChromatogramSet)}: Get the phenotype data

\item \code{length(GCIMSChromatogramSet)}: Number of chromatograms (samples) in the set

\item \code{[[}: Extract the chromatogram of a single sample

\item \code{plot(x = GCIMSChromatogramSet, y = ANY)}: plot method

}}
\section{Slots}{

\describe{
\item{\code{chromatograms}}{A named list of \link{GCIMSChromatogram} objects, one per
sample, named after their \code{SampleID}.}

\item{\code{pData}}{A \code{DataFrame} with the phenotype data, or \code{NULL}.}
}}

\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram}},
\code{\link{GCIMSChromatogram-class}},
\code{\link{GCIMSChromatogramSet}},
\code{\link{dtime,GCIMSChromatogram-method}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{findPeaks,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{rtime,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
