% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSChromatogramSet.R
\name{GCIMSChromatogramSet}
\alias{GCIMSChromatogramSet}
\title{Create a \linkS4class{GCIMSChromatogramSet} object}
\usage{
GCIMSChromatogramSet(chromatograms = list(), pData = NULL)
}
\arguments{
\item{chromatograms}{A named list of \link{GCIMSChromatogram} objects, one per
sample, named after their \code{SampleID}.}

\item{pData}{A \code{data.frame}/\code{DataFrame}/tibble with the phenotype data, or \code{NULL}.}
}
\value{
A \linkS4class{GCIMSChromatogramSet} object
}
\description{
Create a \linkS4class{GCIMSChromatogramSet} object
}
\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram}},
\code{\link{GCIMSChromatogram-class}},
\code{\link{GCIMSChromatogramSet-class}},
\code{\link{dtime,GCIMSChromatogram-method}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{findPeaks,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{rtime,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
