% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSDataset.R,
%   R/getTIS_getRIC-GCIMSDataset.R
\docType{class}
\name{GCIMSDataset-class}
\alias{GCIMSDataset-class}
\alias{GCIMSDataset}
\title{GCIMSDataset}
\arguments{
\item{sample}{Either an integer (sample index) or a string (sample name)}
}
\value{
The GCIMSSample object

A matrix with the reverse ion chromatograms for all samples
}
\description{
GCIMSDataset is an S4 reference class to store a dataset

When the dataset is created, the \code{on_ram} option controls whether the actual
data is stored not in memory or it is read/saved from/to files as
needed, so the dataset object scales with large number of samples.

Get a sample from a GCIMSDataset
}
\section{Fields}{

\describe{
\item{\code{pData}}{A data frame with at least the SampleID and filename columns.}

\item{\code{scratch_dir}}{A directory to save intermediate results.}

\item{\code{delayed_ops}}{Delayed operations}

\item{\code{TIS}}{A matrix of n_samples vs drift time, with the Total Ion Spectrum of each sample}

\item{\code{RIC}}{A matrix of n_samples vs retention time, with the Reverse Ion Chromatogram of each sample}

\item{\code{dt_ref}}{A numeric drift time of reference}

\item{\code{rt_ref}}{A numeric retention time of reference}
}}


\examples{
dataset <- GCIMSDataset(
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
}
