% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSDataset.R
\name{GCIMSDataset-old-constructor}
\alias{GCIMSDataset-old-constructor}
\title{GCIMSDataset old constructor}
\arguments{
\item{pData}{A data frame with at least the \code{SampleID} and \code{FileName} columns.}

\item{base_dir}{A directory containing the file names described in \code{pData}}

\item{parser}{Either a string \code{"default"} or a function that takes a filename and returns a \linkS4class{GCIMSSample} object}

\item{scratch_dir}{A directory to save intermediate results.}

\item{keep_intermediate}{A logical. Whether to keep or not intermediate files when realizing the GCIMSDataset object}

\item{on_ram}{A logical. If \code{TRUE}, samples are kept on RAM. This is faster
as long as you have enough memory to keep all samples. If this is \code{TRUE}, then
\code{scratch_dir} and \code{keep_intermediate} are ignored.}
}
\value{
A GCIMSDataset object
}
\description{
GCIMSDataset old constructor
}
\examples{
# Create a new GCIMSDataset with the convenient constructor function:
dummy_obj <-GCIMSDataset(
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
}
