% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSDataset.R
\name{GCIMSDataset_fromList}
\alias{GCIMSDataset_fromList}
\title{GCIMSDataset_fromList}
\usage{
GCIMSDataset_fromList(
  samples,
  pData = NULL,
  scratch_dir = NULL,
  keep_intermediate = FALSE,
  on_ram = TRUE
)
}
\arguments{
\item{samples}{A named list of \link{GCIMSSample} objects. names should match \code{pData$SampleID}}

\item{pData}{A data frame with at least the SampleID and filename columns.}

\item{scratch_dir}{A directory to save intermediate results.}

\item{keep_intermediate}{Whether to keep sample files for intermediate results. Only used if \code{on_ram=FALSE}}

\item{on_ram}{logical. Whether the dataset should be kept stored on RAM or on disk.}
}
\value{
A GCIMSDataset object
}
\description{
GCIMSDataset_fromList
}
\examples{
# Create a new GCIMSDataset with the convenient constructor function:
sample1 <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
dummy_obj <- GCIMSDataset_fromList(
  pData = data.frame(SampleID = "Sample1", Sex = "female"),
  samples = list(Sample1 = sample1)
)
}
