% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSDelayedOp.R
\docType{class}
\name{GCIMSDelayedOp-class}
\alias{GCIMSDelayedOp-class}
\title{GCIMSDelayedOp class}
\description{
GCIMSDelayedOp is an S4 class to store a delayed operation

Delayed operations are not applied to the dataset immediately, but rather
when some data from the dataset is required. Delaying the calculation has
the advantage that some operations that are applied to each sample can be applied
sequentially to each sample, reducing the amount of reads and writes to
the data files, and increasing the performance.
}
\section{Slots}{

\describe{
\item{\code{name}}{A named for de delayed operation, only used for printing.}

\item{\code{fun}}{A function that takes a \link{GCIMSSample} and returns a \link{GCIMSSample} (modified)}

\item{\code{params}}{A named list with additional arguments to be passed to function}

\item{\code{params_iter}}{A named list with additinoal arguments to be passed to
function. Compared to \code{params}, each argument must be a named list of length the number of samples, so
each sample will receive its corresponding parameter according to its name}

\item{\code{fun_extract}}{A function that takes a modified \link{GCIMSSample} and returns an extracted object.}

\item{\code{fun_aggregate}}{A function that takes a \link{GCIMSDataset} and a list of extracted objects and returns a modified \link{GCIMSDataset}.}
}}

