% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx-deprecated-helpers.R
\name{GCIMSDelayedOp}
\alias{GCIMSDelayedOp}
\title{Deprecated constructor}
\usage{
GCIMSDelayedOp(
  name,
  fun = NULL,
  params = list(),
  params_iter = list(),
  fun_extract = NULL,
  fun_aggregate = NULL
)
}
\arguments{
\item{name}{A named for de delayed operation, only used for printing.}

\item{fun}{A function that takes a sample and returns a modified sample}

\item{params}{A named list with additional arguments to be passed to function}

\item{params_iter}{A named list with additional arguments to be passed to
function. Compared to \code{params}, each argument must be a named list of length the number of samples, so
each sample will receive its corresponding parameter according to its name}

\item{fun_extract}{A function that takes a modified sample and returns an extracted object.}

\item{fun_aggregate}{A function that takes a dataset and a list of extracted objects and returns a modified dataset.}
}
\description{
Deprecated constructor
}
\keyword{internal}
