% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{GCIMSSample-class}
\alias{GCIMSSample-class}
\title{GCIMSSample class}
\value{
A GCIMSSample object
}
\description{
GCIMS Sample is an S4 class to store one sample
with the drift and retention time ranges and other relevant attributes
(GC column, drift tube length...) if available
}
\details{
The actual spectra is stored in the `data` slot, in a matrix,
where the first index (rows) corresponds to drift times and the
second to retention times (columns).
}
\section{Slots}{

\describe{
\item{\code{drift_time}}{numeric. (required)}

\item{\code{retention_time}}{numeric. (required)}

\item{\code{data}}{matrix A matrix with drift time in the rows and retention time
in columns. (required)}
}}

