% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GCIMSSample-methods}
\alias{GCIMSSample-methods}
\alias{[.GCIMS}
\alias{subset.GCIMSSample}
\alias{subset}
\alias{[[<-,GCIMSSample,ANY,ANY,ANY-method}
\title{Methods for the GCIMSSample class}
\usage{
\method{[}{GCIMS}(x, i, j, ...)

\method{subset}{GCIMSSample}(x, dt_idx = NULL, rt_idx = NULL, ...)

\S4method{[[}{GCIMSSample,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{i}{index for drift time to subset}

\item{j}{index for retention time to subset}

\item{...}{ignored}

\item{dt_idx}{A vector of drift time indices to keep}

\item{rt_idx}{A vector of retention time indices to keep}

\item{value}{A vector of length number of samples with metadata to add; \strong{note}:
can pass \code{NULL} to remove metadata or an associated object}
}
\value{
\code{[}: object \code{x} with features \code{i} and cells \code{j}

\code{subset}: A subsetted \code{GCIMSSample} object

\code{[[<-}: \code{x} with the metadata or associated objects added
as \code{i}; if \code{value} is \code{NULL}, removes metadata \code{i}
from object \code{x}
}
\description{
Methods for the GCIMSSample class
}
\section{Functions}{
\itemize{
\item \code{[.GCIMS}: Simple subsetter for \code{GCIMSSample} objects

\item \code{subset.GCIMSSample}: Subset a \code{\link{GCIMSSample-class}} object

\item \code{[[<-,GCIMSSample,ANY,ANY,ANY-method}: Add metadata
}}

\examples{
# `[' examples

}
\seealso{
\code{\link[base]{subset}}
}
