% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-intensity-GCIMSSample.R
\name{GCIMSSample-rtime-dtime-intensity}
\alias{GCIMSSample-rtime-dtime-intensity}
\alias{dtime,GCIMSSample-method}
\alias{rtime,GCIMSSample-method}
\alias{intensity,GCIMSSample-method}
\alias{intensity<-,GCIMSSample-method}
\title{Drift time, Retention time, Intensity of GCIMSSamples}
\usage{
\S4method{dtime}{GCIMSSample}(object)

\S4method{rtime}{GCIMSSample}(object)

\S4method{intensity}{GCIMSSample}(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL
)

\S4method{intensity}{GCIMSSample}(object) <- value
}
\arguments{
\item{object}{A GCIMSSample object.}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}

\item{value}{A matrix of dimensions \code{dim(object)}}
}
\value{
The drift time of the sample

The retention time of the sample
}
\description{
Functions to extract the drift time, the retention time and
the intensity.
}
\section{Functions}{
\itemize{
\item \code{dtime(GCIMSSample)}: Get the drift time vector

\item \code{rtime(GCIMSSample)}: Get the retention time vector

\item \code{intensity(GCIMSSample)}: Get the intensity matrix

\item \code{intensity(GCIMSSample) <- value}: Set the intensity matrix

}}
\examples{
mea_file <- system.file("extdata",  "sample_formats", "small.mea.gz", package = "GCIMS")
gcims_sample <- read_mea(mea_file)
my_matrix <- intensity(gcims_sample, dt_range = c(7, 8), rt_range = c(1,30))
mea_file <- system.file("extdata",  "sample_formats", "small.mea.gz", package = "GCIMS")
gcims_sample <- read_mea(mea_file)
my_matrix <- intensity(gcims_sample)
intensity(gcims_sample) <- my_matrix/100
}
