% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSSample.R
\name{GCIMSSample}
\alias{GCIMSSample}
\title{Create a \linkS4class{GCIMSSample} object}
\usage{
GCIMSSample(drift_time, retention_time, data, ...)
}
\arguments{
\item{drift_time, retention_time, data, ...}{See the Slots section in \linkS4class{GCIMSSample} page}
}
\value{
A GCIMSSample object
}
\description{
Create a \linkS4class{GCIMSSample} object
}
\examples{
# Create a new GCIMSSample with the convenient constructor function:
dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
}
