% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSpectrum-class.R,
%   R/GCIMSSpectrum-dtime-rtime-intensity.R
\docType{class}
\name{GCIMSSpectrum-class}
\alias{GCIMSSpectrum-class}
\alias{GCIMSSpectrum}
\alias{dtime,GCIMSSpectrum-method}
\alias{rtime,GCIMSSpectrum-method}
\alias{intensity,GCIMSSpectrum-method}
\title{GCIMSSpectrum class}
\usage{
GCIMSSpectrum(...)

\S4method{dtime}{GCIMSSpectrum}(object)

\S4method{rtime}{GCIMSSpectrum}(object)

\S4method{intensity}{GCIMSSpectrum}(object, dt_range = NULL, dt_idx = NULL)
}
\arguments{
\item{...}{See the slots section in this page}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}
}
\value{
A GCIMSSpectrum object
}
\description{
GCIMSSpectrum is an S4 class to store a GCIMS Spectrum. It can be a single spectrum
or the aggregation of several spectra.
}
\section{Functions}{
\itemize{
\item \code{GCIMSSpectrum()}: class

\item \code{dtime(GCIMSSpectrum)}: Get the drift time vector

\item \code{rtime(GCIMSSpectrum)}: Get the retention time where this spectrum was extracted

\item \code{intensity(GCIMSSpectrum)}: Get the intensity matrix

}}
\section{Slots}{

\describe{
\item{\code{drift_time}}{A numeric vector with drift times}

\item{\code{intensity}}{A numeric vector witht the corresponding intensities}

\item{\code{retention_time_idx}}{The index or indices used to get the intensity}

\item{\code{retention_time_s}}{The retention times corresponding to the retention time indices.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}
}}

