% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSSpectrumSet.R,
%   R/plot-GCIMSSpectrumSet.R
\docType{class}
\name{GCIMSSpectrumSet-class}
\alias{GCIMSSpectrumSet-class}
\alias{sampleNames,GCIMSSpectrumSet-method}
\alias{sampleNames<-,GCIMSSpectrumSet,ANY-method}
\alias{pData,GCIMSSpectrumSet-method}
\alias{length,GCIMSSpectrumSet-method}
\alias{[[,GCIMSSpectrumSet-method}
\alias{plot,GCIMSSpectrumSet,ANY-method}
\title{GCIMSSpectrumSet class}
\usage{
\S4method{sampleNames}{GCIMSSpectrumSet}(object)

\S4method{sampleNames}{GCIMSSpectrumSet,ANY}(object) <- value

\S4method{pData}{GCIMSSpectrumSet}(object)

\S4method{length}{GCIMSSpectrumSet}(x)

\S4method{[[}{GCIMSSpectrumSet}(x, i)

\S4method{plot}{GCIMSSpectrumSet,ANY}(x, color_by = "SampleID", ...)
}
\arguments{
\item{object}{A \link{GCIMSSpectrumSet} object}

\item{value}{A character vector of length the number of spectra with the new sample names}

\item{x}{A \link{GCIMSSpectrumSet} object to plot}

\item{i}{A number or a string with the sample index or name}

\item{color_by}{The name of a \code{pData(x)} column (or \code{"SampleID"}) used to
color the spectra}

\item{...}{Ignored}
}
\value{
A character vector with the sample names

The \link{GCIMSSpectrumSet} object, with samples renamed in both
\code{spectra} and \code{pData()}

A tibble with the phenotype data, or \code{NULL} if not set

An integer with the number of spectra

The \link{GCIMSSpectrum} of the requested sample

A ggplot2 plot object
}
\description{
GCIMSSpectrumSet is an S4 class to store one \link{GCIMSSpectrum} per sample of
a \link{GCIMSDataset}, together with a copy of \code{pData()} so plots can use the
dataset's annotations.

Samples are not required to share a common drift time axis: each spectrum
keeps its own, exactly as extracted from its sample. No interpolation is
performed.
}
\section{Functions}{
\itemize{
\item \code{sampleNames(GCIMSSpectrumSet)}: Get the sample names

\item \code{sampleNames(object = GCIMSSpectrumSet) <- value}: Set the sample names

\item \code{pData(GCIMSSpectrumSet)}: Get the phenotype data

\item \code{length(GCIMSSpectrumSet)}: Number of spectra (samples) in the set

\item \code{[[}: Extract the spectrum of a single sample

\item \code{plot(x = GCIMSSpectrumSet, y = ANY)}: plot method

}}
\section{Slots}{

\describe{
\item{\code{spectra}}{A named list of \link{GCIMSSpectrum} objects, one per sample,
named after their \code{SampleID}.}

\item{\code{pData}}{A \code{DataFrame} with the phenotype data, or \code{NULL}.}
}}

\seealso{
Other GCIMSSpectrum: 
\code{\link{GCIMSSpectrumSet}},
\code{\link{findPeaks,GCIMSSpectrum-method}}
}
\concept{GCIMSSpectrum}
