% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSSpectrumSet.R
\name{GCIMSSpectrumSet}
\alias{GCIMSSpectrumSet}
\title{Create a \linkS4class{GCIMSSpectrumSet} object}
\usage{
GCIMSSpectrumSet(spectra = list(), pData = NULL)
}
\arguments{
\item{spectra}{A named list of \link{GCIMSSpectrum} objects, one per sample,
named after their \code{SampleID}.}

\item{pData}{A \code{data.frame}/\code{DataFrame}/tibble with the phenotype data, or \code{NULL}.}
}
\value{
A \linkS4class{GCIMSSpectrumSet} object
}
\description{
Create a \linkS4class{GCIMSSpectrumSet} object
}
\seealso{
Other GCIMSSpectrum: 
\code{\link{GCIMSSpectrumSet-class}},
\code{\link{findPeaks,GCIMSSpectrum-method}}
}
\concept{GCIMSSpectrum}
