% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getEIC}
\alias{getEIC}
\title{Get extracted ion chromatogram}
\usage{
getEIC(object, ...)
}
\arguments{
\item{object}{An object that has ion chromatograms}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A numeric vector with the extracted ion chromatogram
}
\description{
Get extracted ion chromatogram
}
\examples{
x <- dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
getEIC(x) # c(`1` = 3, `2` = 7, `3` = 11)
}
