% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/xxx-deprecated-getSample-getSpectrum-getChromatogram.R
\name{getIMS}
\alias{getIMS}
\title{Deprecated function. Use getSpectrum() instead}
\usage{
getIMS(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=getSpectrum]{getSpectrum()}}}
}
\description{
Deprecated function. Use getSpectrum() instead
}
