% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getIMSScan}
\alias{getIMSScan}
\title{Get IMS scans}
\usage{
getIMSScan(object, ...)
}
\arguments{
\item{object}{An object that has IMS scans}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
An ion mobility spectrum
}
\description{
Get IMS scans
}
