% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-getTIS_getRIC.R
\name{getRIC,GCIMSSample-method}
\alias{getRIC,GCIMSSample-method}
\title{Get the reverse ion chromatogram}
\usage{
\S4method{getRIC}{GCIMSSample}(object)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}
}
\value{
A numeric vector with the reverse ion chromatogram
}
\description{
Get the reverse ion chromatogram
}
\examples{
sample_file <- system.file("extdata", "sample_formats", "small.mea.gz", package = "GCIMS")
s <- read_mea(sample_file)
ric <- getRIC(s)
}
