% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getRIC}
\alias{getRIC}
\title{Get the Reverse Ion Chromatogram}
\usage{
getRIC(object, ...)
}
\arguments{
\item{object}{An object to extract its reverse ion chromatogram}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The Reverse Ion Chromatogram, as a numeric vector or a
matrix (depending if the object is one sample or several)
}
\description{
Get the Reverse Ion Chromatogram
}
