% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSample-GCIMSDataset.R
\name{getSample,GCIMSDataset-method}
\alias{getSample,GCIMSDataset-method}
\title{Get a sample from a GCIMSDataset object}
\usage{
\S4method{getSample}{GCIMSDataset}(object, sample)
}
\arguments{
\item{sample}{Either an integer (sample index) or a string (sample name)}
}
\value{
The corresponding \link{GCIMSSample}
}
\description{
Get a sample from a GCIMSDataset object
}
