% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx-deprecated-helpers.R
\name{getSample}
\alias{getSample}
\title{Get a sample from a GCIMSDataset object}
\usage{
getSample(object, sample)
}
\arguments{
\item{object}{a \link{GCIMSDataset} object}

\item{sample}{Either an integer (sample index) or a string (sample name)}
}
\value{
The corresponding \link{GCIMSSample}
}
\description{
Get a sample from a GCIMSDataset object
}
\keyword{internal}
