% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSSample.R
\name{getSpectrum}
\alias{getSpectrum}
\title{Get IMS spectrum from a sample}
\usage{
getSpectrum(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL,
  aggregate = rowSums
)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}

\item{aggregate}{Function that takes the subsetted intensity matrix according to the region of interest
and aggregates the retention times, returning a vector representing the spectrum intensity. \code{rowSums} by default.}
}
\description{
Get IMS spectrum from a sample
}
\examples{
x <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
getSpectrum(x, rt_idx = 2)

# Take the maximum intensity in the region for each drift time:
sp1 <- getSpectrum(x, aggregate = function(x) apply(x, 1, max))
}
