% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIS_getRIC-GCIMSDataset.R
\name{getTIS,GCIMSDataset-method}
\alias{getTIS,GCIMSDataset-method}
\title{Get Total Ion Spectra matrix}
\usage{
\S4method{getTIS}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}
}
\value{
A matrix with samples in rows and the drift time in columns
}
\description{
Get Total Ion Spectra matrix
}
