% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIS_getRIC-GCIMSSample.R
\name{getTIS,GCIMSSample-method}
\alias{getTIS,GCIMSSample-method}
\title{Get the total ion spectrum}
\usage{
\S4method{getTIS}{GCIMSSample}(object)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}
}
\value{
A numeric vector with the total ion spectrum
}
\description{
Get the total ion spectrum
}
\examples{
sample_file <- system.file("extdata", "sample_formats", "small.mea.gz", package = "GCIMS")
s <- read_mea(sample_file)
tis <- getTIS(s)
}
