% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getTIS}
\alias{getTIS}
\title{Get the Total Ion Spectrum}
\usage{
getTIS(object, ...)
}
\arguments{
\item{object}{An object to extract its total ion spectrum}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The Total Ion Spectrum as a numeric vector or a matrix
(depending if the object is one sample or several)
}
\description{
Get the Total Ion Spectrum
}
