% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputePeakTable.R
\name{imputePeakTable}
\alias{imputePeakTable}
\title{Impute a Peak table}
\usage{
imputePeakTable(peak_table, dataset, cluster_stats)
}
\arguments{
\item{peak_table}{A matrix, with samples in rows and clusters in columns. It must have row names and column names.}

\item{dataset}{The dataset object to extract samples from}

\item{cluster_stats}{A data frame with the \code{dt_min_ms} and family columns}
}
\value{
The imputed peak_table
}
\description{
Impute a Peak table
}
