% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratePeaks-GCIMSDataset.R
\name{integratePeaks,GCIMSDataset-method}
\alias{integratePeaks,GCIMSDataset-method}
\title{Integrate peaks in a GCIMSDataset}
\usage{
\S4method{integratePeaks}{GCIMSDataset}(
  object,
  peak_list,
  integration_size_method = c("fixed_size", "free_size"),
  rip_saturation_threshold = 0.1
)
}
\arguments{
\item{object}{The \link{GCIMSDataset} object, modified inline}

\item{peak_list}{A data frame with peak lists}

\item{integration_size_method}{Either fixed_size or free_size}

\item{rip_saturation_threshold}{The threshold}
}
\value{
A modified \link{GCIMSDataset} object
}
\description{
Integrate peaks in a GCIMSDataset
}
