% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-integratePeaks.R
\name{integratePeaks,GCIMSSample-method}
\alias{integratePeaks,GCIMSSample-method}
\title{Peak integration for a GCIMSSample}
\usage{
\S4method{integratePeaks}{GCIMSSample}(
  object,
  peak_list,
  integration_size_method = c("fixed_size", "free_size"),
  rip_saturation_threshold = 0.1
)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{verbose}{If \code{TRUE}, debug information will be printed}
}
\value{
The modified \link{GCIMSSample}, with an updated peak list
}
\description{
Peak integration for a GCIMSSample
}
