% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-integratePeaks.R
\name{integratePeaks}
\alias{integratePeaks}
\title{Title}
\usage{
integratePeaks(
  object,
  peak_list,
  integration_size_method = c("fixed_size", "free_size"),
  rip_saturation_threshold = 0.1
)
}
\arguments{
\item{object}{The \link{GCIMSDataset} object}

\item{rip_saturation_threshold}{}
}
\description{
Title
}
