% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-intensity-GCIMSChromatogram.R
\name{intensity,GCIMSChromatogram-method}
\alias{intensity,GCIMSChromatogram-method}
\title{Get the intensity vector}
\usage{
\S4method{intensity}{GCIMSChromatogram}(object, rt_range = NULL, rt_idx = NULL)
}
\arguments{
\item{object}{A GCIMSChromatogram object}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}
}
\value{
The retention intensity vector
}
\description{
Get the intensity vector
}
