% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-S3.R
\name{intensity,GCIMSSample_S3-method}
\alias{intensity,GCIMSSample_S3-method}
\title{Get the intensity matrix of a GCIMSSample_S3 object}
\usage{
\S4method{intensity}{GCIMSSample_S3}(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL
)
}
\arguments{
\item{object}{The GCIMSSample_S3 object}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}
}
\description{
Get the intensity matrix of a GCIMSSample_S3 object
}
\examples{
m <- matrix(1:15, nrow = 3, ncol = 5)
obj <- new_GCIMSSampleS3(
  drift_time = 1:3,
  retention_time = 1:5,
  intensity = m,
  metadata = list(Name = "Sample1")
)
intensity(obj)
}
