% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{intensityMatrix}
\alias{intensityMatrix}
\title{Get the intensity matrix}
\usage{
intensityMatrix(object, ...)
}
\arguments{
\item{object}{An object with an intensity matrix}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A numeric matrix with drift time in the rows and retention time in columns
}
\description{
Get the intensity matrix
}
