% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{modgauss}
\alias{modgauss}
\title{Create a modified Gaussian (non symmetric) from two Gaussians}
\usage{
modgauss(x, mu, sigma1, sigma2)
}
\arguments{
\item{x}{A vector of indexes.}

\item{mu}{Integer.Position of the modified Gaussian maximum.
It corresponds to the mean of the left and right Gaussian
that generate the modified Gaussian}

\item{sigma1}{Numeric. Standard deviation of the left Gaussian.}

\item{sigma2}{Numeric. Standard deviation of the right Gaussian.}
}
\value{
A vector of intensities of the modified Gaussian.
}
\description{
Create a modified Gaussian (non symmetric) from two Gaussians
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{shift}()}
}
\concept{Dummy data functions}
