% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-S3.R
\name{new_GCIMSSampleS3}
\alias{new_GCIMSSampleS3}
\title{Create a GCIMSSample (S3 version)}
\usage{
new_GCIMSSampleS3(drift_time, retention_time, intensity, metadata)
}
\arguments{
\item{drift_time}{A numeric vector}

\item{retention_time}{A numeric vector}

\item{intensity}{A matrix, with size length(drift_time) x length(retention_time)}

\item{metadata}{A named list of name-value annotations}
}
\value{
A GCIMSSampleS3 object
}
\description{
This is mostly compatible with our previous list-based implementation.
}
\examples{
obj <- new_GCIMSSampleS3(
  drift_time = 1:3,
  retention_time = 1:5,
  intensity = matrix(1:15, nrow = 3, ncol = 5),
  metadata = list(Name = "Sample1")
)
}
