% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTable.R
\name{omit_times}
\alias{omit_times}
\title{Omit ROIs present in certain retention and drift times}
\usage{
omit_times(peak_list, rt_time_2_omit = NULL, dt_time_2_omit = NULL)
}
\arguments{
\item{peak_list}{The output of \code{\link[=peaks]{peaks()}}. Also, you can create your own peak table
and use it as input value for \code{peak_list}}

\item{rt_time_2_omit}{A vector including a set of retention times where ROIs detected
should not be considered. As default is is set as \code{NULL}}

\item{dt_time_2_omit}{A vector including a set of drift times where ROIs detected
should not be considered. As default is is set as \code{NULL}}
}
\value{
A \code{peak_list} without the ROIs present in the retention and drift times
not desired.
}
\description{
Extract the volume of each ROI across samples to create a peak table.
}
\examples{

peak_list <- data.frame(
  rt_apex_s =  c(1, 2, 3, 3, 4, 4,  5, 5, 6, 6),
  dt_apex_ms = c(2, 4, 6, 4, 8, 4, 10, 4, 4, 12)
)
peak_list_filt <- omit_times(peak_list, dt_time_2_omit = 4)


}
