% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW.R
\name{optimize_align_tr}
\alias{optimize_align_tr}
\title{This function finds the optimal polynomial degree to apply ptw to GMIMS data
when correcting retention time axis. The figure of merit is the mean of the
correlation between the reference RIC and samples RIC.}
\usage{
optimize_align_tr(dir_in, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{samples}{Numeric vector of integers. Identifies the set of
sample to be visualized from the dataset.}
}
\value{
An Integer between 0 and 5. If zero, no correction
in retention time is needed. Between 1 and 5 it indicates
the polynomial degree of the warping function.
}
\description{
This function finds the optimal polynomial degree to apply ptw to GMIMS data
when correcting retention time axis. The figure of merit is the mean of the
correlation between the reference RIC and samples RIC.
}
