% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{optimize_cow}
\alias{optimize_cow}
\title{Obtain the optimum warping for (COW)}
\usage{
optimize_cow(dir_in, dir_out, samples, by_rows, seg_vector, slack_vector)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{by_rows}{Logical. Direction to apply the function. If TRUE it
is applied by rows (drift time direction).
If FALSE, applied by columns
(that is the retention time direction).}

\item{seg_vector}{Vector of segment lengths.}

\item{slack_vetor}{Vector of slacks.}
}
\value{
A matrix with the optimum warping.
}
\description{
Obtain the optimum warping for (COW)
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Alignment functions: 
\code{\link{apply_cow}()},
\code{\link{cow}()},
\code{\link{cut_samples}()},
\code{\link{gcims_alignment}()},
\code{\link{gcims_cut_samples}()}
}
\concept{Alignment functions}
