% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-GCIMSSample.R
\name{overlay_peaklist}
\alias{overlay_peaklist}
\title{Overlay a peak list to a plot}
\usage{
overlay_peaklist(
  peaklist = NULL,
  ...,
  dt_range = NULL,
  rt_range = NULL,
  pdata = NULL,
  color_by = NULL,
  mapping_roi = c(dt_min_ms = "dt_min_ms", dt_max_ms = "dt_max_ms", rt_min_s =
    "rt_min_s", rt_max_s = "rt_max_s"),
  palette = P40
)
}
\arguments{
\item{peaklist}{A data frame with at least the columns: \code{dt_min_ms}, \code{dt_max_ms}, \code{rt_min_s}, \code{rt_max_s}
and optionally additional columns (e.g. the column given to \code{color_by})}

\item{...}{Ignored.}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{pdata}{A phenotype data data frame, with a \code{SampleID} column. This column will be used to
merge \code{pdata} with \code{peaklist}, so \code{color_by} can specify a phenotype.}

\item{color_by}{A character with a column name of \code{peaklist} or \code{pdata}. Used to color the border of
the added rectangles and apices. A string with a color name is also acceptable.}

\item{palette}{A character vector with color names to use drawing the rectangles. Use \code{NULL} to let \code{ggplot2} set the defaults.}

\item{mapping_row}{A 4-elements named character vector with the names of the columns from \code{peaklist} that will
be used as the rectangle coordinates.}
}
\value{
A list with the ggplot layers to overlay, including \code{geom_rect} and possibly \code{geom_point} and \code{scale_fill_manual}.
}
\description{
Overlay a peak list to a plot
}
\details{
If \code{peaklist} includes \code{dt_apex_ms} and \code{rt_apex_s} a cross will be plotted on the peak apex.
}
