% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData-GCIMSDataset.R
\name{pData<-,GCIMSDataset,ANY-method}
\alias{pData<-,GCIMSDataset,ANY-method}
\title{Set the phenotype data}
\usage{
\S4method{pData}{GCIMSDataset,ANY}(object) <- value
}
\arguments{
\item{value}{The data frame with annotations, it should have a FileName column and a SampleID column.}
}
\description{
Set the phenotype data
}
