% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks-GCIMSDataset.R
\name{peaks<-,GCIMSDataset-method}
\alias{peaks<-,GCIMSDataset-method}
\title{Set the peak list}
\usage{
\S4method{peaks}{GCIMSDataset}(object) <- value
}
\arguments{
\item{value}{The data frame with a peak list}
}
\description{
Set the peak list
}
