% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Make a plot interactive}
\usage{
plot_interactive(plt)
}
\arguments{
\item{plt}{A ggplot plot}
}
\value{
A plotly plot
}
\description{
Wraps the \code{plt} with \code{\link[plotly:ggplotly]{plotly::ggplotly()}} and sets the \code{xaxis} and \code{yaxis}
ticks to \code{"auto"}, so the axis labels are updated when zooming.
}
\examples{
d <- data.frame(x = c(1,2), y=c(1,2))
plt <- ggplot2::ggplot(d) + ggplot2::geom_point(aes(x = x, y = y))
plot_interactive(plt)
}
