% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-plotRaw.R
\name{plotRaw,GCIMSSample-method}
\alias{plotRaw,GCIMSSample-method}
\title{Topographical plot of a GCIMSSample object}
\usage{
\S4method{plotRaw}{GCIMSSample}(object, dt_range = NULL, rt_range = NULL)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{dt_range}{The drift time range to plot (in milliseconds)}

\item{rt_range}{The retention time range to plot (in seconds)}
}
\value{
A plot of the GCIMSSample
}
\description{
Topographical plot of a GCIMSSample object
}
\examples{
dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
plotRaw(dummy_obj)
}
