% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotRaw}
\alias{plotRaw}
\title{Plot raw data}
\usage{
plotRaw(object, ...)
}
\arguments{
\item{object}{An object to plot raw data}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A plot
}
\description{
Plot raw data
}
