% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIS_getRIC-GCIMSDataset.R
\name{plotRIC,GCIMSDataset-method}
\alias{plotRIC,GCIMSDataset-method}
\title{Plot Reverse Ion Chromatograms}
\usage{
\S4method{plotRIC}{GCIMSDataset}(object, rt_range = NULL, sample = NULL)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{sample}{A number or a string with the sample index or name. If \code{NULL}, all samples are returned}
}
\value{
A plot
}
\description{
Plot Reverse Ion Chromatograms
}
