% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIS_getRIC-GCIMSSample.R
\name{plotRIC,GCIMSSample-method}
\alias{plotRIC,GCIMSSample-method}
\title{Plot the reverse ion chromatogram of a sample}
\usage{
\S4method{plotRIC}{GCIMSSample}(object, rt_range = NULL)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}
}
\value{
A ggplot2 plot object
}
\description{
Plot the reverse ion chromatogram of a sample
}
