% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotRIC}
\alias{plotRIC}
\title{Plot Reverse Ion Chromatogram}
\usage{
plotRIC(object, ...)
}
\arguments{
\item{object}{An object to plot its reverse ion chromatogram}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A plot
}
\description{
Plot Reverse Ion Chromatogram
}
