% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIS_getRIC-GCIMSSample.R
\name{plotTIS,GCIMSSample-method}
\alias{plotTIS,GCIMSSample-method}
\title{Plot the total ion spectrum of a sample}
\usage{
\S4method{plotTIS}{GCIMSSample}(object, dt_range = NULL)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}
}
\value{
A ggplot2 plot object
}
\description{
Plot the total ion spectrum of a sample
}
