% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotTIS}
\alias{plotTIS}
\title{Plot total ion spectrum}
\usage{
plotTIS(object, ...)
}
\arguments{
\item{object}{An object to plot the total ion spectrum}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A plot
}
\description{
Plot total ion spectrum
}
