% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{prealign,GCIMSSample-method}
\alias{prealign,GCIMSSample-method}
\title{Align a GCIMSSample object, in drift time and to the injection point in retention time}
\usage{
\S4method{prealign}{GCIMSSample}(object, align_dt, align_ip, rip_ref_ms, min_start, rt_ref)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{align_dt}{if \code{TRUE}, align the drift time axis using a multiplicative correction}

\item{align_ip}{if \code{TRUE} a multiplicative correction will be done in retention time before applying the other algorithm}

\item{rip_ref_ms}{The reference position of the Reactant Ion Peak in the dataset (in ms)}

\item{min_start}{minimun injection point, to calculate where to begin the spectrums and cut as few points as posible, to be used in injection point alignment}

\item{rt_ref}{retention time reference for alignment to injection point}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample object, in drift time and to the injection point in retention time
}
