% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions.R
\name{psalsa}
\alias{psalsa}
\title{Estimate spectra baseline using psalsa}
\usage{
psalsa(spectra, lambda = 1e+07, p = 0.001, k = -1, maxit = 25)
}
\arguments{
\item{spectra}{matrix with one spectrum per row}

\item{lambda}{smoothing parameter (generally 1e5 - 1e8)}

\item{p}{asymmetry parameter}

\item{k}{peak height parameter (usually 5\% of maximum intensity)}

\item{maxit}{max number of iterations}
}
\description{
Estimate spectra baseline using psalsa
}
\seealso{
Other Baseline removal functions: \code{\link{gcims_baseline_removal}}
}
\concept{Baseline removal functions}
