% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions.R
\name{psalsa_one}
\alias{psalsa_one}
\title{Estimate baseline of a curve using psalsa}
\usage{
psalsa_one(y, lambda = 1e+07, p = 0.001, k = -1, maxit = 25)
}
\arguments{
\item{y}{Either a chromatogram or a spectrum}

\item{lambda}{Smoothing parameter (generally 1e5 - 1e8)}

\item{p}{Asymmetry parameter}

\item{k}{Peak height parameter (usually 5\% of maximum
intensity)}

\item{maxit}{Max number of iterations}
}
\value{
The estimated baseline
}
\description{
Estimate baseline of a curve using psalsa
}
\references{
{

Oller-Moreno, S., Pardo, A., Jiménez-Soto, J. M., Samitier, J., & Marco, S. (2014, February).
Adaptive Asymmetric Least Squares baseline estimation for analytical instruments.
In 2014 IEEE 11th International Multi-Conference on Systems, Signals & Devices (SSD14) (pp. 1-5). IEEE.

 }
}
\seealso{
Other Baseline removal functions: 
\code{\link{gcims_baseline_removal}()}
}
\author{
Sergio Oller-Moreno,  \email{soller@.ibecbarcelona.eu}


Institute for Bioengieering of Catalonia. Signal and Information Processing for Sensing Systems group. Barcelona, Spain.
}
\concept{Baseline removal functions}
