% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCIMS.R
\name{readGCIMS}
\alias{readGCIMS}
\title{GC-IMS Samples Reading}
\usage{
readGCIMS(dir_in, dir_out)
}
\arguments{
\item{x}{Name of the sample}
}
\value{
An R object that contains the matrix and the retention times
}
\description{
GC-IMS Samples Reading
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Reading functions}
