% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{realize}
\alias{realize}
\title{Realize the object by executing all delayed operations}
\usage{
realize(object, ...)
}
\arguments{
\item{object}{An object with delayed operations}

\item{...}{Additional arguments for downstream methods}
}
\value{
The object, without delayed operations
}
\description{
Realize the object by executing all delayed operations
}
