% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{retentionTime}
\alias{retentionTime}
\title{Get retention time vector}
\usage{
retentionTime(object)
}
\arguments{
\item{object}{An object with retention time}
}
\value{
A numeric vector with the retention time
}
\description{
Get retention time vector
}
