% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-intensity-GCIMSChromatogram.R
\name{rtime,GCIMSChromatogram-method}
\alias{rtime,GCIMSChromatogram-method}
\title{Get the retention time vector}
\usage{
\S4method{rtime}{GCIMSChromatogram}(object)
}
\arguments{
\item{object}{A GCIMSChromatogram}
}
\value{
The retention time vector (in s)
}
\description{
Get the retention time vector
}
\seealso{
Other GCIMSChromatogram: 
\code{\link{GCIMSChromatogram}},
\code{\link{GCIMSChromatogram-class}},
\code{\link{dtime,GCIMSChromatogram-method}},
\code{\link{estimateBaseline,GCIMSChromatogram-method}},
\code{\link{findPeaks,GCIMSChromatogram-method}},
\code{\link{intensity,GCIMSChromatogram-method}},
\code{\link{smooth,GCIMSChromatogram-method}}
}
\concept{GCIMSChromatogram}
