% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-GCIMSDataset.R
\name{rtime,GCIMSDataset-method}
\alias{rtime,GCIMSDataset-method}
\title{Get a reference retention time vector for the dataset}
\usage{
\S4method{rtime}{GCIMSDataset}(object, sample = NULL)
}
\arguments{
\item{object}{A GCIMSDataset}

\item{sample}{A number or a string with the sample index or name. If \code{NULL}, the reference drift time is returned}
}
\value{
a retention time vector
}
\description{
Get a reference retention time vector for the dataset
}
