% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-S3.R
\name{rtime,GCIMSSample_S3-method}
\alias{rtime,GCIMSSample_S3-method}
\title{Get the retention time vector of a GCIMSSample_S3 object}
\usage{
\S4method{rtime}{GCIMSSample_S3}(object)
}
\arguments{
\item{object}{The GCIMSSample_S3 object}
}
\description{
Get the retention time vector of a GCIMSSample_S3 object
}
\examples{
obj <- new_GCIMSSampleS3(
  drift_time = 1:3,
  retention_time = 1:5,
  intensity = matrix(1:15, nrow = 3, ncol = 5),
  metadata = list(Name = "Sample1")
)
all(rtime(obj) == 1:5)
}
