% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames-GCIMSDataset.R
\name{sampleNames,GCIMSDataset-method}
\alias{sampleNames,GCIMSDataset-method}
\alias{sampleNames<-,GCIMSDataset,ANY-method}
\title{Sample names}
\usage{
\S4method{sampleNames}{GCIMSDataset}(object)

\S4method{sampleNames}{GCIMSDataset,ANY}(object) <- value
}
\arguments{
\item{object}{GCIMSDataset object}

\item{value}{A character vector of length the number of samples with the sample names}
}
\value{
The \link{GCIMSDataset} object
}
\description{
Sample names
}
\section{Functions}{
\itemize{
\item \code{sampleNames(object = GCIMSDataset) <- value}: Sample names

}}
