% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{shift}
\alias{shift}
\title{Generate a random shift in peak positions (in retention or drift axes)}
\usage{
shift(min_shift, max_shift, npeaks)
}
\arguments{
\item{min_shift}{Integer. Minimum shift in indexes allowed.}

\item{max_shift}{Integer. Maximum shift in indexes allowed.}

\item{npeaks}{Integer. Number of peaks to be shifted.}
}
\value{
A vector of shifts (one per peak to be shifted).
}
\description{
Generate a random shift in peak positions (in retention or drift axes)
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Dummy data functions: 
\code{\link{add_baseline}()},
\code{\link{add_tail}()},
\code{\link{fill_concentration}()},
\code{\link{force_interpolation}()},
\code{\link{gcims_create_dummy_params}()},
\code{\link{gcims_create_dummy_sample}()},
\code{\link{gcims_create_dummy_set}()},
\code{\link{modgauss}()}
}
\concept{Dummy data functions}
