% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-GCIMSDataset.R
\name{smooth,GCIMSDataset-method}
\alias{smooth,GCIMSDataset-method}
\title{Smoothing a GCIMS dataset using a Savitzky-Golay filter}
\usage{
\S4method{smooth}{GCIMSDataset}(x, dt_length_ms = 0.14, rt_length_s = 3, dt_order = 2, rt_order = 2)
}
\arguments{
\item{x}{A \link{GCIMSDataset} object, modified in-place}

\item{dt_length_ms}{the length of the filter in drift time (in ms)}

\item{rt_length_s}{The length of the filter in retention time (in s)}

\item{dt_order}{The order of the filter in drift time}

\item{rt_order}{The order of the filter in retention time}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
Smoothing a GCIMS dataset using a Savitzky-Golay filter
}
