% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-GCIMSSample.R
\name{smooth,GCIMSSample-method}
\alias{smooth,GCIMSSample-method}
\title{Smoothing a GCIMS sample using a Savitzky-Golay filter}
\usage{
\S4method{smooth}{GCIMSSample}(x, dt_length_ms, rt_length_s, dt_order = 2, rt_order = 2)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{dt_length_ms}{the length of the filter in drift time (in ms)}

\item{rt_length_s}{The length of the filter in retention time (in s)}

\item{dt_order}{The order of the filter in drift time}

\item{rt_order}{The order of the filter in retention time}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Smoothing a GCIMS sample using a Savitzky-Golay filter
}
