% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSpectrum-aaa-class.R
\name{smooth,GCIMSSpectrum-method}
\alias{smooth,GCIMSSpectrum-method}
\title{Smoothing a GCIMS Spectrum using a Savitzky-Golay filter}
\usage{
\S4method{smooth}{GCIMSSpectrum}(x, dt_length_ms, dt_order = 2)
}
\arguments{
\item{x}{A \link{GCIMSSpectrum} object}

\item{dt_length_ms}{the length of the filter in drift time (in ms)}

\item{dt_order}{The order of the filter in drift time}
}
\value{
The modified \link{GCIMSSpectrum}
}
\description{
Smoothing a GCIMS Spectrum using a Savitzky-Golay filter
}
