% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaw-GCIMSSample.R
\name{tidy.GCIMSSample}
\alias{tidy.GCIMSSample}
\title{Tidy the raw data into a long data frame}
\usage{
\method{tidy}{GCIMSSample}(x, dt_range = NULL, rt_range = NULL, dt_idx = NULL, rt_idx = NULL, ...)
}
\arguments{
\item{x}{A \link{GCIMSSample} object}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}

\item{...}{unused}
}
\value{
A data frame with \code{dt_ms}, \code{rt_s} and \code{Intensity} columns
}
\description{
Tidy the raw data into a long data frame
}
