% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample.R
\name{UpdateGCIMSSample}
\alias{UpdateGCIMSSample}
\title{Updates old saved GCIMSSample object to the latest version}
\usage{
UpdateGCIMSSample(object)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}
}
\value{
The updated \link{GCIMSSample} object
}
\description{
This function is useful when you have saved a \link{GCIMSSample} object
with a previous version of the GCIMS package and you want to load it
using a new version of the package.
}
\details{
The function allows you to update the old object, adding missing
slots, etc so it is fully compatible with the new class definition.
}
\examples{
obj <- GCIMSSample(drift_time=1:2, retention_time=1:3, data = matrix(1:6, nrow=2, ncol=3))
# Here we modify the object as if it was older:
attr(obj, 'history') <- NULL
attr(obj, "class_version") <- numeric_version("0.0.1")
# Update the object:
newobj <- UpdateGCIMSSample(obj)

}
