% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-updateObject.R
\name{updateObject,GCIMSSample-method}
\alias{updateObject,GCIMSSample-method}
\title{Updates old saved GCIMSSample object to the latest version}
\usage{
\S4method{updateObject}{GCIMSSample}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A \link{GCIMSSample} object, typically that has been serialized and loaded from disk}
}
\value{
The updated \link{GCIMSSample} object
}
\description{
This function is useful when you have saved a \link{GCIMSSample} object
with a previous version of the GCIMS package and you want to load it
using a new version of the package.
}
\details{
The function allows you to update the old object, adding missing
slots, etc so it is fully compatible with the new class definition.
}
\examples{
obj <- GCIMSSample(drift_time=1:2, retention_time=1:3, data = matrix(1:6, nrow=2, ncol=3))
# Update the object:
newobj <- updateObject(obj)

}
