% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-GCIMSSample.R
\name{add_peaklist_rect}
\alias{add_peaklist_rect}
\title{Add peak list rectangles to a raw plot}
\usage{
add_peaklist_rect(
  plt,
  peaklist,
  color_by = NULL,
  col_prefix = "",
  pdata = NULL,
  palette = P40
)
}
\arguments{
\item{plt}{The output of \code{\link[=plot]{plot()}} when applied to a \link{GCIMSSample}}

\item{peaklist}{A data frame with at least the columns: \code{dt_min_ms}, \code{dt_max_ms}, \code{rt_min_s}, \code{rt_max_s}
and optionally additional columns (e.g. the column given to \code{color_by})}

\item{color_by}{A character with a column name of \code{peaklist}. Used to color the border of
the added rectangles}

\item{col_prefix}{After clustering, besides \code{dt_min_ms}, we also have}

\item{pdata}{A phenotype data data frame, with a SampleID column to be merged into peaklist so color_by can specify
a phenotype
\code{freesize_dt_min_ms}. Use \code{col_prefix = "freesize_"} to plot the \code{freesize} version}

\item{palette}{A character vector with color names to use drawing the rectangles. Use \code{NULL} to let \code{ggplot2} set the defaults.}
}
\value{
The given \code{plt} with rectangles showing the ROIs and crosses showing the apexes
}
\description{
Add peak list rectangles to a raw plot
}
\details{
If \code{peaklist} includes \code{dt_apex_ms} and \code{rt_apex_s} a cross will be plotted on the peak apex.
}
\examples{
dt <- 1:10
rt <- 1:10
int <- matrix(0.0, nrow = length(dt), ncol = length(rt))

int[2, 4:8] <- c(.5, .5, 1, .5, 0.5)
int[3, 4:8] <- c(0.5, 2, 2, 2, 0.5)
int[4, 4:8] <- c(1, 2, 5, 2, 1)
int[5, 4:8] <- c(0.5, 2, 2, 2, 0.5)
int[6, 4:8] <- c(.5, .5, 1, .5, 0.5)

dummy_obj <-GCIMSSample(
  drift_time = dt,
  retention_time = rt,
  data = int
)
plt <- plot(dummy_obj)

# Add a rectangle on top of the plot
rect <- data.frame(
  dt_min_ms = 2.75,
  dt_max_ms = 5.6,
  rt_min_s = 4.6,
  rt_max_s = 7.4,
)

add_peaklist_rect(
  plt = plt,
  peaklist = rect
)
}
