% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-plotRaw.R
\name{add_peaklist_rect}
\alias{add_peaklist_rect}
\title{Add peak list rectangles to a raw plot}
\usage{
add_peaklist_rect(plt, peaklist)
}
\arguments{
\item{plt}{The output of \code{\link[=plotRaw]{plotRaw()}} when applied to a \link{GCIMSSample}}

\item{peaklist}{The output of \code{\link[=peaks]{peaks()}} of a GCIMSSample}
}
\value{
The given \code{plt} with rectangles showing the ROIs and crosses showing the apexes
}
\description{
Add peak list rectangles to a raw plot
}
