% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-GCIMSSample.R
\name{add_peaklist_rect}
\alias{add_peaklist_rect}
\title{Add peak list rectangles to a raw plot}
\usage{
add_peaklist_rect(
  plt,
  peaklist,
  color_by = NULL,
  col_prefix = "",
  pdata = NULL,
  palette = P40
)
}
\arguments{
\item{plt}{The output of \code{\link[=plot]{plot()}} when applied to a \link{GCIMSSample}}

\item{peaklist}{A data frame with at least the columns: \code{dt_min_ms}, \code{dt_max_ms}, \code{rt_min_s}, \code{rt_max_s}
and optionally additional columns (e.g. the column given to \code{color_by})}

\item{color_by}{A character with a column name of \code{peaklist}. Used to color the border of
the added rectangles}

\item{col_prefix}{After clustering, besides \code{dt_min_ms}, we also have}

\item{pdata}{A phenotype data data frame, with a SampleID column to be merged into peaklist so color_by can specify
a phenotype
\code{freesize_dt_min_ms}. Use \code{col_prefix = "freesize_"} to plot the \code{freesize} version}

\item{palette}{A character vector with color names to use drawing the rectangles. Use \code{NULL} to let \code{ggplot2} set the defaults.}
}
\value{
The given \code{plt} with rectangles showing the ROIs and crosses showing the apexes
}
\description{
Add peak list rectangles to a raw plot
}
\details{
If \code{peaklist} includes \code{dt_apex_ms} and \code{rt_apex_s} a cross will be plotted on the peak apex.
}
