% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSDataset.R
\name{align,GCIMSDataset-method}
\alias{align,GCIMSDataset-method}
\title{Align a GCIMS dataset}
\usage{
\S4method{align}{GCIMSDataset}(object, method = "pow", align_ip = TRUE, ...)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{method}{Method for alignment, should be "ptw" or "pow"}

\item{align_ip}{if TRUE a multiplicative correction will be done in retention time before applying the other algorithm}

\item{...}{additional parameters for POW alignment}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
The alignment uses a multiplicative correction in drift time and a
Parametric Time Warping correction in retention time
}
