% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSDataset.R
\name{align,GCIMSDataset-method}
\alias{align,GCIMSDataset-method}
\title{Align a GCIMS dataset}
\usage{
\S4method{align}{GCIMSDataset}(object, method_rt = "ptw", align_dt = TRUE, align_ip = TRUE, ...)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{method_rt}{Method for alignment, should be "ptw" or "pow"
if pow is selected the package "pow must be installed, to do so visit:
https://github.com/sipss/pow}

\item{align_dt}{if \code{TRUE} the drift time axis will be aligned using a multiplicative correction}

\item{align_ip}{if TRUE a multiplicative correction will be done in retention time before applying the other algorithm}

\item{...}{additional parameters for POW alignment}
}
\value{
The modified \link{GCIMSDataset}
}
\description{
The alignment uses a multiplicative correction in drift time and a
Parametric Time Warping correction in retention time
}
