% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{align,GCIMSSample-method}
\alias{align,GCIMSSample-method}
\title{Align a GCIMSSample object, in retention time}
\usage{
\S4method{align}{GCIMSSample}(object, method_rt, ric_ref, ric_ref_rt, ...)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{method_rt}{Method for alignment, should be "ptw" or "pow"}

\item{ric_ref}{The reference Reverse Ion Chromatogram}

\item{ric_ref_rt}{The retention times corresponding to \code{ric_ref}}

\item{...}{Additional arguments passed on to the alignment method.}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample object, in retention time
}
