% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-GCIMSSample.R
\name{align,GCIMSSample-method}
\alias{align,GCIMSSample-method}
\title{Align a GCIMSSample object, in drift and retention time}
\usage{
\S4method{align}{GCIMSSample}(
  object,
  rip_ref_ms,
  ric_ref,
  ric_ref_rt,
  ip_ref_s,
  method_rt,
  align_dt,
  align_ip,
  ...
)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rip_ref_ms}{The reference position of the Reactant Ion Peak in the dataset (in ms)}

\item{ric_ref}{The reference Reverse Ion Chromatogram}

\item{ric_ref_rt}{The retention times corresponding to \code{ric_ref}}

\item{ip_ref_s}{Injection point reference for retention time}

\item{method_rt}{Method for alignment, should be "ptw" or "pow"}

\item{align_ip}{if TRUE a multiplicative correction will be done in retention time before applying the other algorithm}

\item{...}{Additional arguments passed on to the alignment method.}
}
\value{
The modified \link{GCIMSSample}
}
\description{
Align a GCIMSSample object, in drift and retention time
}
