% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions_NEW.R
\name{align_tr}
\alias{align_tr}
\title{Correction of retention time axis using Parametric Time Warping (PTW)}
\usage{
align_tr(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where aligned data files are stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
sample to be visualized from the dataset.}
}
\description{
This chromatrogram computes the RIC of the reference samples and
uses it to correct each of the EIC of a sample, for the all the selected
samples on the dataset. The warping function that relates the
retention time axes of the reference and the sample to be corrected
is linear.
}
