% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{alignDt}
\alias{alignDt}
\title{Align an object in drift time}
\usage{
alignDt(x, y, ...)
}
\arguments{
\item{x}{An object to align}

\item{y}{Another object to use as reference}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
The object, modified
}
\description{
Align an object in drift time
}
