% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{build_peak_table}
\alias{build_peak_table}
\title{Build a peak table}
\usage{
build_peak_table(peak_list_clustered, aggregate_conflicting_peaks = NULL)
}
\arguments{
\item{peak_list_clustered}{The output of \link{gcims_figures_of_merit}.}

\item{aggregate_conflicting_peaks}{\code{NULL} or a function. What to do, in case two peaks from the same sample
have been assigned to the same cluster. If \code{NULL}, throw an error. If \code{mean}, \code{max} or any other function,
we will summarize all the conflicting volumes into that number (e.g. "take the maximum of the peaks")}
}
\value{
A list with the peak table and the ROI duplicity information. The peak table
is a data frame with
}
\description{
Build a peak table
}
\examples{
pl <- data.frame(
  SampleID = c("S1", "S1", "S2", "S2"),
  cluster = c(1, 2, 1, 2),
  Volume = c(10, 20, 8, 18)
)
build_peak_table(pl)
}
