% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{build_peak_table}
\alias{build_peak_table}
\title{Build a peak table}
\usage{
build_peak_table(peak_list_clustered, aggregate_conflicting_peaks = NULL)
}
\arguments{
\item{peak_list_clustered}{The output of \code{\link[=gcims_compute_fom]{gcims_compute_fom()}}. Also, you can create your own peak table
and use it as input value for \code{peak_list_clustered} (see first example below)}

\item{aggregate_conflicting_peaks}{\code{NULL} or a function. What to do, in case two peaks from the same sample
have been assigned to the same cluster. If \code{NULL}, throw an error. If \code{mean}, \code{max} or any other function,
we will summarize all the conflicting volumes into that number (e.g. "take the maximum of the peaks")}
}
\value{
A list with three fields: \code{peak_table}, \code{peak_table_matrix}, and \code{peak_table_duplicity}.
\code{peak_table}, and \code{peak_table_matrix}, provide information of the peak table. \code{peak_table} is a dataframe
containing cluster volumes, whose columns represent samples and rows clusters. \code{peak_table_matrix} presents
the same information content as \code{peak_table} but in matrix form. Note that in \code{peak_table} columns represent
clusters and rows samples. Finally, \code{peak_table_duplicity} is a dataframe that shows ROI duplicity information
among clusters. Ideally, only one peak per sample should belong to a cluster.
}
\description{
Extract the volume of each ROI across samples to create a peak table.
}
\examples{
\donttest{
# Create your peak table from scratch:
pl <- data.frame(
  SampleID = c("S1", "S1", "S2", "S2"),
  cluster = c(1, 2, 1, 2),
  Volume = c(10, 20, 8, 18)
)
build_peak_table(pl)

# Create a peak table from the output of the function gcims_figure_of_merit()
dir_in <- system.file("extdata", package = "GCIMS")
peak_list_fom <- readRDS(file.path(dir_in, "peak_list_fom.rds"))
peak_table <- build_peak_table(peak_list_fom, aggregate_conflicting_peaks = max)

peak_table$peak_table_matrix
}

}
