% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaw-GCIMSSample.R
\name{cubic_root_trans}
\alias{cubic_root_trans}
\title{Cubic root transformation}
\usage{
cubic_root_trans()
}
\value{
A scale transformation object of name "cubic_root"
}
\description{
A scales transformation to be used with ggplot2.
}
\details{
This function is exported because we are using it in vignettes, but it may
become unavailable in future versions
}
