% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_samples_functions.R
\name{cut_samples}
\alias{cut_samples}
\title{Cuts retention time and drift time}
\usage{
cut_samples(loaded_sample, rt_range = NULL, dt_range = NULL)
}
\arguments{
\item{loaded_sample}{Current sample to be cut.}

\item{rt_range}{A  vector of two components. Beginning and end
of the retention time cut.}

\item{dt_range}{A  vector of two components. Beginning and end
of the drift time cut.}
}
\value{
An cut matrix.
}
\description{
Cuts retention time and drift time
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}


}
\seealso{
Other Cutting functions: 
\code{\link{gcims_cut_samples}()}
}
\concept{Cutting functions}
