% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-AllGenerics.R
\name{describeAsList}
\alias{describeAsList}
\title{Format as a list}
\usage{
describeAsList(object, ...)
}
\value{
A list with a brief description/representation of the object
}
\description{
Format as a list
}
\keyword{internal}
