% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-rtime-GCIMSDataset.R
\name{dtime,GCIMSDataset-method}
\alias{dtime,GCIMSDataset-method}
\title{Get A reference drift time vector for the dataset}
\usage{
\S4method{dtime}{GCIMSDataset}(object, sample = NULL)
}
\arguments{
\item{object}{A GCIMSDataset}

\item{sample}{A number or a string with the sample index or name. If \code{NULL}, the reference drift time is returned}
}
\value{
a drift time vector
}
\description{
Get A reference drift time vector for the dataset
}
