% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-GCIMSChromatogram.R
\name{estimateBaseline,GCIMSChromatogram-method}
\alias{estimateBaseline,GCIMSChromatogram-method}
\alias{baseline,GCIMSChromatogram-method}
\alias{baseline<-,GCIMSChromatogram-method}
\title{Estimate the baseline of a GCIMS Chromatogram using a connect local minima algorithm}
\usage{
\S4method{estimateBaseline}{GCIMSChromatogram}(object, rt_length_s)

\S4method{baseline}{GCIMSChromatogram}(object, rt_range = NULL, rt_idx = NULL, .error_if_missing = TRUE)

\S4method{baseline}{GCIMSChromatogram}(object) <- value
}
\arguments{
\item{object}{A \link{GCIMSChromatogram} object}

\item{rt_length_s}{The length of the baseline region. It should be
comparable or longer than the peak width}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{rt_idx}{A numeric vector with the retention time indices to extract (or a logical vector of the length of retention time)}

\item{.error_if_missing}{A logical. If \code{TRUE} (default) give an error if baseline is not estimated. Returns \code{NULL} otherwise.}

\item{value}{A vector with a baseline of the same length as \code{intensity(object)}}
}
\value{
The modified \link{GCIMSChromatogram}
}
\description{
The baseline is estimated by connecting local minima and interpolating from those.
The local minima are identified as "the minima in each region of length x"
The length of the regions are given in seconds in the \code{region_s} parameter.
}
\section{Functions}{
\itemize{
\item \code{baseline(GCIMSChromatogram)}: Get the baseline

\item \code{baseline(GCIMSChromatogram) <- value}: Set the baseline

}}
