% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-estimateBaseline.R
\name{estimateBaseline,GCIMSSample-method}
\alias{estimateBaseline,GCIMSSample-method}
\alias{baseline,GCIMSSample-method}
\alias{baseline<-,GCIMSSample-method}
\title{Estimate the baseline of a GCIMS Sample using a connect local minima algorithm}
\usage{
\S4method{estimateBaseline}{GCIMSSample}(object, dt_peak_fwhm_ms, dt_region_multiplier, rt_length_s)

\S4method{baseline}{GCIMSSample}(object)

\S4method{baseline}{GCIMSSample}(object) <- value
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{dt_peak_fwhm_ms}{Full Width at Half Maximum in milliseconds. Used to
determine the length of the regions where local minima are searched.}

\item{dt_region_multiplier}{A multiplier to calculate the region}

\item{rt_length_s}{The length of the baseline region. It should be
comparable or longer than the peak width}

\item{value}{A matrix with the sample baseline of the same dimensions as \code{dim(object)}}
}
\value{
The modified \link{GCIMSSample}
}
\description{
The baseline is estimated by connecting local minima and interpolating from those.
The local minima are identified as "the minima in each region of length x"
The length of the regions are estimated as \verb{fwhm * a multiplier / 2.3482}. This
assumes it's several times
}
\section{Functions}{
\itemize{
\item \code{baseline(GCIMSSample)}: Get the baseline

\item \code{baseline(GCIMSSample) <- value}: Set the baseline

}}
