% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSpectrum-estimateBaseline.R
\name{estimateBaseline,GCIMSSpectrum-method}
\alias{estimateBaseline,GCIMSSpectrum-method}
\alias{baseline,GCIMSSpectrum-method}
\alias{baseline<-,GCIMSSpectrum-method}
\title{Estimate the baseline of a GCIMS Spectrum using a connect local minima algorithm}
\usage{
\S4method{estimateBaseline}{GCIMSSpectrum}(object, dt_peak_fwhm_ms, dt_region_multiplier = 12)

\S4method{baseline}{GCIMSSpectrum}(object, dt_range = NULL, dt_idx = NULL, .error_if_missing = TRUE)

\S4method{baseline}{GCIMSSpectrum}(object) <- value
}
\arguments{
\item{object}{A \link{GCIMSSpectrum} object}

\item{dt_peak_fwhm_ms}{Full Width at Half Maximum in milliseconds. Used to
determine the length of the regions where local minima are searched.}

\item{dt_region_multiplier}{A multiplier to calculate the region}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{.error_if_missing}{A logical. If \code{TRUE} (default) give an error if baseline is not estimated. Returns \code{NULL} otherwise.}

\item{value}{A vector with a baseline of the same length as \code{intensity(object)}}
}
\value{
The modified \link{GCIMSSpectrum}
}
\description{
The baseline is estimated by connecting local minima and interpolating from those.
The local minima are identified as "the minima in each region of length x"
The length of the regions are estimated as \verb{fwhm * a multiplier / 2.3482}. This
assumes it's several times
}
\section{Functions}{
\itemize{
\item \code{baseline(GCIMSSpectrum)}: Get the baseline

\item \code{baseline(GCIMSSpectrum) <- value}: Set the baseline

}}
