% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-GCIMSSample.R
\name{filterRt,GCIMSSample-method}
\alias{filterRt,GCIMSSample-method}
\title{Filter GCIMSSample samples by retention time}
\usage{
\S4method{filterRt}{GCIMSSample}(object, rt_range)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}
}
\value{
A subset of the sample, only in the selected \code{rt_range}
}
\description{
Filter GCIMSSample samples by retention time
}
\examples{
sample_file <- system.file("extdata", "sample_formats", "small.mea.gz", package = "GCIMS")
s <- read_mea(sample_file)
s <- filterRt(s, rt_range = c(5, 50))
}
