% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-peaks.R
\name{findPeaks,GCIMSSample-method}
\alias{findPeaks,GCIMSSample-method}
\title{Peak detection for a GCIMSSample}
\usage{
\S4method{findPeaks}{GCIMSSample}(object, noise_level = 3, verbose = FALSE)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{noise_level}{Scalar number. The number of times the standard deviation
above the noise level needed to detect a peak.}

\item{verbose}{If \code{TRUE}, debug information will be printed}
}
\value{
The modified \link{GCIMSSample}, with a peak list
}
\description{
Peak detection for a GCIMSSample
}
