% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks-GCIMSSample.R
\name{findPeaks,GCIMSSample-method}
\alias{findPeaks,GCIMSSample-method}
\title{Peak detection for a GCIMSSample}
\usage{
\S4method{findPeaks}{GCIMSSample}(object, ...)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{...}{
  Arguments passed on to \code{\link[=findPeaksImpl]{findPeaksImpl}}
  \describe{
    \item{\code{verbose}}{If \code{TRUE} information will be printed on screen}
    \item{\code{dt_length_ms,rt_length_s}}{Length of the filters used to compute the
second derivative. See details.}
    \item{\code{dt_peakwidth_range_ms,rt_peakwidth_range_s}}{A vector of length 2 with the minimum and maximum peak width. See details}
    \item{\code{exclude_rip}}{Whether to exclude ROIs with a drift time apex smaller than the RIP drift time end.}
    \item{\code{iou_overlap_threshold}}{A number, between 0 and 1. Pairs of ROIs with an intersection over union larger than this threshold are merged.}
    \item{\code{debug_idx}}{A list with two numeric vectors named \code{dt} and \code{rt} each of them having a the indices to where debug info is kept}
  }}
}
\value{
The modified \link{GCIMSSample}, with a peak list
}
\description{
Peak detection for a GCIMSSample
}
