% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks-GCIMSSample.R
\name{findPeaks,GCIMSSample-method}
\alias{findPeaks,GCIMSSample-method}
\title{Peak detection for a GCIMSSample}
\usage{
\S4method{findPeaks}{GCIMSSample}(
  object,
  noise_level = 3,
  verbose = FALSE,
  dt_length_ms = 0.14,
  rt_length_s = 3,
  iou_overlap_threshold = 0.2
)
}
\arguments{
\item{object}{A \link{GCIMSSample} object}

\item{noise_level}{A number used to determine the minimum peak height threshold}

\item{verbose}{If \code{TRUE} information will be printed on screen}

\item{dt_length_ms}{Length (in ms) used in the Savitzky-Golay filter to compute the second derivative with respect to drift time.}

\item{rt_length_s}{Length (in s) used in the Savitzky-Golay filter to compute the second derivative with respect to retention time.}

\item{iou_overlap_threshold}{A number, between 0 and 1. Regions of interest b}
}
\value{
The modified \link{GCIMSSample}, with a peak list
}
\description{
Peak detection for a GCIMSSample
}
