% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaksImpl}
\alias{findPeaksImpl}
\title{Peak and ROI detection}
\usage{
findPeaksImpl(
  drift_time,
  retention_time,
  int_mat,
  noise_level,
  verbose = FALSE,
  dt_length_pts = 11,
  rt_length_pts = 21,
  iou_overlap_threshold = 0.2
)
}
\arguments{
\item{drift_time}{The drift time vector}

\item{retention_time}{The retention time vector}

\item{int_mat}{The intensity matrix, of dimensions \code{length(drift_time)} rows and \code{length(retention_time)} columns}

\item{noise_level}{A number used to determine the minimum peak height threshold}

\item{verbose}{If \code{TRUE} information will be printed on screen}

\item{dt_length_pts, rt_length_pts}{Length of the filters used to compute the second derivative. See details}

\item{iou_overlap_threshold}{A number, between 0 and 1. Regions of interest b}
}
\description{
Detects regions of interest with peaks in a sample.
}
\details{
Peaks are detected on the partial second derivative of the intensity matrix.
The peak detection happens by scanning for peaks in the drift time direction, using
the second derivative

In detail, the approach is as follows:
\enumerate{
\item We compute the second derivative with respect to the drift and retention times.
\item We detect the RIP and based on its peak width we estimate the minimum peak distance between peaks in drift time
\item Based on a region without peaks in the original matrix, we estimate the amplitude of noise peaks. This amplitude
is used as a basis to determine the minimum peak height required for peak detection.
\item We run pracma::findPeaks on both partial derivatives
\item We intersect the maxima to discard saddle points
\item We merge similar ROIs using a threshold on the intersection over union
\item Get some ROI metrics and return.
}
}
\keyword{internal}
