% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-AllGenerics.R
\name{GCIMS-generics}
\alias{GCIMS-generics}
\alias{dtime}
\alias{getChromatogram}
\alias{getSpectrum}
\alias{getSample}
\alias{getTIS}
\alias{getRIC}
\alias{plotRaw}
\alias{plotTIS}
\alias{plotRIC}
\alias{filterDt}
\alias{decimate}
\alias{align}
\alias{alignDt}
\alias{estimateBaseline}
\alias{baseline}
\alias{baseline<-}
\alias{integratePeaks}
\title{GCIMS Generics}
\usage{
dtime(object, ...)

getChromatogram(object, ...)

getSpectrum(object, ...)

getSample(object, ...)

getTIS(object, ...)

getRIC(object, ...)

plotRaw(object, ...)

plotTIS(object, ...)

plotRIC(object, ...)

filterDt(object, ...)

decimate(object, ...)

align(object, ...)

alignDt(x, y, ...)

estimateBaseline(object, ...)

baseline(object, ...)

baseline(object) <- value

integratePeaks(object, ...)
}
\arguments{
\item{object}{An object to get the baseline}

\item{...}{Additional arguments for downstream methods}

\item{x}{An object to align}

\item{y}{Another object to use as reference}

\item{value}{baseline to set}
}
\value{
A numeric vector with the drift time

A chromatogram

An ion mobility spectrum

A sample object

The Total Ion Spectrum as a numeric vector or a matrix
(depending if the object is one sample or several)

The Reverse Ion Chromatogram, as a numeric vector or a
matrix (depending if the object is one sample or several)

A plot

A plot

A plot

The object, modified

The object, modified

The object, modified

The object, modified

The object, with a baseline estimated

The baseline of the object

The object

The object, with integrated peaks
}
\description{
Generics defined at the GCIMS package. We are open to moving them
to an existing generics-only package if you need so.
}
\section{Functions}{
\itemize{
\item \code{dtime()}: Get drift time vector

\item \code{getChromatogram()}: Get chromatogram

\item \code{getSpectrum()}: Get spectrums

\item \code{getSample()}: Get a Sample

\item \code{getTIS()}: Get the Total Ion Spectrum

\item \code{getRIC()}: Get the Reverse Ion Chromatogram

\item \code{plotRaw()}: Plot raw data

\item \code{plotTIS()}: Plot total ion spectrum

\item \code{plotRIC()}: Plot Reverse Ion Chromatogram

\item \code{filterDt()}: Filter in Drift time

\item \code{decimate()}: Decimate an object

\item \code{align()}: Align an object

\item \code{alignDt()}: Align an object in drift time

\item \code{estimateBaseline()}: Estimate the baseline in an object

\item \code{baseline()}: Get the baseline of an object

\item \code{baseline(object) <- value}: Set the baseline of an object

\item \code{integratePeaks()}: Integrate peaks of an object

}}
\examples{
x <- GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
dtime(x) # c(1,2)

}
