% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{gcims_alignment}
\alias{gcims_alignment}
\title{Peak alignment using  Correlation Optimized Warping (COW)}
\usage{
gcims_alignment(dir_in, dir_out, samples, by_rows, Seg, Slack)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{by_rows}{Logical. Direction to apply the function. If TRUE it by rows (drift time direction).
If FALSE, applied by columns (that is the retention time direction).}

\item{Seg}{The segment length. The number of segments is N = floor(ncol(X)/Seg)
if Seg is a two row matrix, the first row contains indexes of the T
array, ranging from 1 to length(T). The second row contains the indexes
for the X matrix, ranging from 1 to ncol(X). These indexes define the
boundaries for each segment}

\item{Slack}{Maximum range or degree of warping in each segment.}
}
\description{
Peak alignment using  Correlation Optimized Warping (COW)
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Alignment functions: \code{\link{cow}}
}
\concept{Alignment functions}
