% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{gcims_alignment}
\alias{gcims_alignment}
\title{Peak alignment using Correlation Optimized Warping (COW)}
\usage{
gcims_alignment(dir_in, dir_out, samples, time, seg_vector, slack_vector)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}

\item{seg_vector}{Vector of segment lengths.}

\item{slack_vector}{Vector of slacks.}
}
\value{
An aligned gcims dataset.
}
\description{
Peak alignment using Correlation Optimized Warping (COW)
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Alignment functions: 
\code{\link{apply_cow}()},
\code{\link{cow}()},
\code{\link{optimize_cow}()}
}
\concept{Alignment functions}
