% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{gcims_alignment}
\alias{gcims_alignment}
\title{Peak alignment using Correlation Optimized Warping (COW)}
\usage{
gcims_alignment(dir_in, dir_out, samples, time, seg_vector, slack_vector)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are
loaded from.}

\item{dir_out}{Output directory. Where data files after alignment
correction are stored.}

\item{samples}{A vector. Set of samples to be aligned(e.g.: c(1, 2,
3)).}

\item{time}{Sets the dimension to be corrected: drift time or
retention time. Introduce "Retention" for retention time; or "Drift" for
drift time.}

\item{seg_vector}{Vector of segment lengths.}

\item{slack_vector}{Vector of slacks.}
}
\value{
A set of S3 objects.
}
\description{
Peak alignment using Correlation Optimized Warping (COW)
}
\details{
\code{gcims_baseline_removal} aligns sample chromatograms (in
retention time) or spectra (in drift time).
The alignment in drift time is achieved by transforming the selected temporal
axis of each sample according to the reference using the Correlation Optimized
Warping (COW).

In COW, the reference signal of length \emph{N} is divided
in \emph{L} segments of length \emph{I} and so it is done for the signal to
be aligned. When the correction is applied along retention time direction,
the signals used to estimate COW model are Total Ion Chromatograms (TIC).
In case of correcting drift time misalignments, COW model is estimated
using Total Ion Spectra (TIS).

The alignment between reference and sample
representatives is performed segment by segment, allowing slight
modifications on the segment lengths of a sample. Segment lengths are
controlled by the so-called slack parameter, \emph{t}. For each of the
segments of a sample, segment length is varied so that the correlation
between sample and reference is maximized. At this point a linear
interpolation is realized to recover sample segments of the original length
and sample period. Once the COW model is known, it is applied to the rest
of chromatrograms (if the correction is retention time) or spectra (if the
correction is in drift time) of a sample. This is done for all samples
selected by the argument \code{samples}.

Cow method is optimized by selecting the parameters that maximize
the average correlation between the reference and the rest of the samples.
In particular, the arguments \code{seg_vector} and \code{slack_vector}
provide, respectively, the sets of possible values of \emph{L} and \emph{t}
for COW parameter optimization.

The alignment in retention time is achieved applying parametric time warping.
The arguments \code{seg_vector} and \code{slack_vector} are not necessary for
the retention time alignment.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3, 7)

# Example of Drift time Alignment:
# Before:
gcims_plot_spec(dir_in, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")

# After:
time <- "Drift"
seg_vector <- seq(from = 10, to = 30, by = 5)
slack_vector <- seq(from = 1, to = (seg_vector[length(seg_vector)] - 4), by = 3)
gcims_alignment(dir_in, dir_out, samples, time, seg_vector, slack_vector)
gcims_plot_spec(dir_out, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
}

}
\references{
{
Niels-Peter Vest Nielsen, Jens Micheal Carstensen and Jørn Smedegaard 'Aligning of singel and multiple
wavelength chromatographic profiles for chemometric data analysis using correlation optimised warping'
J. Chrom. A 805(1998)17-35

Correlation optimized warping and dynamic time warping as preprocessing methods for chromatographic Data
Giorgio Tomasi, Frans van den Berg and Claus Andersson, Journal of Chemometrics 18(2004)231-241
}
}
\author{
Sergio Oller-Moreno,  \email{soller@ibecbarcelona.eu}

Inspired on the work by:

Giorgio Tomasi, \email{gt@kvl.dk}

Frans van den Berg 070821 (GT) \email{fb@kvl.dk}

Royal Agricultural and Veterinary University - Department of Food Science
Quality and Technology - Spectroscopy and Chemometrics group - Denmark
www.models.kvl.dk
}
