% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions.R
\name{gcims_baseline_removal}
\alias{gcims_baseline_removal}
\title{Baseline Removal using Psalsa algorithm}
\usage{
gcims_baseline_removal(dir_in, dir_out, samples, by_rows, lambda, p, k)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{by_rows}{Logical. Direction to apply the function. If TRUE it by rows (drift time direction).
If FALSE, applied by columns (that is the retention time direction).}

\item{lambda}{Smoothing parameter (generally 1e5 - 1e8)}

\item{p}{Asymmetry parameter}

\item{k}{Peak height parameter (usually 5\% of maximum intensity)}
}
\value{
A baseline removed gcims dataset.
}
\description{
Baseline Removal using Psalsa algorithm
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Baseline removal functions: \code{\link{psalsa}}
}
\concept{Baseline removal functions}
