% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_removal_functions.R
\name{gcims_baseline_removal}
\alias{gcims_baseline_removal}
\title{Remove baseline of a dataset using psalsa}
\usage{
gcims_baseline_removal(
  dir_in,
  dir_out,
  samples,
  time,
  lambda = 1e+07,
  p = 0.001,
  k = -1
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are
loaded from.}

\item{dir_out}{Output directory. Where data files after baseline
correction are stored.}

\item{samples}{A vector. Set of samples to which remove the baseline
(e.g.: c(1, 2, 3)).}

\item{time}{Sets the dimension to be corrected: drift time or
retention time. Introduce "Retention" for retention time; or "Drift" for
drift time.}

\item{lambda}{Smoothing parameter (generally 1e3 - 1e9).}

\item{p}{Asymmetry parameter (generally 1e-3 - 1e-6).}

\item{k}{Peak height parameter (usually 5\% of maximum intensity).}
}
\value{
A set of S3 objects.
}
\description{
Remove baseline of a dataset using psalsa
}
\details{
\code{gcims_baseline_removal} performs baseline correction on sample
chromatograms (in retention time) or spectra (in drift time).This method
relies in is \emph{psalsa} algorithm. \emph{psalsa} is based on the Asymmetric Least
Squares (ALS) algorithm for baseline estimation, although slightly modified
to reject the effects of large peaks above the baseline more easily. \emph{psalsa}
depends on three parameters: a penalty on the second derivative
(\code{lambda}), a penalty on the value of the baseline with respect the
value of the signal (\code{p}), and third parameter (\code{k}), that
modifies substantially the value of (\code{p}) in the event of large
intensity peaks.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3, 7)

# Example of Drift time Baseline Removal:
# Before:
gcims_plot_spec(dir_in, samples, rt_value = 60, dt_range = NULL, colorby = "Class")

# After:
time <- "Drift"
lambda <- 1E3
p <- 0.001
k <- -1
gcims_baseline_removal(dir_in, dir_out, samples, time, lambda, p, k)
gcims_plot_spec(dir_out, samples, rt_value = 60, dt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\references{
{

Eilers, P.H.C. (2003) A perfect smoother. Anal. Chem.  p. 3631–6.
\doi{10.1021/ac034173t}

Oller-Moreno, S., Pardo, A., Jiménez-Soto, J. M., Samitier, J., & Marco, S.
(2014, February). Adaptive Asymmetric Least Squares baseline estimation for
analytical instruments. In 2014 IEEE 11th International Multi-Conference on
Systems, Signals & Devices (SSD14) (pp. 1-5). IEEE.
\doi{10.1109/SSD.2014.6808837}

}
}
\author{
Sergio Oller-Moreno,  \email{soller@.ibecbarcelona.eu}

Institute for Bioengineering of Catalonia. Signal and Information Processing for Sensing Systems group. Barcelona, Spain.
}
\concept{Baseline removal functions}
