% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{gcims_create_reference}
\alias{gcims_create_reference}
\title{Select or create a reference sample}
\usage{
gcims_create_reference(dir_in, dir_out, samples, ref_number)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{ref_number}{Reference sample. If NULL the reference
is the mean sample.}
}
\value{
A reference sample
}
\description{
Select or create a reference sample
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
