% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_samples_functions.R
\name{gcims_cut_samples}
\alias{gcims_cut_samples}
\title{Cut samples in a retention time - drift time rectangle}
\usage{
gcims_cut_samples(dir_in, dir_out, samples, ret_cut, drift_cut)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{ret_cut}{A  vector of two components. Beginning and end
of the retention time cut.}

\item{drift_cut}{A  vector of two components. Beginning and end
of the drift time cut.}
}
\value{
An cut gcims dataset.
}
\description{
Cut samples in a retention time - drift time rectangle
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Alignment functions: 
\code{\link{apply_cow}()},
\code{\link{cow}()},
\code{\link{cut_samples}()},
\code{\link{gcims_alignment}()},
\code{\link{optimize_cow}()}
}
\concept{Alignment functions}
