% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_cut_samples}
\alias{gcims_cut_samples}
\title{Cut samples in a retention time - drift time rectangle}
\usage{
gcims_cut_samples(dir_in, dir_out, samples, rt_range, dt_range)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where cut data files are
stored.}

\item{samples}{A vector. Set of samples to be aligned(e.g.: c(1, 2,
3)).}

\item{rt_range}{A  vector of two components. Beginning and end of the
retention time cut.If NULL the complete retention time range is used.}

\item{dt_range}{A  vector of two components. Beginning and end of the
drift time cut. If NULL the complete drift time range is used.}
}
\value{
A set of S3 objects.
}
\description{
Cut samples in a retention time - drift time rectangle
}
\details{
\code{cut_samples} cuts a sample in a retention time - drift time
  rectangle according to the retention time / drift time ranges given by
  function arguments \code{rt_range} / \code{dt_range}. Use this function to
  focus on the retention time - drift time region where chemical information
  is more abundant, that is, where you can find a high peak densities by
  visual inspection.
}
\note{
By reducing the size of data, the computational time of signal
  pre-processing stage can reduced substantially.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example Sample data cutting:
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
rt_range <-c(70, 125)
dt_range <- c(8, 9.25)
gcims_cut_samples(dir_in, dir_out, samples, rt_range, dt_range)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)
}
\seealso{
Other Utility functions: 
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
