% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_decimate}
\alias{gcims_decimate}
\title{Decimation}
\usage{
gcims_decimate(dir_in, dir_out, samples, p_rt, q_rt, p_dt, q_dt)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{p_rt}{The current number of averaged scans.}

\item{q_rt}{The new number of averaged scans.}

\item{p_dt}{The current number of averaged scans.}

\item{q_dt}{The new number of averaged scans.}
}
\value{
A reduced  gcims dataset.
}
\description{
Decimation
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()

# Before:
setwd(dir_in)
samples <- c(3, 7)
gcims_plot_chrom(dir_in, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")
gcims_plot_spec(dir_in, samples, rt_value = NULL,  dt_range = NULL, colorby = "Class")

# After:
p_rt <- 1
q_rt <- 5
p_dt <- 1
q_dt <- 2
gcims_decimate(dir_in, dir_out, samples, p_rt, q_rt, p_dt, q_dt)
setwd(dir_out)
gcims_plot_chrom(dir_out, samples, dt_value = NULL,  rt_range = c(50, 226) , colorby = "Class")
gcims_plot_spec(dir_out, samples, rt_value = NULL,  dt_range = c(7.75, 9.6), colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
