% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_decimate}
\alias{gcims_decimate}
\title{Decimation}
\usage{
gcims_decimate(dir_in, dir_out, samples, q_rt, q_dt)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where decimated data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to be decimated.}

\item{q_rt}{Numeric. Binning factor in retention time.}

\item{q_dt}{Numeric. Binning factor in drift time.}
}
\value{
A set of S3 objets.
}
\description{
Decimation
}
\details{
\code{gcims_decimate} performs downsampling in retention and drift
time axes of GCIMS data. In particular, it reduces sampling frequency in
retention and dritf times respectively, by the  factors \code{q_rt} and
\code{q_dt}. Use this function if you are interested in both increasing the
signal to noise ratio of data and compress it. Please take in to account
that decimation also reduces data resolution.
}
\note{
\code{gcims_decimate} introduces a delay in retention and drift time
axes.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()

# Before:
setwd(dir_in)
samples <- c(3, 7)
gcims_plot_chrom(dir_in, samples, dt_value = NULL,  rt_range = NULL, colorby = "Class")
gcims_plot_spec(dir_in, samples, rt_value = NULL,  dt_range = NULL, colorby = "Class")

# After:
q_rt <- 5
q_dt <- 2
gcims_decimate(dir_in, dir_out, samples, q_rt, q_dt)
setwd(dir_out)
gcims_plot_chrom(dir_out, samples, dt_value = NULL,  rt_range = c(50, 226) , colorby = "Class")
gcims_plot_spec(dir_out, samples, rt_value = NULL,  dt_range = c(7.75, 9.6), colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\references{
{ Oppenheim, Alan V.; Schafer, Ronald W.; Buck, John R. (1999).
"4". Discrete-Time Signal Processing (2nd ed.). Upper Saddle River, N.J.:
Prentice Hall. p. 168. ISBN 0-13-754920-2. }
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
