% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_figures_of_merit.R
\name{gcims_figures_of_merit}
\alias{gcims_figures_of_merit}
\title{Figures Of Merit Calculation}
\usage{
gcims_figures_of_merit(
  dir_in,
  dir_out,
  peak_list,
  cluster_stats,
  integration_size = c("cluster_roi", "individual_roi"),
  rip_saturation_threshold = 0.1
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where files containing the Figures of Meir of each ROI are
stored.}

\item{peak_list}{A data frame. The peak list where we will add the figures of merit.}

\item{cluster_stats}{A data frame with cluster statistics (ROI limits...)}

\item{integration_size}{Either "cluster_roi" or "individual_roi". When computing the volume,
the integration size can be set individually for each sample or use the reference cluster size given by \code{cluster_stats}.}

\item{rip_saturation_threshold}{A number. The fraction of the maximum RIP. If the RIP at the ROI is below that
fraction, we will consider the peak to be saturated.}
}
\value{
The given peak list, with the added columns
}
\description{
Calculates the area, volume asymmetry and saturation of each peak roi in \code{peak_list}
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
roi_selection <- tempdir()
fom <- tempdir()

# Use BiocParallel for pararellization
library(BiocParallel)
# Enable parallellization with 3 workers
register(SnowParam(workers = 3, progressbar = TRUE, exportglobals = FALSE), default = TRUE)

# Example of Calculating the Figures of Merit'
peak_list <- gcims_rois_selection(dir_in, roi_selection, samples = c(3, 7), noise_level = 3)
clustering <- group_peak_list(
  peak_list,
  distance_method = "sd_scaled_euclidean",
  clustering = list(method = "kmedoids", Nclusters = 13)
)
roi_fusion_out <- gcims_rois_fusion(clustering$peak_list_clustered, clustering$cluster_stats)
peak_list_fom <- gcims_figures_of_merit(
  dir_in = roi_selection,
  dir_out = fom,
  peak_list = roi_fusion_out$peak_list_clustered,
  cluster_stats = roi_fusion_out$cluster_stats
)
head(peak_list_fom)

files_roi <- list.files(path = roi_selection, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files_roi))

files_fom <- list.files(path = fom, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files_fom))
}
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
