% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_interpolate}
\alias{gcims_interpolate}
\title{Interpolates data according to retention and drift time sampling frequencies}
\usage{
gcims_interpolate(dir_in, dir_out, samples, time)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}
}
\value{
An interpolated dataset
}
\description{
Interpolates data according to retention and drift time sampling frequencies
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", "to_interpolate", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Drift time Interpolation
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
time <- "Drift"
gcims_interpolate(dir_in, dir_out, samples, time)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
