% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_interpolate}
\alias{gcims_interpolate}
\title{Interpolates data to obtain constant sampling frequencies in retention and drift times}
\usage{
gcims_interpolate(dir_in, dir_out, samples, time)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where interpolated data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which their data matrices need to be interpolated.}

\item{time}{Sets the dimension to be corrected: drift time or
retention time. Introduce "Retention" for retention time; or "Drift" for
drift time.}
}
\value{
A set of S3 objects.
}
\description{
Interpolates data to obtain constant sampling frequencies in retention and drift times
}
\details{
\code{gcims_interpolate} performs a linear interpolation on gcims
data along the time axis selected in \code{gcims_interpolate}, and for all
samples in \code{samples}.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", "to_interpolate", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Drift time Interpolation
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
time <- "Drift"
gcims_interpolate(dir_in, dir_out, samples, time)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\references{
{ Oppenheim, Alan V.; Schafer, Ronald W.; Buck, John R. (1999).
  "4". Discrete-Time Signal Processing (2nd ed.). Upper Saddle River, N.J.:
  Prentice Hall. p. 168. ISBN 0-13-754920-2. }
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
