% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{gcims_missing_imputation}
\alias{gcims_missing_imputation}
\title{Missing Values Imputation}
\usage{
gcims_missing_imputation(peak_table, dir_in, cluster_stats)
}
\arguments{
\item{peak_table}{The matrix to be imputed, with samples in rows (see example)}

\item{dir_in}{Directory with baseline removed samples}

\item{cluster_stats}{Cluster statistics (reference ROI limits)}
}
\value{
A matrix with samples in rows, clusters in columns and volumes
as values, without missing values
}
\description{
Missing Values Imputation
}
\details{
\code{gcims_missing_imputation} calculates the volume of the ROIs that
have missing values. It uses the coordinates of the reference ROI for each
cluster and substitute the missing value by the volume in this region without
baseline.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
peak_table_obj <- readRDS(file.path(dir_in, "peak_table.rds"))
roi_fusion_out <- readRDS(file.path(dir_in, "roi_fusion_out.rds"))
peak_table_to_impute <- peak_table_obj$peak_table_mat

# Peak table before imputation
peak_table_to_impute


peak_table_imputed <- gcims_missing_imputation(peak_table = peak_table_to_impute,
                                               dir_in = dir_in, # use bsln
                                               cluster_stats = roi_fusion_out$cluster_stats
                                               )
# Peak table after imputation
peak_table_imputed
}

}
\concept{Imputation functions}
