% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{gcims_missing_imputation}
\alias{gcims_missing_imputation}
\title{Missing Values Imputation}
\usage{
gcims_missing_imputation(peak_table, dir_in, cluster_stats)
}
\arguments{
\item{peak_table}{The matrix to be imputed, with samples in rows (see example)}

\item{dir_in}{Directory with baseline removed samples}

\item{cluster_stats}{Cluster statistics (reference ROI limits)}
}
\value{
A matrix with samples in rows, clusters in columns and volumes
as values, without missing values
}
\description{
Missing Values Imputation
}
\details{
\code{gcims_missing_imputation} calculates the volume of the ROIs that
have missing values. It uses the coordinates of the reference ROI for each
cluster and subistitue the missing value by the volume in this region without
baseline.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
roi_selection <- tempfile("dir")
fom <- tempfile("dir")
pl <- gcims_rois_selection(dir_in, roi_selection, samples = c(3, 7), noise_level = 3)
clust <- group_peak_list(
  pl,
  distance_method = "sd_scaled_euclidean",
  clustering = list(method = "kmedoids", Nclusters = 13)
)
# Baseline removal should be here, disabled for speed
# bsln <- tempfile("dir)
# gcims_remove_baseline(
#   dir_in = roi_selection,
#   dir_out = bsln,
#   samples = c(3, 7),
#   clust$peak_list_clustered
# )
#
fusion_res <- gcims_rois_fusion(
  clust$peak_list_clustered,
  clust$cluster_stats
)
peak_list_foms <- gcims_figures_of_merit(
  dir_in = roi_selection, # use bsln
  dir_out = fom,
  peak_list = fusion_res$peak_list_clustered,
  cluster_stats = fusion_res$cluster_stats
)
peak_table_obj <- build_peak_table(peak_list_foms, aggregate_conflicting_peaks = max)
# Ideally dir_in should point to baseline corrected, but we skip this
# here for brevity.
peak_table_to_impute <- peak_table_obj$peak_table_mat
peak_table_imputed <- gcims_missing_imputation(
  peak_table = peak_table_to_impute,
  dir_in = fom, # use bsln
  cluster_stats = fusion_res$cluster_stats
)
head(peak_table_imputed)
}
\seealso{
Other Imputation functions: 
\code{\link{gcims_peak_imputation}()}
}
\concept{Imputation functions}
