% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{gcims_peak_imputation}
\alias{gcims_peak_imputation}
\title{Peak Table imputation using RF}
\usage{
gcims_peak_imputation(dir_in, dir_out, prop_samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{prop_samples}{Mimimum proportion of hits per feature in the dataset to
consider a feature as valid.}
}
\value{
A roi table for the whole dataset.
}
\description{
Peak Table imputation using RF
}
\seealso{
Other Imputation functions: 
\code{\link{gcims_missing_imputation}()}
}
\concept{Imputation functions}
