% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_picking_functions.R
\name{gcims_peak_picking}
\alias{gcims_peak_picking}
\title{Two dimensional peak picking}
\usage{
gcims_peak_picking(
  dir_in,
  dir_out,
  samples,
  rem_baseline = TRUE,
  min_length_tr = 50,
  min_length_td = 10,
  preprocess = TRUE
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{Samples to which perform the peak picking.}

\item{rem_baseline}{Perform a baseline removal in retention / drift time
axes. TRUE by default.}

\item{min_length_tr}{Minimum peak length in retention time (now indexes).}

\item{min_length_td}{Minimum peak length in drift time (now indexes).}

\item{preprocess}{Preprocess the samples?. TRUE or FALSE}
}
\value{
A peak table per sample.
}
\description{
Two dimensional peak picking
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3
# Example of Retention Two-Dimensional Peak Picking
  gcims_peak_picking(dir_in, dir_out, samples, rem_baseline = TRUE,
  min_length_tr = 50, min_length_td = 10, preprocess = TRUE)

setwd(dir_out)
M3 <- readRDS("M3.rds")
roi_df <- M3$data$roi_df
head(roi_df)
files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)
}
\concept{Peak Peaking functions}
