% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_chrom}
\alias{gcims_plot_chrom}
\title{Plots TICs or EICs.}
\usage{
gcims_plot_chrom(dir_in, samples, dt_value = NULL, rt_range = NULL, colorby)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{dt_value}{Selected drift time value. NULL by default.}

\item{rt_range}{Min a Max retention time values. NULL by
default.}

\item{colorby}{Metadata parameter used for color the samples.}
}
\value{
A set of Chromatograms.
}
\description{
Plots TICs or EICs.
}
\examples{
wd <- getwd()
dir_in <- tempdir()
dir_out <-  file.path(dir_in,"dummy")
dir.create (dir_out, FALSE)
samples_per_class <- 2
params <- gcims_create_dummy_params()
gcims_create_dummy_set(dir_in, dir_out, samples_per_class, params)
samples <- 1
dt_value <- 8
gcims_plot_chrom(dir_out, samples, dt_value)
unlink(dir_out, recursive = TRUE)
setwd(wd)
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_spec}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
