% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_chrom}
\alias{gcims_plot_chrom}
\title{Plots TICs or EICs.}
\usage{
gcims_plot_chrom(dir_in, samples, td_ind = NULL, rt_range = NULL)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{td_ind}{Selected drift time value. NULL by default.}

\item{rt_range}{Min a Max retention time values. NULL by
default.}
}
\value{
A set of Chromatograms.
}
\description{
Plots TICs or EICs.
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_spec}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
