% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_spec}
\alias{gcims_plot_spec}
\title{Plots TISa or spectra}
\usage{
gcims_plot_spec(dir_in, samples, tr_ind = NULL, dt_range = NULL)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{tr_ind}{Selected retention time value. NULL by default.}

\item{dt_range}{Min a Max drift time values. NULL by
default.}
}
\value{
A set of spectra.
}
\description{
Plots TISa or spectra
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_visualization}()}
}
\concept{Visualization functions}
