% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_spec}
\alias{gcims_plot_spec}
\title{Plots TISa or spectra}
\usage{
gcims_plot_spec(
  dir_in,
  samples,
  rt_value = NULL,
  dt_range = NULL,
  colorby = "Name"
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{rt_value}{Selected retention time value. NULL by default.}

\item{dt_range}{Min a Max drift time values. NULL by
default.}

\item{colorby}{Metadata parameter used for color the samples.}
}
\value{
A set of spectra.
}
\description{
Plots TISa or spectra
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
samples <- c(3, 7 ,8, 14, 20, 21)
gcims_plot_spec(dir_in, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")
setwd(current_dir)

}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
