% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_spec}
\alias{gcims_plot_spec}
\title{Plots TISa or spectra}
\usage{
gcims_plot_spec(
  dir_in,
  samples,
  rt_value = NULL,
  dt_range = NULL,
  colorby = "Name"
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be
to be visualized.}

\item{rt_value}{Selected retention time value. NULL by default.}

\item{dt_range}{Min a Max drift time values. NULL by
default.}

\item{colorby}{Metadata parameter used for color the samples.}
}
\value{
A set of spectra.
}
\description{
Plots TISa or spectra
}
\examples{
\dontrun{
wd <- getwd()
dir_in <- tempdir()
dir_out <-  file.path(dir_in,"dummy")
dir.create (dir_out, FALSE)
samples_per_class <- 2
params <- gcims_create_dummy_params()
gcims_create_dummy_set(dir_in, dir_out, samples_per_class, params)
samples <- 1
gcims_plot_spec(dir_out, samples)
unlink(dir_out, recursive = TRUE)
setwd(wd)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
