% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_plot_spec}
\alias{gcims_plot_spec}
\title{Plots TIS or spectra}
\usage{
gcims_plot_spec(
  dir_in,
  samples,
  rt_value = NULL,
  dt_range = NULL,
  colorby = "Name"
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{samples}{Numeric vector of integers. Identifies the set of
sample to be visualized from the dataset.}

\item{rt_value}{Numeric. Selects the retention time to show the
corresponding sample spectra . If NULL, the Total Ion Spectra (TIS) of
samples is shown. NULL by default.}

\item{dt_range}{Min and Max drift time values. If NULL the complete
drift time range is used. NULL by default.}

\item{colorby}{A variable included in the metadata used to color
sample spectra.}
}
\value{
A plot of spectra colored according to some variable included in the
  metadata.
}
\description{
Plots TIS or spectra
}
\details{
\code{gcims_plot_spec} uses the function
  \code{\link[ggplot2]{geom_line}} to plot a set of sample spectra at given retention
  time (or their TIS) and for a specific range of drift times. Use \code{gcims_plot_spec}
  to visualize the effects of digital smoothing, baseline correction and signal alignment
  algorithms along the drift time axis.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
samples <- c(3, 7 ,8, 14, 20, 21)
gcims_plot_spec(dir_in, samples, rt_value = NULL, dt_range = NULL, colorby = "Class")
setwd(current_dir)

}
\references{
{
Wickham, Hadley. "ggplot2." Wiley Interdisciplinary Reviews: Computational Statistics 3.2 (2011): 180-185.
 }
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_view_sample}()}
}
\concept{Visualization functions}
