% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{gcims_prepare_data}
\alias{gcims_prepare_data}
\title{Data preparation}
\usage{
gcims_prepare_data(dir_in, dir_out, samples, params)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where prepared data files are
stored.}

\item{samples}{A vector. Set of samples to which remove the baseline
(e.g.: c(1, 2, 3)).}

\item{params}{A list of lists with parameter values to perform
filtering and decimation in retention and drift time axes.}
}
\value{
A set of S3 objects. Additionally, it returns a list with two variables needed to perform sample alignment in drift
and retention time axes: \code{tis} and \code{rics}. First variable (a matrix) corresponds to the Total Ion Spectra of
samples, while second one (a matrix) to their Reactive Ion Chromatograms, respectively. These outputs are input variables for
the function \code{gcims_align_data()}.
}
\description{
Interpolates, filters and decimates data.
}
\details{
\code{gcims_prepare_data()} prepares samples for the subsequent data
pre-processing stage. Each of the samples is interpolated, filtered (optional),
and decimated (optional). To select which pre-processing steps apply to data use the input
variable \code{params}, described below:
\describe{
\item{\verb{filter$do, filter$order_rt, filter$length_rt, filter$order_dt}, \code{filter$length_dt}}{
Set of variables that control data filtering using Savitzky-Golay filters.  \code{do} is a Boolean variable.
If TRUE, digital filters are applied. \code{order_rt} and \code{length_rt} are positive integers that control
the polynomial order and the filter length in retention time, respectively. The analogous control
variables for drift time filtering are \code{order_dt} and \code{length_dt}.}
\item{\verb{decimate$do, decimate$dfactor_rt, decimate$factor_dt}}{
Set of variables that control data decimation. \code{do} is a Boolean variable is a Boolean variable.
If TRUE, decimation is performed on data. \code{factor_rt} and \code{factor_dt} are positive integers that
control the decimation factors in retention and drift times, respectively.}
}
}
\note{
Note that filter length must be an odd number bigger than the
polynomial order of the filter.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3,7)

# Example of digital smoothing and decimation, in both axes.
# To show the effect data preparation two EIC are compared:

# Before:
gcims_plot_chrom(dir_in, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

# Data preparation:
alignment_data <- gcims_prepare_data(
  dir_in = dir_in,
  dir_out = dir_out,
  samples = samples,
  params = list(
    filter = list(
      do = TRUE,
      order_rt = 2,
      length_rt = 19,
      order_dt = 2,
      length_dt = 19
    ),
    decimate = list(
      do = TRUE,
      factor_rt = 2,
      factor_dt = 2
    )
  )
)

# After:
gcims_plot_chrom(dir_out, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
}

}
\references{
{García, S., Luengo, J. and Herrera, F., 2015. Data preprocessing
in data mining (Vol. 72, pp. 59-139). Cham, Switzerland: Springer International
Publishing.
\doi{10.1007/978-3-319-10247-4}
}
}
\concept{Data preparation functions}
