% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_mat}
\alias{gcims_read_mat}
\title{Read GC-IMS Samples from MATLAB files}
\usage{
gcims_read_mat(dir_in, dir_out)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}
}
\value{
An invisible NULL. This function is called to convert mat files to rds files.
}
\description{
Read GC-IMS Samples from MATLAB files
}
\examples{
# We can create a mat file:
mat_samples_dir <- tempfile("mat_samples")
dir.create(mat_samples_dir)
matfile <- file.path(mat_samples_dir, "sample.mat")

obj <- list(
  metadata = list(Name = 1),
  data = list(
    drift_time = 1:3,
    retention_time = 1:4,
    data_df = matrix(1:12, nrow = 3, ncol = 4)
  )
)
R.matlab::writeMat(matfile, dd = obj)
dir_out <- tempfile("rds_samples")
dir.create(dir_out)
gcims_read_mat(mat_samples_dir, dir_out)
list.files(dir_out, full.names = TRUE)
}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_metadata}()}
}
\concept{Reading functions}
