% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_mat}
\alias{gcims_read_mat}
\title{Read GC-IMS Samples from matlab}
\usage{
gcims_read_mat(dir_in, dir_out)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}
}
\value{
An R object that contains the matrix and the retention times
}
\description{
Read GC-IMS Samples from matlab
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_metadata}()}
}
\concept{Reading functions}
