% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_metadata}
\alias{gcims_read_metadata}
\title{Read metadata form a csv}
\usage{
gcims_read_metadata(dir_in, samples, file)
}
\arguments{
\item{dir_in}{The input directory.}

\item{samples}{The set of samples to be processed.}

\item{file}{Name of the file that contains the metadata. It must be an excel file}
}
\value{
An R object that contains the metadata information
}
\description{
Read metadata form a csv
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata","add_meta", package = "GCIMS")
setwd(dir_in)
M1 <- readRDS("M1.rds")
samples <- 1
print(M1$metadata)
file <- "Metadata.xlsx"
gcims_read_metadata(dir_in, samples, file)
M1 <- readRDS("M1.rds")
print(M1$metadata)
M1$metadata$Class <- NULL
M1$metadata$Bottle <- NULL
M1$metadata$Brand <- NULL
print(M1$metadata)
saveRDS(M1, "M1.rds")
setwd(current_dir)

}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_mat}()}
}
\concept{Reading functions}
