% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{gcims_read_samples}
\alias{gcims_read_samples}
\title{Read GC-IMS Samples from .csv files}
\usage{
gcims_read_samples(dir_in, dir_out, file, skip = 0)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{file}{Name of the metadata file.}

\item{skip}{Number of lines to skip before reading gcims data.}
}
\value{
An R object that contains the matrix and the retention times
}
\description{
Read GC-IMS Samples from .csv files
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Reading functions: 
\code{\link{gcims_read_mat}()}
}
\concept{Reading functions}
