% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_baseline_removal.R
\name{gcims_remove_baseline}
\alias{gcims_remove_baseline}
\title{Remove baseline of a dataset}
\usage{
gcims_remove_baseline(dir_in, dir_out, samples, peak_list)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are
loaded from.}

\item{dir_out}{Output directory. Where data files after baseline
correction are stored.}

\item{samples}{A vector. Set of samples to which remove the baseline
(e.g.: c(1, 2, 3)).}

\item{peak_list}{A peak list after containing information regarding the 2d peaks found in data.}
}
\value{
A set of S3 objects.
}
\description{
Remove baseline of a dataset
}
\details{
\code{gcims_remove_baseline} performs baseline correction on sample
spectra (in drift time) and chromatograms (in retention time) .This function corrects
the baseline in drift time and afterwards, in retention time. To do it, it cuts the
axis in regions of equal length and looks for the local minima of each region. A simple
linear interpolation is performed to create a baseline of the same length as the number
of data points per axis.The region width is estimated differently in each of the the axes:
In drift time, it is computed as 12 times the standard deviation of the RIP width, while
in retention time as the third quantile of peak width distributions in chromatograms (obtained
from the peak table). Take into account that when removing the baseline in retention
time, the shape of the Reactant Ion Peak is severely affected.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
peak_list <- readRDS(file.path(dir_in, "peak_list.rds"))
dir_out <- tempdir()
samples <- sample_num <- 3

# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = c(155, 175), dt_range = c(8.75, 9.1), transform = FALSE)

# Example of baseline removal:
gcims_remove_baseline(dir_in, dir_out, samples, peak_list)

# After:
gcims_view_sample(dir_out, sample_num = samples, rt_range = c(155, 175), dt_range = c(8.75, 9.1), transform = FALSE)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
}

}
\concept{Baseline removal functions}
