% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_remove_rip}
\alias{gcims_remove_rip}
\title{Removes RIP from all samples}
\usage{
gcims_remove_rip(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}
}
\value{
A RIP removed dataset
}
\description{
Removes RIP from all samples
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

}
}
\seealso{
Other Utility functions: 
\code{\link{gcims_interpolate}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
