% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_remove_rip}
\alias{gcims_remove_rip}
\title{Removes the Reactant Ion Peak (RIP) from samples}
\usage{
gcims_remove_rip(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where RIP removed data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which their RIP has to be removed.}
}
\value{
A Set of S3 objects.
}
\description{
Removes the Reactant Ion Peak (RIP) from samples
}
\details{
\code{gcims_remove_rip} substitutes the RIP by its corresponding
linear approximation to the RIP baseline, for every spectrum in a sample.
This process is repeated for all samples in \code{samples}. Use this
function if you are interested in enhancing the contrast of peaks of sample
images / chromatograms / spectra to be obtained from
\code{gcims_view_sample} / \code{gcims_plot_chrom} /
\code{gcims_plot_spec}.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Reactant Ion Peak removal
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
gcims_remove_rip(dir_in, dir_out, samples)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\seealso{
Other Utility functions: 
\code{\link{gcims_compute_fom}()},
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_shift_rt}()}
}
\concept{Utility functions}
