% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_reshape_samples}
\alias{gcims_reshape_samples}
\title{Reshape samples}
\usage{
gcims_reshape_samples(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where reshaped samples are stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to be reshaped.}
}
\value{
A set of S3 objects.
}
\description{
Reshape samples
}
\details{
\code{gcims_reshape_samples} ensures that all samples in a dataset
have the same dimensions (number of data points) in retention and drift
times. \code{gcims_reshape_samples} checks what are the minimum retention /
drift time ranges a cuts all samples according to these ranges.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- c(3, 7, 8, 14, 20, 21, 22)

# Example of reshaping samples
# (all samples must have the same
# retention and drift time dimensions).

# Before:
nrow_before <- matrix(0, nrow = length(samples), ncol = 1)
ncol_before <- matrix(0, nrow = length(samples), ncol = 1)
for (i in seq_along(samples)){
  aux_string <- paste0("M", samples[i], ".rds")
  aux_list <- readRDS(file.path(dir_in, aux_string))
  nrow_before[i, 1] <- nrow(as.matrix(aux_list$data$data_df))
  ncol_before[i, 1] <- ncol(as.matrix(aux_list$data$data_df))
}

# After:
gcims_reshape_samples(dir_in, dir_out, samples)
nrow_after <- matrix(0, nrow = length(samples), ncol = 1)
ncol_after <- matrix(0, nrow = length(samples), ncol = 1)
for (i in seq_along(samples)){
  aux_string <- paste0("M", samples[i], ".rds")
  aux_list <- readRDS(file.path(dir_out, aux_string))
  nrow_after[i, 1] <- nrow(as.matrix(aux_list$data$data_df))
  ncol_after[i, 1] <- ncol(as.matrix(aux_list$data$data_df))
}

reshaping_info <- cbind(nrow_before, ncol_before, nrow_after, ncol_after)
colnames(reshaping_info) <- c("rows_before", "columns_before", "rows_after", "columns_after")
rownames(reshaping_info) <- c("M3", "M7", "M8", "M14", "M20", "M21", "M22")

print(reshaping_info)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
