% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakDetection.R
\name{gcims_rois_selection}
\alias{gcims_rois_selection}
\title{ROIs Selection}
\usage{
gcims_rois_selection(dir_in, dir_out, samples, noise_level)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where RIP removed data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which their RIP has to be removed.}

\item{noise_level}{Scalar number. The number of times the standard deviation
above the noise level needed to detect a peak.}
}
\value{
A Set of S3 objects.
}
\description{
ROIs Selection
}
\details{
\code{gcims_remove_rip} substitutes the RIP by its corresponding
linear approximation to the RIP baseline, for every spectrum in a sample.
This process is repeated for all samples in \code{samples}. Use this
function if you are interested in enhancing the contrast of peaks of sample
images / chromatograms / spectra to be obtained from
\link{gcims_view_sample}, \link{gcims_plot_chrom} or \link{gcims_plot_spec}.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Reactant Ion Peak removal
# Before:
gcims_view_sample(dir_in, sample_num = samples, rt_range = NULL, dt_range = NULL)

# After:
gcims_remove_rip(dir_in, dir_out, samples)
gcims_view_sample(dir_out, sample_num = samples, rt_range = NULL, dt_range = NULL)

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)
}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_shifting}()},
\code{\link{gcims_unfold}()}
}
\concept{Utility functions}
