% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{gcims_smoothing}
\alias{gcims_smoothing}
\title{Data Smoothing using Savitzky-Golay filters}
\usage{
gcims_smoothing(dir_in, dir_out, samples, by_rows, filter_length,
  polynomial_order)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{by_rows}{Logical. Direction to apply the function. If TRUE it by rows (drift time direction).
If FALSE, applied by columns (that is the retention time direction).}

\item{filter_length}{Numerical. Length of the filter.}

\item{polynomial_order}{Numerical. Order of the polynomial.}
}
\value{
A filtered  gcims dataset.
}
\description{
Data Smoothing using Savitzky-Golay filters
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Smoothing functions}
