% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{gcims_smoothing}
\alias{gcims_smoothing}
\title{Data Smoothing using Savitzky-Golay filters}
\usage{
gcims_smoothing(
  dir_in,
  dir_out,
  samples,
  time,
  filter_length = 19,
  polynomial_order = 2
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where smoothed data files are
stored.}

\item{samples}{A vector. Set of samples to be filtered (e.g.: c(1,
2, 3)).}

\item{time}{Sets the dimension to be corrected: drift time or
retention time. Introduce "Retention" for retention time; or "Drift" for
drift time.}

\item{filter_length}{Numerical. Length of the filter.}

\item{polynomial_order}{Numerical. Order of the polynomial.}
}
\value{
A set of S3 objects.
}
\description{
Data Smoothing using Savitzky-Golay filters
}
\details{
\code{gcims_smoothing} performs digital smoothing on sample
chromatograms (in retention time) or spectra (in drift time). This
smoothing process is required to reduce the noise level of data samples.
\code{gcims_smoothing} relies in Savitzky-Golay filters to reduce noise
contribution in data. In a nutshell, Savitzky-Golay filters reconstruct
data by fitting successive sets of data points of given length to
low-degree polynomials. Short \emph{filter_length} values don't reduce
noise significantly, while large ones may distort signal shape. Similarly,
polynomials with low \emph{polynomial_order} may not be flexible enough to
properly reconstruct the signal, while polynomials with high
\emph{polynomial_order} can be too flexible and model noise on the
reconstructed signal. A trade-off for these parameter values should be
found.
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Retention time smoothing:
# Before:
gcims_plot_chrom(dir_in, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

# After:
gcims_smoothing(dir_in, dir_out, samples, time = "Retention")
gcims_plot_chrom(dir_out, samples, dt_value = 8.5,  rt_range = NULL, colorby = "Class")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))
setwd(current_dir)

}
\references{
{ Savitzky, A. and Golay, M.J.E. (1964) Smoothing and
Differentiation of Data by Simplified Least Squares Procedures. Analytical
Chemistry, 36, 1627–39. https://doi.org/10.1021/ac60214a047.

Schafer, R.W. (2011) What is a savitzky-golay filter? IEEE Signal Processing
Magazine, Institute of Electrical and Electronics Engineers Inc. 28, 111–7.
https://doi.org/10.1109/MSP.2011.941097. }
}
\concept{Smoothing functions}
