% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{gcims_smoothing}
\alias{gcims_smoothing}
\title{Data Smoothing using Savitzky-Golay filters}
\usage{
gcims_smoothing(
  dir_in,
  dir_out,
  samples,
  time,
  filter_length = 19,
  polynomial_order = 2
)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}

\item{filter_length}{Numerical. Length of the filter.}

\item{polynomial_order}{Numerical. Order of the polynomial.}
}
\value{
A filtered  gcims dataset.
}
\description{
Data Smoothing using Savitzky-Golay filters
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\concept{Smoothing functions}
