% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_unfold}
\alias{gcims_unfold}
\title{Converts Sample Matrices to Feature Vectors}
\usage{
gcims_unfold(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{dir_out}{Output directory. Where unfold data files are
stored.}

\item{samples}{Numeric vector of integers. Identifies the set of
samples to which their data matrices need to be unfolded.}
}
\value{
A set of S3 objects
}
\description{
Converts Sample Matrices to Feature Vectors
}
\details{
\code{gcims_unfold} reshapes \emph{n} x \emph{m} GCIMS data into
\emph{1} x (\emph{n} x \emph{m}) vectors. This is done for all samples in
\code{samples}. \code{gcims_unfold} allows the latter use of pattern
recognition techniques (e.g. Principal Component Analysis) that need to be
fed with tabular data.
}
\note{
It is recommended to perform a previous reduction of data
dimensionality using the functions \code{gcims_decimate} and
\code{gcims_cut_samples} before applying \code{gcims_unfold} to the data.
}
\examples{
dir_in <- system.file("extdata", "to_interpolate", package = "GCIMS")
dir_out <- tempdir()
samples <- 3

# Example of Sample Data unfolding
# Before:
M3 <- readRDS(file.path(dir_in, "M3.rds"))
data <- M3$data$data_df
data_dimensions <- dim(data)

message("Before unfolding data has ", data_dimensions[1],
" rows and ", data_dimensions[2], " columns.")

# After:
gcims_unfold(dir_in, dir_out, samples)
M3 <- readRDS(file.path(dir_out, "M3.rds"))
data <- M3$data$data_df
data_length <- length(data)

message("After unfolding data is a vector of length ", data_length,".
This value corresponds to product of the number
of columns by the number of rows of the original data.")

files <- list.files(path = dir_out, pattern = ".rds", all.files = FALSE, full.names = TRUE)
invisible(file.remove(files))

}
\seealso{
Other Utility functions: 
\code{\link{gcims_cut_samples}()},
\code{\link{gcims_decimate}()},
\code{\link{gcims_figures_of_merit}()},
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{gcims_reshape_samples}()},
\code{\link{gcims_rois_fusion}()},
\code{\link{gcims_rois_selection}()},
\code{\link{gcims_shifting}()}
}
\concept{Utility functions}
