% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gcims_unfold}
\alias{gcims_unfold}
\title{Converts Sample Matrices to Feature Vectors}
\usage{
gcims_unfold(dir_in, dir_out, samples)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}
}
\value{
A set of feature vectors
}
\description{
Converts Sample Matrices to Feature Vectors
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Utility functions: 
\code{\link{gcims_interpolate}()},
\code{\link{gcims_remove_rip}()},
\code{\link{reshape_samples}()}
}
\concept{Utility functions}
