% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_view_sample}
\alias{gcims_view_sample}
\title{Sample matrix visualization}
\usage{
gcims_view_sample(dir_in, sample_num, rt_range = NULL, dt_range = NULL)
}
\arguments{
\item{dir_in}{The input directory.}

\item{sample_num}{The number corresponding to the sample
to be visualized.}

\item{rt_range}{Min a Max retention time values. NULL by
default.}

\item{dt_range}{Min a Max drift time values. NULL by
default.}
}
\value{
An image of the sample matrix.
}
\description{
Sample matrix visualization
}
\examples{
current_dir <- getwd()
dir_in <- system.file("extdata", package = "GCIMS")
sample_num <- 3
gcims_view_sample(dir_in, sample_num, rt_range = NULL, dt_range = NULL)
setwd(current_dir)

}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_plot_spec}()}
}
\concept{Visualization functions}
