% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx-deprecated-gcims_visualization.R
\name{gcims_view_sample}
\alias{gcims_view_sample}
\title{Sample matrix visualization}
\usage{
gcims_view_sample(
  dir_in,
  sample_num,
  rt_range = NULL,
  dt_range = NULL,
  transform = FALSE
)
}
\arguments{
\item{dir_in}{Input directory. Where input data files are loaded
from.}

\item{sample_num}{Numeric. Identifies the sample to be visualized from
the dataset.}

\item{rt_range}{Min and Max retention time values. If NULL the
complete retention time range is used. NULL by default.}

\item{dt_range}{Min and Max drift time values. If NULL the complete
drift time range is used. NULL by default.}

\item{transform}{Boolean. IF TRUE, performs the cubic root of data to
enhance image contrast. FALSE by default.}
}
\value{
An image of selected data sample.
}
\description{
Visualize 2-D gcims data.
}
\details{
\code{gcims_view_sample()} represents gcims data as a raster. In
this plot, retention time increases along the y-axis from down to up, while
drift time does it along the x-axis left to right. \code{gcims_view_sample()} provides to the user
qualitative information about what are regions of interest to be analyzed in a sample.
}
\note{
\code{gcims_view_sample()} can't provide a reliable visualization of a
data sample if its sampling frequencies  along drift and/or retention time
are not constants. To overcome this problem consider using the function
\code{gcims_prepare_samples()}to interpolate data.
}
\examples{
dir_in <- system.file("extdata", package = "GCIMS")
sample_num <- 3
gcims_view_sample(dir_in, sample_num, rt_range = NULL, dt_range = NULL)

}
\references{
{
Wickham, Hadley. "ggplot2." Wiley Interdisciplinary Reviews: Computational Statistics 3.2 (2011): 180-185.
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_chrom}()},
\code{\link{gcims_plot_spec}()}
}
\concept{Visualization functions}
