% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcims_visualization.R
\name{gcims_visualization}
\alias{gcims_visualization}
\title{Sample matrix visualization}
\usage{
gcims_visualization(dir_in, sample_num)
}
\arguments{
\item{dir_in}{The input directory.}

\item{sample_num}{The number corresponding to the sample
to be visualized.}
}
\value{
An image of the sample matrix.
}
\description{
Sample matrix visualization
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Visualization functions: 
\code{\link{gcims_plot_eics}()},
\code{\link{gcims_plot_spec}()},
\code{\link{gcims_plot_tics}()},
\code{\link{gcims_plot_tisa}()}
}
\concept{Visualization functions}
