% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSChromatogram-aaa-class.R
\docType{class}
\name{GCIMSChromatogram-class}
\alias{GCIMSChromatogram-class}
\alias{GCIMSChromatogram}
\title{GCIMSChromatogram class}
\usage{
GCIMSChromatogram(...)
}
\arguments{
\item{...}{See the slots section in this page}
}
\value{
A GCIMSChromatogram object
}
\description{
GCIMSChromatogram is an S4 class to store a GCIMS Chromatogram It can be a single chromatogram
or the aggregation of several chromatograms.
}
\section{Functions}{
\itemize{
\item \code{GCIMSChromatogram()}: class

}}
\section{Slots}{

\describe{
\item{\code{retention_time}}{A numeric vector with retention times}

\item{\code{intensity}}{A numeric vector with the corresponding intensities}

\item{\code{drift_time_idx}}{The index or indices used to get the intensity}

\item{\code{drift_time_ms}}{The drift times corresponding to \code{drift_time_idx}.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}
}}

