% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSChromatogram-.R,
%   R/description-GCIMSChromatogram.R, R/peaks-GCIMSChromatogram.R,
%   R/plot-GCIMSChromatogram.R
\docType{class}
\name{GCIMSChromatogram-class}
\alias{GCIMSChromatogram-class}
\alias{GCIMSChromatogram}
\alias{description,GCIMSChromatogram-method}
\alias{description<-,GCIMSChromatogram,ANY-method}
\alias{peaks,GCIMSChromatogram-method}
\alias{peaks<-,GCIMSChromatogram-method}
\alias{plot,GCIMSChromatogram,ANY-method}
\title{GCIMSChromatogram class}
\usage{
GCIMSChromatogram(...)

\S4method{description}{GCIMSChromatogram}(object)

\S4method{description}{GCIMSChromatogram,ANY}(object) <- value

\S4method{peaks}{GCIMSChromatogram}(object)

\S4method{peaks}{GCIMSChromatogram}(object) <- value

\S4method{plot}{GCIMSChromatogram,ANY}(x, y, ...)
}
\arguments{
\item{...}{See the slots section below}

\item{object}{A GCIMSChromatogram object}

\item{value}{A data frame with the peak list}

\item{x}{A GCIMSChromatogram object to plot}

\item{y}{unused}
}
\value{
A GCIMSChromatogram object
}
\description{
GCIMSChromatogram is an S4 class to store a GCIMS Chromatogram It can be a single chromatogram
or the aggregation of several chromatograms.
}
\section{Functions}{
\itemize{
\item \code{GCIMSChromatogram()}: Friendly constructor

\item \code{description(GCIMSChromatogram)}: Get the description

\item \code{description(object = GCIMSChromatogram) <- value}: Set the description

\item \code{peaks(GCIMSChromatogram)}: Get the peak list

\item \code{peaks(GCIMSChromatogram) <- value}: Set the peak list

\item \code{plot(x = GCIMSChromatogram, y = ANY)}: plot method

}}
\section{Slots}{

\describe{
\item{\code{retention_time}}{A numeric vector with retention times}

\item{\code{intensity}}{A numeric vector with the corresponding intensities}

\item{\code{baseline}}{A numeric vector of the same length as \code{intensity} with the corresponding baseline,
or \code{NULL} if not set. Use \code{\link[=estimateBaseline]{estimateBaseline()}} to estimate it, \code{\link[=baseline]{baseline()}} to directly access it.}

\item{\code{drift_time_idx}}{The index or indices used to get the intensity}

\item{\code{drift_time_ms}}{The drift times corresponding to \code{drift_time_idx}.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}

\item{\code{peaks}}{A data frame with peaks, use \code{\link[=findPeaks]{findPeaks()}} to fill it, or \code{\link[=peaks]{peaks()}} to set/get it}

\item{\code{peaks_debug_info}}{A list with arbitrary debug infor from \code{\link[=findPeaks]{findPeaks()}}}
}}

