% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSChromatogram-aaa-class.R,
%   R/GCIMSChromatogram-description.R
\docType{class}
\name{GCIMSChromatogram-class}
\alias{GCIMSChromatogram-class}
\alias{GCIMSChromatogram}
\alias{description,GCIMSChromatogram-method}
\alias{description<-,GCIMSChromatogram,ANY-method}
\title{GCIMSChromatogram class}
\usage{
GCIMSChromatogram(...)

\S4method{description}{GCIMSChromatogram}(object)

\S4method{description}{GCIMSChromatogram,ANY}(object) <- value
}
\arguments{
\item{...}{See the slots section below}

\item{object}{A GCIMSChromatogram object}

\item{value}{A string with the description}
}
\value{
A GCIMSChromatogram object
}
\description{
GCIMSChromatogram is an S4 class to store a GCIMS Chromatogram It can be a single chromatogram
or the aggregation of several chromatograms.
}
\section{Functions}{
\itemize{
\item \code{GCIMSChromatogram()}: Friendly constructor

\item \code{description(GCIMSChromatogram)}: Get the description

\item \code{description(object = GCIMSChromatogram) <- value}: Set the description

}}
\section{Slots}{

\describe{
\item{\code{retention_time}}{A numeric vector with retention times}

\item{\code{intensity}}{A numeric vector with the corresponding intensities}

\item{\code{baseline}}{A numeric vector of the same length as \code{intensity} with the corresponding baseline.
Or \code{NULL} if not set. Use \code{\link[=estimateBaseline]{estimateBaseline()}} to estimate it, \code{\link[=baseline]{baseline()}} to directly access it.}

\item{\code{drift_time_idx}}{The index or indices used to get the intensity}

\item{\code{drift_time_ms}}{The drift times corresponding to \code{drift_time_idx}.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}
}}

