% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSDataset.R,
%   R/getGCIMSSample-GCIMSDataset.R, R/getTIS_getRIC-GCIMSDataset.R,
%   R/pData-GCIMSDataset.R, R/peaks-GCIMSDataset.R
\docType{class}
\name{GCIMSDataset-class}
\alias{GCIMSDataset-class}
\alias{GCIMSDataset}
\alias{getSample,GCIMSDataset-method}
\alias{getTIS,GCIMSDataset-method}
\alias{plotRIC,GCIMSDataset-method}
\alias{pData,GCIMSDataset-method}
\alias{pData<-,GCIMSDataset,ANY-method}
\alias{peaks,GCIMSDataset-method}
\alias{peaks<-,GCIMSDataset-method}
\title{GCIMSDataset class}
\usage{
GCIMSDataset(
  pData,
  base_dir,
  scratch_dir = tempfile("GCIMSDataset_tempdir_"),
  keep_intermediate = FALSE,
  on_ram = FALSE
)

\S4method{getSample}{GCIMSDataset}(object, sample)

\S4method{getTIS}{GCIMSDataset}(object)

\S4method{plotRIC}{GCIMSDataset}(object, rt_range = NULL, sample = NULL)

\S4method{pData}{GCIMSDataset}(object)

\S4method{pData}{GCIMSDataset,ANY}(object) <- value

\S4method{peaks}{GCIMSDataset}(object)

\S4method{peaks}{GCIMSDataset}(object) <- value
}
\arguments{
\item{pData}{A data frame with at least the \code{SampleID} and \code{FileName} columns.}

\item{base_dir}{A directory containing the file names described in \code{pData}}

\item{scratch_dir}{A directory to save intermediate results.}

\item{keep_intermediate}{A logical. Whether to keep or not intermediate files when realizing the GCIMSDataset object}

\item{on_ram}{A logical. If \code{TRUE}, samples are kept on RAM. This is faster
as long as you have enough memory to keep all samples. If this is \code{TRUE}, then
\code{scratch_dir} and \code{keep_intermediate} are ignored.}

\item{object}{A \link{GCIMSDataset} object}

\item{sample}{A number or a string with the sample index or name. If \code{NULL}, all samples are returned}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{value}{The data frame with a peak list}
}
\value{
A GCIMSDataset object

The corresponding \link{GCIMSSample}

A matrix with samples in rows and the drift time in columns

A plot
}
\description{
GCIMSDataset is an S4 class to store a dataset
}
\details{
The actual data is not stored in memory, but read/saved from/to files as
needed, so the dataset object scales with large number of samples.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getSample(GCIMSDataset)}: class

\item \code{getTIS(GCIMSDataset)}: Get Total Ion Spectra matrix

\item \code{plotRIC(GCIMSDataset)}: Plot Reverse Ion Chromatograms

}}
\section{Functions}{
\itemize{
\item \code{GCIMSDataset()}: Constructor method

\item \code{pData(GCIMSDataset)}: Get the phenotype data

\item \code{pData(object = GCIMSDataset) <- value}: Set the phenotype data

\item \code{peaks(GCIMSDataset)}: Get the peak list

\item \code{peaks(GCIMSDataset) <- value}: Set the peak list

}}
\examples{
# Create a new GCIMSDataset with methods::new()
dummy_obj <- methods::new(
  "GCIMSDataset",
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
# Create a new GCIMSDataset with the convenient constructor function:
dummy_obj <-GCIMSDataset(
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
}
