% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-aaa-class.R,
%   R/GCIMSDataset-getGCIMSSample.R, R/GCIMSDataset-getTIS_getRIC.R
\docType{class}
\name{GCIMSDataset-class}
\alias{GCIMSDataset-class}
\alias{GCIMSDataset}
\alias{getGCIMSSample,GCIMSDataset-method}
\alias{getTIS,GCIMSDataset-method}
\alias{plotRIC,GCIMSDataset-method}
\title{GCIMSDataset class}
\usage{
GCIMSDataset(
  pData,
  base_dir,
  scratch_dir = tempfile("GCIMSDataset_tempdir_"),
  keep_intermediate = FALSE
)

\S4method{getGCIMSSample}{GCIMSDataset}(object, sample)

\S4method{getTIS}{GCIMSDataset}(object)

\S4method{plotRIC}{GCIMSDataset}(object)
}
\arguments{
\item{pData}{A data frame with at least the \code{SampleID} and \code{FileName} columns.}

\item{base_dir}{A directory containing the file names described in \code{pData}}

\item{scratch_dir}{A directory to save intermediate results.}

\item{keep_intermediate}{A logical. Whether to keep or not intermediate files when realizing the GCIMSDataset object}

\item{object}{A \link{GCIMSDataset} object}

\item{sample}{Either an integer (sample index) or a string (sample name)}
}
\value{
A GCIMSDataset object

The corresponding \link{GCIMSSample}

A matrix with samples in rows and the drift time in columns

A plot
}
\description{
GCIMSDataset is an S4 class to store a dataset
}
\details{
The actual data is not stored in memory, but read/saved from/to files as
needed, so the dataset object scales with large number of samples.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getGCIMSSample(GCIMSDataset)}: class

\item \code{getTIS(GCIMSDataset)}: Get Total Ion Spectra matrix

\item \code{plotRIC(GCIMSDataset)}: Plot Reverse Ion Chromatograms

}}
\section{Functions}{
\itemize{
\item \code{GCIMSDataset()}: Constructor method

}}
\examples{
# Create a new GCIMSDataset with methods::new()
dummy_obj <- methods::new(
  "GCIMSDataset",
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
# Create a new GCIMSDataset with the convenient constructor function:
dummy_obj <-GCIMSDataset(
  pData = data.frame(SampleID = character(), filename = character(0)),
  base_dir = tempdir()
)
}
