% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDelayedOp-class.R
\docType{class}
\name{GCIMSDelayedOp-class}
\alias{GCIMSDelayedOp-class}
\alias{GCIMSDelayedOp}
\title{GCIMSDelayedOp class}
\usage{
GCIMSDelayedOp(...)
}
\arguments{
\item{...}{See the slots section in this page}
}
\value{
A GCIMSDelayedOp object
}
\description{
GCIMSDelayedOp is an S4 class to store a delayed operation

Delayed operations are not applied to the dataset immediately, but rather
when some data from the dataset is required. Delaying the calculation has
the advantage that some operations that are applied to each sample can be applied
sequentially to each sample, reducing the amount of reads and writes to
the data files, and increasing the performance.
}
\section{Functions}{
\itemize{
\item \code{GCIMSDelayedOp()}: class

}}
\section{Slots}{

\describe{
\item{\code{name}}{A named for de delayed operation, only used for printing.}

\item{\code{fun}}{A function that takes a \link{GCIMSSample} and returns a \link{GCIMSSample} (modified)}

\item{\code{params}}{A named list with additional arguments to be passed to function}

\item{\code{fun_extract}}{A function that takes a modified \link{GCIMSSample} and returns an extracted object.}

\item{\code{fun_aggregate}}{A function that takes a \link{GCIMSDataset} and a list of extracted objects and returns a modified \link{GCIMSDataset}.}
}}

