% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{GCIMSSample-class}
\alias{GCIMSSample-class}
\title{GCIMSSample class}
\value{
A GCIMSSample object
}
\description{
GCIMS Sample is an S4 class to store one sample
with the drift and retention time ranges and other relevant attributes
(GC column, drift tube length...) if available
}
\details{
The actual spectra is stored in the `data` slot, in a matrix,
where the first index (rows) corresponds to drift times and the
second to retention times (columns).
}
\section{Slots}{

\describe{
\item{\code{drift_time}}{numeric. (required)}

\item{\code{retention_time}}{numeric. (required)}

\item{\code{data}}{matrix A matrix with drift time in the rows and retention time
in columns. (required)}

\item{\code{gc_column}}{character. (optional) The type of chromatographic column used}

\item{\code{drift_tube_length}}{numeric (optional) The length of the drift tube, in mm}

\item{\code{params}}{list (optional) Arbitrary list of parameters and annotations}
}}

\examples{
dummy_obj <-methods::new(
  "GCIMSSample",
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_tube_length = 98.0 # in mm
)
}
