% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample-class.R, R/GCIMSSample-plotRaw.R
\docType{class}
\name{GCIMSSample-class}
\alias{GCIMSSample-class}
\alias{GCIMSSample}
\alias{plotRaw,GCIMSSample-method}
\title{GCIMSSample class}
\usage{
GCIMSSample(...)

\S4method{plotRaw}{GCIMSSample}(object, dt_range = NULL, rt_range = NULL)
}
\arguments{
\item{...}{See the slots section in this page}

\item{x}{A GCIMSSample object}
}
\value{
A GCIMSSample object

A plot of the GCIMSSample
}
\description{
GCIMS Sample is an S4 class to store one sample
with the drift and retention time ranges and other relevant attributes
(GC column, drift tube length...) if available

The actual spectra is stored in the \code{data} slot, in a matrix,
where the first index (rows) corresponds to drift times and the
second to retention times (columns).
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotRaw(GCIMSSample)}: Topographical plot of a GC-IMS Sample

}}
\section{Functions}{
\itemize{
\item \code{GCIMSSample()}: Create a GCIMSSample object

}}
\section{Slots}{

\describe{
\item{\code{drift_time}}{numeric. (required)}

\item{\code{retention_time}}{numeric. (required)}

\item{\code{data}}{matrix A matrix with drift time in the rows and retention time
in columns. (required)}

\item{\code{gc_column}}{character. (optional) The type of chromatographic column used}

\item{\code{drift_tube_length}}{numeric (optional) The length of the drift tube, in mm}

\item{\code{drift_gas}}{character. (optional) The drift gas used (e.g "nitrogen")}

\item{\code{params}}{list (optional) Arbitrary list of parameters and annotations}

\item{\code{history}}{character. A character vector with a summary of information of the
processing details the sample has gone through already}

\item{\code{filepath}}{character. A string with the path to the raw data}

\item{\code{description}}{A string (optional). A sample name or ID or description used in plots}

\item{\code{class_version}}{"numeric_version" (internal) The GCIMSSample object defines
internally a class version, so if a GCIMSSample object is saved, the GCIMS
package is updated and the GCIMSSample class has changed during the upgrade
it may be possible to upgrade the previously saved object when it's loaded.}
}}

\examples{
# Create a new GCIMSSample with methods::new()
dummy_obj <-methods::new(
  "GCIMSSample",
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
# Create a new GCIMSSample with the convenient constructor function:
dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
plotRaw(dummy_obj)
}
\seealso{
\link{GCIMSSample-methods}
}
