% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GCIMSSample-methods}
\alias{GCIMSSample-methods}
\alias{[.GCIMS}
\alias{subset.GCIMSSample}
\alias{subset}
\title{Methods for the GCIMSSample class}
\usage{
\method{[}{GCIMS}(x, i, j, ...)

\method{subset}{GCIMSSample}(x, dt_idx = NULL, rt_idx = NULL, ...)
}
\arguments{
\item{x}{A GCIMSSample object}

\item{i}{index for drift time to subset}

\item{j}{index for retention time to subset}

\item{...}{ignored}

\item{dt_idx}{A vector of drift time indices to keep}

\item{rt_idx}{A vector of retention time indices to keep}
}
\value{
\code{[}: object \code{x} with features \code{i} and cells \code{j}

\code{subset}: A subsetted \code{GCIMSSample} object
}
\description{
Methods for the GCIMSSample class
}
\section{Functions}{
\itemize{
\item \code{[.GCIMS}: Simple subsetter for \code{GCIMSSample} objects

\item \code{subset.GCIMSSample}: Subset a \code{\link[=GCIMSSample-class]{GCIMSSample-class()}} object
}}

\examples{
# `[' examples

}
\seealso{
\code{\link[base:subset]{base::subset()}}
}
