% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSSample.R
\name{GCIMSSample-methods}
\alias{GCIMSSample-methods}
\alias{dtime,GCIMSSample-method}
\alias{rtime,GCIMSSample-method}
\alias{intensity,GCIMSSample-method}
\alias{[.GCIMSSample}
\alias{dim.GCIMSSample}
\alias{subset.GCIMSSample}
\alias{subset}
\alias{getEIC,GCIMSSample-method}
\alias{getIMSScan,GCIMSSample-method}
\title{Methods for the GCIMSSample class}
\usage{
\S4method{dtime}{GCIMSSample}(object)

\S4method{rtime}{GCIMSSample}(object)

\S4method{intensity}{GCIMSSample}(
  object,
  dt_range = NULL,
  rt_range = NULL,
  dt_idx = NULL,
  rt_idx = NULL
)

\method{[}{GCIMSSample}(x, i, j, ...)

\method{[}{GCIMSSample}(x, i, j, ...)

\method{dim}{GCIMSSample}(x)

\method{subset}{GCIMSSample}(x, dt_idx = NULL, rt_idx = NULL, dt_range = NULL, rt_range = NULL, ...)

\S4method{getEIC}{GCIMSSample}(object, dt_range = NULL, rt_range = NULL)

\S4method{getIMSScan}{GCIMSSample}(object, dt_range = NULL, rt_range = NULL)
}
\arguments{
\item{object}{A GCIMSSample object}

\item{dt_range}{The drift time range to keep (in milliseconds)}

\item{rt_range}{The retention time range to keep (in seconds)}

\item{dt_idx}{A vector of drift time indices to keep}

\item{rt_idx}{A vector of retention time indices to keep}

\item{x}{A GCIMSSample object}

\item{i}{index for drift time to subset}

\item{j}{index for retention time to subset}

\item{...}{ignored}
}
\value{
\code{[}: object \code{x} with features \code{i} and cells \code{j}

\code{[}: object \code{x} with features \code{i} and cells \code{j}

An integer vector with the number of rows and columns of the matrix

\code{subset}: A subsetted \code{GCIMSSample} object
}
\description{
Methods for the GCIMSSample class
}
\section{Functions}{
\itemize{
\item \code{dtime(GCIMSSample)}: Get the drift time vector

\item \code{rtime(GCIMSSample)}: Get the retention time vector

\item \code{intensity(GCIMSSample)}: Get the intensity matrix

\item \code{[}: Simple subsetter for \linkS4class{GCIMSSample} objects

\item \code{[}: Simple subsetter for \linkS4class{GCIMSSample} objects

\item \code{dim(GCIMSSample)}: Dimension of the data matrix

\item \code{subset(GCIMSSample)}: Subset a \linkS4class{GCIMSSample} object

\item \code{getEIC(GCIMSSample)}: Get the extracted ion chromatogram

\item \code{getIMSScan(GCIMSSample)}: Get IMS spectrum

}}
\examples{
mea_file <- system.file("extdata",  "sample_formats", "small.mea.gz", package = "GCIMS")
gcims_sample <- read_mea(mea_file)
my_matrix <- intensity(gcims_sample, dt_range = c(7, 8), rt_range = c(1,30))
# `[' examples

# `[' examples

obj <- GCIMSSample(drift_time=1:2, retention_time=1:3, data = matrix(1:6, nrow=2, ncol=3))
dim(obj)
}
\seealso{
\code{\link[base:subset]{base::subset()}}
}
