% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GCIMSSample}
\alias{GCIMSSample}
\title{Create a GCIMSSample object}
\usage{
GCIMSSample(...)
}
\arguments{
\item{...}{See the slots in \linkS4class{GCIMSSample}}
}
\value{

}
\description{
Create a GCIMSSample object
}
\examples{
dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3),
  gc_column = "Optional column name",
  drift_gas = "nitrogen",
  drift_tube_length = 98.0 # in mm
)
}
