% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-GCIMSSpectrum.R,
%   R/description-GCIMSSpectrum.R, R/dtime-rtime-intensity-GCIMSSpectrum.R,
%   R/peaks-GCIMSSpectrum.R, R/plot-GCIMSSpectrum.R
\docType{class}
\name{GCIMSSpectrum-class}
\alias{GCIMSSpectrum-class}
\alias{GCIMSSpectrum}
\alias{description,GCIMSSpectrum-method}
\alias{description<-,GCIMSSpectrum,ANY-method}
\alias{dtime,GCIMSSpectrum-method}
\alias{rtime,GCIMSSpectrum-method}
\alias{intensity,GCIMSSpectrum-method}
\alias{peaks,GCIMSSpectrum-method}
\alias{peaks<-,GCIMSSpectrum-method}
\alias{plot,GCIMSSpectrum,ANY-method}
\title{GCIMSSpectrum class}
\usage{
GCIMSSpectrum(...)

\S4method{description}{GCIMSSpectrum}(object)

\S4method{description}{GCIMSSpectrum,ANY}(object) <- value

\S4method{dtime}{GCIMSSpectrum}(object)

\S4method{rtime}{GCIMSSpectrum}(object)

\S4method{intensity}{GCIMSSpectrum}(object, dt_range = NULL, dt_idx = NULL)

\S4method{peaks}{GCIMSSpectrum}(object)

\S4method{peaks}{GCIMSSpectrum}(object) <- value

\S4method{plot}{GCIMSSpectrum,ANY}(x, y, ...)
}
\arguments{
\item{...}{See the slots section}

\item{object}{A \link{GCIMSSpectrum} object}

\item{value}{A data frame with the peak list}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{dt_idx}{A numeric vector with the drift time indices to extract (or a logical vector of the length of drift time)}

\item{x}{A GCIMSSpectrum object to plot}

\item{y}{unused}
}
\value{
A GCIMSSpectrum object

A data frame with the peaks in the spectrum

The GCIMSSpectrum object
}
\description{
GCIMSSpectrum is an S4 class to store a GCIMS Spectrum. It can be a single spectrum
or the aggregation of several spectra.
}
\section{Functions}{
\itemize{
\item \code{GCIMSSpectrum()}: Friendly constructor

\item \code{description(GCIMSSpectrum)}: Get the description

\item \code{description(object = GCIMSSpectrum) <- value}: Set the description

\item \code{dtime(GCIMSSpectrum)}: Get the drift time vector

\item \code{rtime(GCIMSSpectrum)}: Get the retention time where this spectrum was extracted

\item \code{intensity(GCIMSSpectrum)}: Get the intensity matrix

\item \code{peaks(GCIMSSpectrum)}: Get the peak list

\item \code{peaks(GCIMSSpectrum) <- value}: Set the peak list

\item \code{plot(x = GCIMSSpectrum, y = ANY)}: plot method

}}
\section{Slots}{

\describe{
\item{\code{drift_time}}{A numeric vector with drift times}

\item{\code{intensity}}{A numeric vector with the corresponding intensities}

\item{\code{baseline}}{A numeric vector of the same length as \code{intensity} with the corresponding baseline,
or \code{NULL} if not set. Use \code{\link[=estimateBaseline]{estimateBaseline()}} to estimate it, \code{\link[=baseline]{baseline()}} to directly access it.}

\item{\code{retention_time_idx}}{The index or indices used to get the intensity}

\item{\code{retention_time_s}}{The retention times corresponding to the retention time indices.}

\item{\code{description}}{A string with a description (used as plot title, useful e.g. to know the sample it came from)}

\item{\code{peaks}}{A data frame with peaks, use \code{\link[=findPeaks]{findPeaks()}} to fill it, or \code{\link[=peaks]{peaks()}} to set/get it}

\item{\code{peaks_debug_info}}{A list with arbitrary debug information from \code{\link[=findPeaks]{findPeaks()}}}
}}

\examples{
spec <- GCIMSSpectrum(drift_time = 1:10, intensity = c(1:5, 6:2))
}
