% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-AllGenerics.R
\name{getEIC}
\alias{getEIC}
\title{Deprecated function. Use getChromatogram() instead}
\usage{
getEIC(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=getChromatogram]{getChromatogram()}}}
}
\description{
Deprecated function. Use getChromatogram() instead
}
