% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getEIC}
\alias{getEIC}
\title{Get extracted ion chromatogram}
\usage{
getEIC(object, ...)
}
\arguments{
\item{object}{An object that has ion chromatograms}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A numeric vector with the extracted ion chromatogram
}
\description{
Get extracted ion chromatogram
}
