% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGCIMSSample}
\alias{getGCIMSSample}
\title{Get a GCIMSSample}
\usage{
getGCIMSSample(object, ...)
}
\arguments{
\item{object}{An object where we can extract GCIMSSamples from}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
A \link{GCIMSSample} object
}
\description{
Get a GCIMSSample
}
