% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/xxx-deprecated-getSample-getSpectrum-getChromatogram.R
\name{getGCIMSSample}
\alias{getGCIMSSample}
\title{Deprecated function. Use getSample() instead}
\usage{
getGCIMSSample(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=getSample]{getSample()}}}
}
\description{
Deprecated function. Use getSample() instead
}
