% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getIMSScan}
\alias{getIMSScan}
\title{Get IMS scans}
\usage{
getIMSScan(object, ...)
}
\arguments{
\item{object}{An object that has IMS scans}

\item{...}{Further arguments, possibly used by downstream methods.}
}
\value{
An ion mobility spectrum
}
\description{
Get IMS scans
}
\examples{
x <- dummy_obj <-GCIMSSample(
  drift_time = 1:2,
  retention_time = 1:3,
  data = matrix(1:6, nrow = 2, ncol = 3)
)
getIMSScan(x) # c(`1` = 9, `2` = 12)
}
