% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-getTIS_getRIC.R
\name{getRIC,GCIMSDataset-method}
\alias{getRIC,GCIMSDataset-method}
\title{Get Reverse Ion Chromatogram}
\usage{
\S4method{getRIC}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}
}
\value{
The RIC matrix
}
\description{
Get Reverse Ion Chromatogram
}
