% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{group_peak_list}
\alias{group_peak_list}
\title{Peak grouping function, exposing a lot of options useful for benchmarking}
\usage{
group_peak_list(
  peaks,
  filter_dt_width_criteria = "IQR",
  filter_rt_width_criteria = "arnau",
  distance_method = "mahalanobis",
  distance_between_peaks_from_same_sample = Inf,
  clustering = list(method = "kmedoids", Nclusters = "max_peaks_sample"),
  aggregate_conflicting_peaks = NULL,
  verbose = FALSE
)
}
\arguments{
\item{peaks}{A data frame with at least the following columns:
\itemize{
\item "UniqueID" A unique ID for each peak
\item "SampleID" The sample ID the peak belongs to
\item "dtapex_ms", "rtapex_s" The peak positions
\item "volume" The size of the peak
\item "dt_max_ms", "dt_min_ms", "rt_max_s", "rt_min_s" (for filtering outlier peaks based on their size)
}}

\item{filter_dt_width_criteria, filter_rt_width_criteria}{A character with the method for outlier detection.
\itemize{
\item "IQR": Remove peaks more than 1.5 interquartile ranges above upper quartile or
below the lower quartile.
\item "arnau": FIXME Adhoc method from Arnau, where he removes peaks above mean+4iqr or below median-0.75iqr
\item "none": Do not remove peaks based on their drift time width or retention time height
}}

\item{distance_method}{A string. One of the distance methods from \link[stats:dist]{stats::dist} or "mahalanobis"}

\item{distance_between_peaks_from_same_sample}{The distance between two peaks from the same sample will be set to \code{distance_between_peaks_from_same_sample*max(distance_matrix)}}

\item{clustering}{A named list with "method" and the supported method, as well as further options.
For \code{method = "kmedoids"}, you must provide \code{Nclusters}, with either the number of clusters
to use in the kmedoids algorithm or the string "max_peaks_sample" to use the maximum number of
detected peaks per sample.

No other clustering method is currently supported}

\item{aggregate_conflicting_peaks}{\code{NULL} or a function. When we build the peak table, with peaks in rows, samples in
columns, peak_table\link{i,j} is the volume of the peak from sample j in cluster i. If the clustering process
clusters together two peaks form the same sample, those peaks will conflict in the peak table. \code{NULL} will error
in that case, another function will be applied on the conflicting volumes (e.g \code{mean} or \code{max} would be reasonable options)}

\item{verbose}{logical, to control printing in the function}
}
\value{
A list with :
\itemize{
\item peak_table: A peak table that includes peak position, median peak minimum/maximum retention and drift times and the peak volume for each sample
\item peak_table_duplicity: How many volume values have been aggregated. Should be 1 for each sample/peak
\item extra_clustering_info: Arbitrary clustering extra information, that depends on the clustering method
}
}
\description{
Peak grouping function, exposing a lot of options useful for benchmarking
}
\examples{
# FIXME
}
