% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeaksClustering.R
\name{group_peak_list}
\alias{group_peak_list}
\title{Group peaks in clusters}
\usage{
group_peak_list(
  peaks,
  filter_dt_width_criteria = "IQR",
  filter_rt_width_criteria = "arnau",
  distance_method = "mahalanobis",
  distance_between_peaks_from_same_sample = 100,
  clustering = list(method = "kmedoids", Nclusters = "max_peaks_sample"),
  verbose = FALSE
)
}
\arguments{
\item{peaks}{A data frame with at least the following columns:
\itemize{
\item "UniqueID" A unique ID for each peak
\item "SampleID" The sample ID the peak belongs to
\item "dt_apex_ms", "rt_apex_s" The peak positions
\item "dt_max_ms", "dt_min_ms", "rt_max_s", "rt_min_s" (for filtering outlier peaks based on their size)
}}

\item{filter_dt_width_criteria, filter_rt_width_criteria}{A character with the method for outlier detection.
\itemize{
\item "IQR": Remove peaks with widths more than 1.5 interquartile ranges above upper quartile or
below the lower quartile.
\item "arnau": FIXME Adhoc method from Arnau, where he removes peaks with widths above mean+4iqr or below median-0.75iqr
\item "none": Do not remove peaks based on their drift time width or retention time height
}}

\item{distance_method}{A string. One of the distance methods from \link[stats:dist]{stats::dist}, "sd_scaled_euclidean" or "mahalanobis"}

\item{distance_between_peaks_from_same_sample}{The distance between two peaks from the same sample will be set to \code{distance_between_peaks_from_same_sample*max(distance_matrix)}}

\item{clustering}{A named list with "method" and the supported method, as well as further options.
For \code{method = "kmedoids"}, you must provide \code{Nclusters}, with either the number of clusters
to use in the kmedoids algorithm (\link[cluster:pam]{cluster::pam}) or the string \code{"max_peaks_sample"} to use the maximum number of
detected peaks per sample.

For \code{method = "hclust"}, you can provide \code{hclust_method}, with the \code{method} passed to \link[stats:hclust]{stats::hclust}.}

\item{verbose}{logical, to control printing in the function}
}
\value{
A list with :
\itemize{
\item peak_table: A peak table that includes peak position, median peak minimum/maximum retention and drift times and the peak Volume for each sample
\item peak_table_duplicity: How many Volume values have been aggregated. Should be 1 for each sample/peak
\item extra_clustering_info: Arbitrary clustering extra information, that depends on the clustering method
}
}
\description{
Peak grouping function, exposing several options useful for benchmarking.
}
\examples{
\donttest{
dir_in <- system.file("extdata", package = "GCIMS")
peak_list <- readRDS(file.path(dir_in, "peak_list.rds"))

peak_clustering  <- group_peak_list(
  peaks = peak_list,
  filter_dt_width_criteria = NULL,
  filter_rt_width_criteria = NULL,
  distance_method = "mahalanobis",
  distance_between_peaks_from_same_sample = Inf,
  clustering = list(method = "kmedoids", Nclusters = "max_peaks_sample"),
  verbose = FALSE
)
}
}
