% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R, R/cow.R
\name{InterpCoeff}
\alias{InterpCoeff}
\title{Function to calculate coefficients for interpolation}
\usage{
InterpCoeff(n, nprime, offs)

InterpCoeff(n, nprime, offs)
}
\arguments{
\item{n}{number of coefficients}

\item{nprime}{??}

\item{offs}{??}

\item{n}{number of coefficients}

\item{nprime}{??}

\item{offs}{??}
}
\value{
a list with Coeff and Index

a list with Coeff and Index
}
\description{
Function to calculate coefficients for interpolation

Function to calculate coefficients for interpolation
}
