% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_functions.R
\name{optimize_cow}
\alias{optimize_cow}
\title{Obtain the optimum warping for (COW)}
\usage{
optimize_cow(dir_in, dir_out, samples, time, seg_vector, slack_vector)
}
\arguments{
\item{dir_in}{The input directory.}

\item{dir_out}{The output directory.}

\item{samples}{The set of samples to be processed.}

\item{seg_vector}{Vector of segment lengths.}

\item{slack_vector}{Vector of slacks.}

\item{time}{It indicates if the correction is going to be in the
drift time or in the retention time. It should be
introduce "Retention" for correcting the retention
time; or "Drift" for the drift time.}
}
\value{
A matrix with the optimum warping.
}
\description{
Obtain the optimum warping for (COW)
}
\examples{
\dontrun{
dataset_2_polarities <- lcms_dataset_load(system.file("extdata",
                                                     "dataset_metadata.rds",
                                                     package = "NIHSlcms"))
dataset_pos <- lcms_filter_polarity(dataset_2_polarities, polarity. = 1)

print(dataset_pos)
}
}
\seealso{
Other Alignment functions:
\code{\link{apply_cow}()},
\code{\link{cow}()},
\code{\link{gcims_alignment}()}
}
\concept{Alignment functions}
