% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-GCIMSSample.R
\name{overlay_peaklist}
\alias{overlay_peaklist}
\title{Overlay a peak list to a plot}
\usage{
overlay_peaklist(
  peaklist = NULL,
  ...,
  dt_range = NULL,
  rt_range = NULL,
  pdata = NULL,
  color_by = NULL,
  mapping_roi = c(dt_min_ms = "dt_min_ms", dt_max_ms = "dt_max_ms", rt_min_s =
    "rt_min_s", rt_max_s = "rt_max_s"),
  palette = P40
)
}
\arguments{
\item{peaklist}{A data frame with at least the columns: \code{dt_min_ms}, \code{dt_max_ms}, \code{rt_min_s}, \code{rt_max_s}
and optionally additional columns (e.g. the column given to \code{color_by})}

\item{...}{Ignored.}

\item{dt_range}{The minimum and maximum drift times to extract (length 2 vector)}

\item{rt_range}{The minimum and maximum retention times to extract (length 2 vector)}

\item{pdata}{A phenotype data data frame, with a \code{SampleID} column. This column will be used to
merge \code{pdata} with \code{peaklist}, so \code{color_by} can specify a phenotype.}

\item{color_by}{A character with a column name of \code{peaklist} or \code{pdata}. Used to color the border of
the added rectangles and apices. A string with a color name is also acceptable.}

\item{mapping_roi}{A 4-elements named character vector with the names of the columns from \code{peaklist} that will
be used as the rectangle coordinates.}

\item{palette}{A character vector with color names to use drawing the rectangles. Use \code{NULL} to let \code{ggplot2} set the defaults.}
}
\value{
A list with the ggplot layers to overlay, including \code{geom_rect} and possibly \code{geom_point} and \code{scale_fill_manual}.
}
\description{
Overlay a peak list to a plot
}
\details{
If \code{peaklist} includes \code{dt_apex_ms} and \code{rt_apex_s} a cross will be plotted on the peak apex.
}
\examples{
dt <- 1:10
rt <- 1:10
int <- matrix(0.0, nrow = length(dt), ncol = length(rt))

int[2, 4:8] <- c(.5, .5, 1, .5, 0.5)
int[3, 4:8] <- c(0.5, 2, 2, 2, 0.5)
int[4, 4:8] <- c(1, 2, 5, 2, 1)
int[5, 4:8] <- c(0.5, 2, 2, 2, 0.5)
int[6, 4:8] <- c(.5, .5, 1, .5, 0.5)

dummy_obj <-GCIMSSample(
  drift_time = dt,
  retention_time = rt,
  data = int
)
plt <- plot(dummy_obj)

# Add a rectangle on top of the plot
rect <- data.frame(
  dt_min_ms = 2.75,
  dt_max_ms = 5.6,
  rt_min_s = 4.6,
  rt_max_s = 7.4
)

plt + overlay_peaklist(
  peaklist = rect
)
}
