% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData-GCIMSDataset.R
\name{pData,GCIMSDataset-method}
\alias{pData,GCIMSDataset-method}
\title{Get the phenotype data}
\usage{
\S4method{pData}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}
}
\description{
Get the phenotype data
}
