% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks-GCIMSDataset.R
\name{peaks,GCIMSDataset-method}
\alias{peaks,GCIMSDataset-method}
\title{Get the peak list}
\usage{
\S4method{peaks}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}
}
\description{
Get the peak list
}
