% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Make a plot interactive}
\usage{
plot_interactive(plt)
}
\arguments{
\item{plt}{A ggplot plot}
}
\value{
A plotly plot
}
\description{
Wraps the \code{plt} with \code{\link[plotly:ggplotly]{plotly::ggplotly()}} and sets the \code{xaxis} and \code{yaxis}
ticks to \code{"auto"}, so the axis labels are updated when zooming.
}
\examples{
plot_interactive()
}
