% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-getTIS_getRIC.R
\name{plotTIS,GCIMSDataset-method}
\alias{plotTIS,GCIMSDataset-method}
\title{Plot Total Ion Spectra}
\usage{
\S4method{plotTIS}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object}
}
\value{
A plot
}
\description{
Plot Total Ion Spectra
}
