% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-realize.R
\name{realize}
\alias{realize}
\title{Runs all delayed operations on the object}
\usage{
realize(object)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}
}
\value{
The same \link{GCIMSDataset} object, without pending operations
}
\description{
Runs all delayed operations on the object
}
\examples{
base_dir <- system.file("extdata", "sample_formats", package = "GCIMS")
annot <- data.frame(SampleID = "Sample1", FileName = "small.mea.gz")
dataset <- GCIMSDataset(annot, base_dir)
print(dataset)
realize(dataset)
print(dataset)

}
