% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realize-GCIMSDataset.R
\name{realize}
\alias{realize}
\title{Runs all delayed operations on the object}
\usage{
realize(object, keep_intermediate = NA)
}
\arguments{
\item{object}{A \link{GCIMSDataset} object, modified in-place}

\item{keep_intermediate}{A logical, whether to keep the intermediate files of
the previous realization once this one finishes. If \code{NA}, keeping will depend
on the \code{object}.}
}
\value{
The same \link{GCIMSDataset} object, without pending operations
}
\description{
Runs all delayed operations on the object
}
\examples{
base_dir <- system.file("extdata", "sample_formats", package = "GCIMS")
annot <- data.frame(SampleID = "Sample1", FileName = "small.mea.gz")
dataset <- GCIMSDataset$new(annot, base_dir)
print(dataset)
realize(dataset)
print(dataset)

}
