% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSChromatogram-dtime-rtime-intensity.R
\name{rtime,GCIMSChromatogram-method}
\alias{rtime,GCIMSChromatogram-method}
\title{Get the retention time vector}
\usage{
\S4method{rtime}{GCIMSChromatogram}(object)
}
\arguments{
\item{object}{A GCIMSChromatogram}
}
\value{
The retention time vector (in s)
}
\description{
Get the retention time vector
}
