% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCIMSDataset-get_dtime_rtime.R
\name{rtime,GCIMSDataset-method}
\alias{rtime,GCIMSDataset-method}
\title{Get a reference retention time vector for the dataset}
\usage{
\S4method{rtime}{GCIMSDataset}(object)
}
\arguments{
\item{object}{A GCIMSDataset}
}
\value{
a retention time vector
}
\description{
Get a reference retention time vector for the dataset
}
