% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-GCIMSChromatogram.R
\name{smooth,GCIMSChromatogram-method}
\alias{smooth,GCIMSChromatogram-method}
\title{Smoothing a GCIMS chromatogram using a Savitzky-Golay filter}
\usage{
\S4method{smooth}{GCIMSChromatogram}(x, rt_length_s = 3, rt_order = 2L)
}
\arguments{
\item{x}{A \link{GCIMSChromatogram} object}

\item{rt_length_s}{The length of the filter in retention time (in s)}

\item{rt_order}{The order of the filter in retention time}
}
\value{
The modified \link{GCIMSChromatogram}
}
\description{
Smoothing a GCIMS chromatogram using a Savitzky-Golay filter
}
