% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{write_mea}
\alias{write_mea}
\title{write a mea file (experimental)}
\usage{
write_mea(object, filename)
}
\arguments{
\item{object}{A \code{GCIMSSample} object}

\item{filename}{The filename to write it to (will have .mea appended if no extension is given)}
}
\value{
Nothing
}
\description{
This function writes a GCIMSSample into a .mea file. The implementation
misses a lot of metadata, and may not be fully compliant with LAV or VOCal,
so it is not recommended in production environments.
}
\examples{
obj <- GCIMSSample(
 drift_time = 1:2,
 retention_time = 1:3,
 data = matrix(1:6, nrow = 2, ncol = 3),
)
mea_file <- tempfile(fileext = ".mea")
write_mea(obj, mea_file)
file.remove(mea_file)
}
